% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_button.R
\name{action_button}
\alias{action_button}
\title{Create an action button with additional styling, icon support, and text
customization}
\usage{
action_button(
  classes = c(),
  id,
  icon = NULL,
  icon_style = NULL,
  text_class = NULL,
  text = NULL
)
}
\arguments{
\item{classes}{<\verb{character vector}> Additional CSS classes to apply to the
button. Default is an empty character vector.}

\item{id}{<\code{character}> A unique identifier for the button.}

\item{icon}{<\code{character}> A character string specifying the name of the
Material Design icon to be used in the button. Default is NULL for no icons. The
icon is made through \code{\link{icon_material}}.}

\item{icon_style}{<\code{character}> Additional CSS classes to apply to the icon.
Default is NULL for no additional classes.}

\item{text_class}{<\code{character}> Additional CSS classes to apply to the button
text. Default is NULL for no additional classes.}

\item{text}{<\code{character}> Button text. Defaults to NULL for no text.}
}
\value{
A Shiny tags$button (with the \code{action-button} class making it
essentially an \code{\link[shiny]{actionButton}}) element with the specified
attributes and styling.
}
\description{
This function creates a custom action button for use in Curbcut. The
button supports additional CSS classes, as well as the inclusion of a Material
Design icon, custom icon styling, and customizable text styling.
}
\details{
It is a reemplementation of \code{\link[shiny]{actionButton}}.
}
