% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_advanced_options.R
\name{adv_opt_lock_selection}
\alias{adv_opt_lock_selection}
\title{Select all the IDs for a given address or postal code}
\usage{
adv_opt_lock_selection(address, lang = NULL)
}
\arguments{
\item{address}{<\code{character}> A character string specifying the address or
postal code to search for.}

\item{lang}{<\code{character}> A character string specifying the language to use
for the error messages. Defaults to NULL which is no translation.}
}
\value{
A vector of IDs for all scales that fit with the input address or
postal code.
}
\description{
This function takes an address or postal code as input and returns the
corresponding IDs of all the scales that fit around the address or postal code.
If the input is a postal code, the function checks if it is within an available
region and returns the corresponding zones. If the input is an address, the function
geocodes the address and checks the closest dissemination areas for the
geocoded point, returning the corresponding IDs. The output is a vector of
IDs that fit with the address or postal code.
}
\seealso{
\code{\link{settings_server}} for the server code that uses this
function.

\code{\link{adv_opt_lock_selection_UI}} That creates the UI for this
function.
}
