% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_advanced_options.R
\name{adv_opt_lock_selection_UI}
\alias{adv_opt_lock_selection_UI}
\title{Create a UI element for selecting and locking a default location}
\usage{
adv_opt_lock_selection_UI(id, lang = NULL)
}
\arguments{
\item{id}{<\code{character}> The ID of the module, in which this UI bit will
be used. Necessary for namespacing reasons.}

\item{lang}{A string indicating the language of the UI element. Defaults to
\code{NULL} which means no translation.}
}
\value{
A Shiny UI element for selecting and locking a default location.
}
\description{
This function creates a Shiny UI element that allows the user to select
and lock in a default location (postal code or address) for further use in
any Curbcut map page. If a location is locked, then at every zoom level on
a map the location will be automatically, as a default, the zone in which
the locked location falls in.
}
