% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_advanced_options.R
\name{adv_opt_region}
\alias{adv_opt_region}
\title{Create a shiny radio button input for selecting a geographical region.}
\usage{
adv_opt_region(id, region, lang = NULL)
}
\arguments{
\item{id}{<\code{character}> The ID of the module, in which this UI bit will
be used. Necessary for namespacing reasons.}

\item{region}{<\code{character}> The current default region, usually region = r$region().}

\item{lang}{<\code{character}> A character string specifying the language in
which the names of the regions should be translated. Defaults to \code{NULL}
which means no translation.}
}
\value{
A shiny radio button input object.
}
\description{
This function creates a shiny radio button input for selecting a geographical
region from a predefined list of regions. The df of regions is obtained
from a global environment variable named \code{regions_dictionary}. This list
contains all the regions that can be picked, with the following columns:
\code{pickable} if the user can select them in the advance options, the \code{name}
column, and the \code{region} code.
}
