% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_controls.R
\name{advanced_controls_server}
\alias{advanced_controls_server}
\title{Advanced Controls (server, although NOT A MODULE)}
\usage{
advanced_controls_server(id, r, label = shiny::reactive("Show"))
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{label}{<\code{character}> Label for the checkbox. Must be a reactive.
Defaults to \code{shiny::reactive("Show")}}
}
\value{
None
}
\description{
This function defines the server-side behavior of the advanced controls module.
The advanced controls feature a checkbox that, when toggled, will show or
hide additional widgets.
}
