% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_controls.R
\name{advanced_controls_UI}
\alias{advanced_controls_UI}
\title{Advanced Controls - UI}
\usage{
advanced_controls_UI(id, label = cc_t("Show"), ...)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}. NOT A MODULE, NO NAMESPACING.}

\item{label}{<\code{character}> Label for the checkbox. Defaults to \code{cc_t("Show")}}

\item{...}{<\code{args}> UIs to be inserted in the advanced controls div.}
}
\value{
A Shiny UI definition
}
\description{
This function creates the user interface for the advanced controls module.
It is designed to be used in tandem with the advanced_controls_server function.
}
