% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovars_helpers.R
\name{autovars_common_widgets}
\alias{autovars_common_widgets}
\title{Get the Common Widgets for a Specific Module ID}
\usage{
autovars_common_widgets(id)
}
\arguments{
\item{id}{<\code{character}> Identifying the module in modules$id

If 'var_lefts' (the variables left to be selected) is a data frame,
it retrieves the variable list from the 'var_code' column.

It retrieves time information from 'variables' object and other widgets
by subsetting the 'group_diff' column of the 'variables' object.}
}
\value{
A list with two components. The first component is a numeric vector
of unique time points, and the second component is a list of other widgets.
The list of other widgets will be empty if no other widgets are common to all variables.
}
\description{
This function retrieves the common widgets (time and other widgets)
associated with a specific module ID. The information is retrieved from
global environment objects named 'modules' and 'variables'.
}
