% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovars_helpers.R
\name{autovars_final_value}
\alias{autovars_final_value}
\title{Determine Final Value}
\usage{
autovars_final_value(id, group_name, picker_vals, previous_var)
}
\arguments{
\item{id}{<\code{character}> Identifying the module in modules$id}

\item{group_name}{<\code{character}> String identifying the group name to only filter
the widgets of the same group.}

\item{picker_vals}{<\code{list}> Selected values from pickers and sliders in a list.
Used to check which of the 'group_diff' values are included.}

\item{previous_var}{<\code{character}> A previously selected variable. It's returned
if the tibble is not a dataframe or there are no fits for the picker values.}
}
\value{
If the retrieved tibble is not a data frame, the function simply returns
the group name (if it isn't null), or the previous variable. If the tibble is a
data frame, the function returns the 'var_code' associated with
the maximum sum of fits.
}
\description{
This function returns a final value based on an id, group name, selected values
from a picker, and a previously selected variable. It operates in the context
of a global object named 'modules'.
}
