% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovars_helpers.R
\name{autovars_groupnames}
\alias{autovars_groupnames}
\title{Retrieve Group Names for a Specific Module}
\usage{
autovars_groupnames(id, pres = FALSE)
}
\arguments{
\item{id}{<\code{character}> Identifying the module in modules$id}

\item{pres}{<\code{logical}> If TRUE, will return a logical on if the there should
be a widgets existing at all. Defaults to FALSE, which returns the default
return of this function.

If 'var_lefts' (the variables left to be selected) is a data frame,
it returns the unique group names from the 'group_name' column.

If 'var_lefts' is a character vector, it returns a fully prepared dropdown list.}
}
\value{
A character vector of group names or a list prepared for a dropdown selection,
depending on the type of 'var_lefts'.
}
\description{
This function retrieves the group names associated with a specific module ID.
The group names are retrieved from a global environment object named 'modules'.
They are used to populate the main dropdown.
}
