% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovars_helpers.R
\name{autovars_placeholder_var}
\alias{autovars_placeholder_var}
\title{Generate Placeholder Variable}
\usage{
autovars_placeholder_var(id)
}
\arguments{
\item{id}{<\code{character}> Identifying the module in modules$id#'}
}
\value{
If the 'var_left' of the module is a data frame, the function returns
the first element of the 'var_code' column. If 'var_left' is a character
vector, the function returns its first element.
}
\description{
This function retrieves the first variable of a specified module.
The module is selected by its ID from a global object named 'modules'.
}
