% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovars_helpers.R
\name{autovars_time_ui}
\alias{autovars_time_ui}
\title{Create UI for time variable selection}
\usage{
autovars_time_ui(
  id,
  times,
  widget_ns,
  time_div_label = "Time",
  time_div_icon = "date_range",
  compare_label = "Compare dates"
)
}
\arguments{
\item{id}{<\code{character}> The namespace ID of the page.}

\item{times}{<\code{vector}> Vector of times (Usually common_widgets$time())}

\item{widget_ns}{<\code{function}> A function for namespacing widget IDs.}

\item{time_div_label}{<\code{character}> A function that returns the label for the
time division.}

\item{time_div_icon}{<\code{character}> Material ion name for the time selection UI, defaults
to "date_range".}

\item{compare_label}{<\code{character}> Label of the checkbox used to compare dates}
}
\value{
<\code{shiny.tagList}> A tag list containing the UI elements for the
time variable selection.
}
\description{
This function creates a user interface element for selecting time variables.
It allows for the creation of either a slider or
a slider with text, depending on the structure of the time data provided.
}
