% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_connect}
\alias{aws_connect}
\title{Establish a Connection to AWS RDS with SSL}
\usage{
aws_connect()
}
\value{
A \code{DBIConnection} object representing the connection to the RDS instance.
}
\description{
Uses the \code{DBI} package to establish a connection to an AWS RDS instance.
It automatically generates an RDS authentication token and determines
the SSL root certificate path for the connection.
}
