% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_generate_rds_auth_token}
\alias{aws_generate_rds_auth_token}
\title{Generate an AWS RDS Authentication Token}
\usage{
aws_generate_rds_auth_token()
}
\value{
A character string containing the generated authentication token.
}
\description{
This function generates an authentication token for AWS RDS using the AWS CLI.
It requires certain environment variables to be set for the function to succeed.
These include AWS credentials and RDS instance details.
}
