% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_get_rootcert}
\alias{aws_get_rootcert}
\title{Retrieve or Download the SSL Root Certificate}
\usage{
aws_get_rootcert()
}
\value{
A character string with the path to the SSL root certificate.
}
\description{
Depending on the execution environment (production or development),
this function retrieves the path to an existing SSL root certificate
or downloads it from a specified AWS S3 bucket.
}
