% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_pool}
\alias{aws_pool}
\title{AWS RDS Database Connection Pool Initialization}
\usage{
aws_pool()
}
\value{
A pooled database connections object on success, or stops with an error
message if unable to connect.
}
\description{
Initializes a connection pool for an AWS RDS database using \code{cc_dbPool}.
It configures the pool based on environment variables and specific AWS settings,
}
