% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bookmark_codes}
\alias{bookmark_codes}
\title{Make shorter bookmarks using codes}
\format{
\subsection{\code{bookmark_codes}}{

A named vector:
\describe{
\item{name}{The real widget namespace}
\item{value}{New widget code}
}
}
}
\usage{
bookmark_codes
}
\description{
Make shorter bookmarks using codes
}
\keyword{datasets}
