% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark_module.R
\name{bookmark_server}
\alias{bookmark_server}
\title{Server function for creating a bookmarkable URL with Curbcut widgets and an
rdeck map}
\usage{
bookmark_server(
  id,
  r,
  select_id = shiny::reactive(NULL),
  map_viewstate = shiny::reactive(NULL),
  exclude_input = c()
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the bookmark_server function
will appear, e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{select_id}{<\verb{reactive character}> the current selected ID, usually
\code{r[[id]]$select_id()}. Defaults to \code{NULL} for when it is not a map module.}

\item{map_viewstate}{<\verb{reactive list}> The map viewstate. Usually the output of the
\code{\link{map_server}}, or of \code{\link[rdeck]{get_view_state}}. Defaults
to \code{NULL} for when it is not a map module.}

\item{exclude_input}{<\verb{character vector}> Widget names to exclude from
bookmarking. By default, the input monitoring the opening of dropdowns
is excluded. Example of exclusion: \code{ccpicker_var}.}
}
\value{
The module server function for creating bookmarkable URLs.
}
\description{
This function creates a module server for bookmarkable app that use Curbcut
widgets and an rdeck map. It bookmarks the tab id, the region, language,
widgets, the map view state (zoom and lat/lon coordinates), the selected id
and the \code{df}.
}
