% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bookmark_shorts}
\alias{bookmark_shorts}
\title{Make shorter all widgets id using codes}
\format{
\subsection{\code{bookmark_shorts}}{

A named vector:
\describe{
\item{name}{The start of the widget \code{id}}
\item{value}{\code{id} for which to switch to make bookmark URL smaller}
}
}
}
\usage{
bookmark_shorts
}
\description{
Make shorter all widgets id using codes
}
\keyword{datasets}
