% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{bookmark_widget_helper}
\alias{bookmark_widget_helper}
\title{Helper function to separate and process bookmarked Shiny widgets}
\usage{
bookmark_widget_helper(id, wgt, lang = NULL)
}
\arguments{
\item{id}{<\code{character}> Page ID for which the bookmark needs to be updated,
e.g. \code{canale}.}

\item{wgt}{<\code{character}> A character string representing the widget inputs
(the \code{wgt} part of the bookmarked URL).}

\item{lang}{<\code{character}> An optional character string specifying the language
used on the app. Default is NULL.}
}
\value{
A list of processed bookmarked Curbcut widgets separated into three
categories: "cbox" (checkboxes), "s_text" (slider text), and "picker" (pickers).
}
\description{
This function takes a bookmarked widget string and separates the widgets into
two groups: those corresponding to \code{\link{bookmark_codes}} and those to
\code{\link{bookmark_shorts}} It then processes the widget IDs and values as
needed and returns them as a list to be used by the \code{\link{use_bookmark}}
function.
}
