% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{breaks_delta}
\alias{breaks_delta}
\title{Calculate Breaks for Delta Legends}
\usage{
breaks_delta(vars, df, character = FALSE, data = NULL)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link{vars_build}}.}

\item{df}{<\verb{reactive character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{character}{<\code{Logical}>, if \code{TRUE}, the breaks are returned as characters.}

\item{data}{<\code{data.frame}> Optional, a data frame to use instead of recovering it
from the file.}
}
\value{
A vector of break points. Numeric if \code{character = FALSE}, otherwise
character.
}
\description{
This function calculates the break points for delta legends based on the data.
It also handles special cases, such as when all data is within a small range.
}
