% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get_helpers.R
\name{breaks_delta_helper}
\alias{breaks_delta_helper}
\alias{breaks_delta_helper.default}
\alias{breaks_delta_helper.positive}
\alias{breaks_delta_helper.negative}
\title{Breaks Delta Helper Function}
\usage{
breaks_delta_helper(data, character, ...)

\method{breaks_delta_helper}{default}(data, character, ...)

\method{breaks_delta_helper}{positive}(data, character, ...)

\method{breaks_delta_helper}{negative}(data, character, ...)
}
\arguments{
\item{data}{<\code{numeric}> Data containg the class fo inform if the values are
only negative, positive or mixed.}

\item{character}{<\code{logical}> If TRUE, returns breaks as character strings.}

\item{...}{Additional arguments}
}
\value{
<\code{numeric/character}> Breaks from negative to positive values.

Return breaks from negative to positive values

Return breaks as only positive values

Return breaks as only negative values
}
\description{
This function and its methods create breaks for positive, negative, or both
types of values in a given vector. It handles outliers and provides options
for output format.
}
\section{Methods (by class)}{
\itemize{
\item \code{breaks_delta_helper(default)}: default method

\item \code{breaks_delta_helper(positive)}: positive method

\item \code{breaks_delta_helper(negative)}: negative method

}}
