% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_helpers.R
\name{calculate_map_zoom_level}
\alias{calculate_map_zoom_level}
\title{Calculate the map zoom level based on top scale value}
\usage{
calculate_map_zoom_level(top_scale, avail_scale_combinations)
}
\arguments{
\item{top_scale}{<\code{character}> The current top scale value.}

\item{avail_scale_combinations}{<\verb{character vector}> Available scale combinations
for the top scale.}
}
\value{
A numeric vector representing the map zoom levels.
}
\description{
Calculate the map zoom level based on top scale value
}
