% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{cc_t_list}
\alias{cc_t_list}
\title{Translate names of elements in a nested list}
\usage{
cc_t_list(x, translation_df)
}
\arguments{
\item{x}{<\code{list}> A nested list where the names of elements will be translated.}

\item{translation_df}{<\code{data.frame}> A data.frame with two columns: en and fr.
It contains the original names in English and the corresponding translated
names in French.}
}
\value{
The same nested list with the names of elements translated.
}
\description{
This function translates the names of elements in a nested list using a
translation data.frame
}
\examples{
translation_df <- data.frame(
  en = c("apple", "banana", "cherry", "fruits", "vegetables"),
  fr = c("pomme", "banane", "cerise", "fruits", "légumes")
)

x <- list(
  fruits = list(apple = 1, banana = 2),
  vegetables = list(carrot = 3, lettuce = 4)
)

cc_t_list(x, translation_df)
}
