% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{cc_t_span}
\alias{cc_t_span}
\title{Translate an object between English and French, returning spans}
\usage{
cc_t_span(..., .envir = parent.frame())
}
\arguments{
\item{...}{<\verb{character objects}> Any objects, including lists or atomic vectors}

\item{.envir}{The parent environment for evaluating expressions in
\code{...}. Used to get the correct value of variables handled by
\code{\link[glue]{glue_safe}}. Defaults to \code{parent.frame()}.}
}
\value{
Returns a \code{\link[shiny]{tagList}} containing two spans, one for
each language (English and French).
}
\description{
Translate input object from English to French, and returns two spans,
one for each language. This is intended for UI elements.
}
\seealso{
\code{\link{cc_t}} for more general translation behavior
}
