% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_checkbox.R
\name{checkbox_server}
\alias{checkbox_server}
\title{Shiny module for a checkbox input control}
\usage{
checkbox_server(
  id,
  r,
  checkbox_id = "cbox",
  label = shiny::reactive(NULL),
  event_reset = shiny::reactive(NULL)
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file.}

\item{checkbox_id}{<\code{character}> string giving the identifier for the checkbox input
object. This will be used as the input's \code{inputId} and will inform the bookmarking.
It needs to be unique in the page. The default value is \code{"cbox"}.}

\item{label}{<\code{character}> A text label to display next to the checkbox
input. If NULL (default), no label will be displayed.}

\item{event_reset}{<\verb{reactive expression}> A change in this reactive will trigger
the reset of the checkbox value to TRUE. If NULL (default), the checkbox value
will never be reset.}
}
\value{
A reactive object representing the value of the checkbox input control.
}
\description{
This function updates a checkbox input control for use in Shiny server. The
checkbox allows users to select a boolean value. The function uses the
\code{\link{checkboxInput}} function and returns its value as a reactive object.
}
\seealso{
\code{\link{checkbox_UI}}
}
