% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_checkbox.R
\name{checkbox_UI}
\alias{checkbox_UI}
\title{Create a checkbox input control}
\usage{
checkbox_UI(id, checkbox_id = "cbx", label = NULL, value = TRUE, ...)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{alp}.}

\item{checkbox_id}{<\code{character}> string giving the identifier for the checkbox input
object. This will be used as the input's \code{inputId} and will inform the bookmarking.
It needs to be unique in the page. The default value is \code{"cbx"}.}

\item{label}{<\code{character}> A text label to display next to the checkbox input.
If NULL (default), no label will be displayed.}

\item{value}{<\code{logical}> The initial value of the checkbox input. Defaults to
\code{TRUE}.}

\item{...}{Additional arguments to pass to the \code{\link{checkboxInput}} function.}
}
\value{
A checkbox input control for use in Shiny UI.
}
\description{
This function creates a checkbox input control for use in Curbcut. The checkbox
allows users to select a logical value. The function uses the \code{\link{checkboxInput}}
function and returns its output.
}
\seealso{
\code{\link{checkbox_server}}
}
