% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_module.R
\name{compare_server}
\alias{compare_server}
\alias{compare_UI}
\title{Server logic for comparing variables in Curbcut}
\usage{
compare_server(
  id,
  r,
  var_list,
  time = shiny::reactive(NULL),
  show_panel = shiny::reactive(TRUE)
)

compare_UI(id, var_list)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the module will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{var_list}{<\verb{reactive list}> Choices to display in the picker input. Normally
made using \code{\link{dropdown_make}}.}

\item{time}{<\verb{reactive numeric vector}> Vector of time values to use for
appending a time to the variables picked. The returned vector will be the
same length as this argument.}

\item{show_panel}{<\verb{reactive logical}> Controls whether the comparison panel
should be shown.}
}
\value{
A Shiny module server function that provides the logic for
comparing variables in Curbcut
}
\description{
This function provides the server logic for comparing the main variable with
different variables in Curbcut.
}
\details{
This function uses the \code{\link{picker_server}} function to
generate a reactive expression (\code{var_right}) that contains the selected
variable(s) for comparison. It also uses the \code{\link[shinyjs]{toggle}}
function to show/hide the comparison panel based on the value of the
\code{show_panel} reactive expression.
}
\section{Functions}{
\itemize{
\item \code{compare_UI()}: Create the UI for the compare module

}}
