% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{convert_unit.degree}
\alias{convert_unit.degree}
\title{Convert a numeric vector into a character vector with a degrees sign}
\usage{
\method{convert_unit}{degree}(var, x, ...)
}
\arguments{
\item{var}{Same as in the generic function.}

\item{x}{Same as in the generic function.}

\item{...}{Additional arguments to be passed to the method.}
}
\value{
A character vector with a percentage sign.
}
\description{
Convert a numeric vector into a character vector with a degrees sign
}
