% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{convert_unit.dollar}
\alias{convert_unit.dollar}
\title{Convert a numeric vector into a character vector with a dollar sign}
\usage{
\method{convert_unit}{dollar}(var, x, compact = FALSE, ...)
}
\arguments{
\item{var}{Same as in the generic function.}

\item{x}{Same as in the generic function.}

\item{compact}{<\code{logical}> Whether to attempt to compact large numbers.}

\item{...}{Additional arguments to be passed to the method.}
}
\value{
A character vector with a dollar sign.
}
\description{
Convert a numeric vector into a character vector with a dollar sign
}
