% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{convert_unit.pct}
\alias{convert_unit.pct}
\title{Convert a numeric vector into a character vector with a percentage sign}
\usage{
\method{convert_unit}{pct}(var, x, decimal = 2, ...)
}
\arguments{
\item{var}{Same as in the generic function.}

\item{x}{Same as in the generic function.}

\item{decimal}{<\code{numeric}> How many decimal numbers should there be. Defaults
to 2.}

\item{...}{Additional arguments to be passed to the method.}
}
\value{
A character vector with a percentage sign.
}
\description{
Convert a numeric vector into a character vector with a percentage sign
}
