% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{cookie_retrieve}
\alias{cookie_retrieve}
\title{Retrieve the value of a cookie for the current Shiny session}
\usage{
cookie_retrieve(input, name)
}
\arguments{
\item{input}{This parameter represents the reactive list of all input values.
It allows the module to access and use the input values provided by the user
in the app. Should be \code{input = input}, as it's the same \code{input} than in the
shinyServer and moduleServer functions.}

\item{name}{<\code{character}> A character string specifying the name of the
cookie to be retrieved.}
}
\value{
The value of the specified cookie, or NULL if the cookie is not set.
}
\description{
This function retrieves the value of a cookie with the specified name from
the current Shiny session. The name of the cookie is specified as an argument
to the function. The value of the cookie is returned.
}
