% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{cookie_set}
\alias{cookie_set}
\title{Set and save a cookie value}
\usage{
cookie_set(session, name, value)
}
\arguments{
\item{session}{This parameter represents the session of the app. It allows the
module to interact with the app and perform various actions, such as updating
the UI or accessing user information.Should be \code{session = session}, as it's the
same \code{session} than in the shinyServer and moduleServer functions.}

\item{name}{<\code{character}> A character string specifying the name of the
cookie to be set, e.g. \code{lang}.}

\item{value}{<\code{string}> The value to be set for the cookie.}
}
\description{
This function sets a cookie value by sending a custom message to the
client-side JavaScript. The cookie name and value are specified as arguments
to the function and can be retrieved in ulterior session of the same user.
}
