% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{cookie_update_value}
\alias{cookie_update_value}
\title{Update the value of a cookie and return the updated value}
\usage{
cookie_update_value(input, name, current_value)
}
\arguments{
\item{input}{This parameter represents the reactive list of all input values.
It allows the module to access and use the input values provided by the user
in the app. Should be \code{input = input}, as it's the same \code{input} than in the
shinyServer and moduleServer functions.}

\item{name}{<\code{character}> A character string specifying the name of the cookie
to retrieve.}

\item{current_value}{<\code{character}> The current value of the reactive value
to update.}
}
\value{
The value of the cookie, or the \verb{current value} if there is no cookie
or if the current value of the variable is the same as the one of the cookie.
}
\description{
This function updates the value of a reactive variable. The current value of the
variable is passed as an argument to the function. If there is no cookie with
the specified name, the current value of the variable is returned
is returned. If the cookie exists and has a different value than the current
value, the cookie value is returned. This function is intended to be used to
update a reactive value in a Shiny app.
}
