% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_functions.R
\name{copy_current_url}
\alias{copy_current_url}
\title{Copy the current URL}
\usage{
copy_current_url()
}
\value{
Copies the URL.
}
\description{
This function lets the user copy the current URL. It uses the JS function
\code{copy_current_url} previous declared by \code{\link{use_curbcut_js}}.
}
