% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ui_server_mods.R
\name{create_ui_server_mods}
\alias{create_ui_server_mods}
\title{Assigns UI and server functions in the global environment for all pages in \code{modules}}
\usage{
create_ui_server_mods(modules)
}
\arguments{
\item{modules}{<\code{data.frame}> The data.frame containing all the pages information.}
}
\value{
This function assigns UI and server functions to the global
environment and returns \code{invisible()}.
}
\description{
This function takes the \code{modules} dataframe as input and creates UI and server
functions for every page it can. The UI function it assigns to the global environment
generates the layout of the page, while the server function handles the logic
behind it. The function then assigns the generated UI and server functions
to the global environment, using the page ID as the function name.
}
