% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get_helpers.R
\name{data_append_breaks}
\alias{data_append_breaks}
\title{Append new columns based on quantile breaks}
\usage{
data_append_breaks(var, data, q3_q5 = "q5", rename_col = "var_left")
}
\arguments{
\item{data}{<\code{data.frame}> Data frame to append columns to.}

\item{q3_q5}{<\code{character}> Specifies whether to use three or five quantiles.
Default is "q5".}

\item{rename_col}{<\code{character}> The column name to rename. Default is
"var_left". Can also be "var_right".}
}
\value{
<\code{data.frame}> Modified data frame with additional columns.
}
\description{
This function appends new columns to the input data based on quantile breaks
for specific variables. The quantile breaks are calculated using the function
\code{find_breaks_quintiles}. The function also preserves other attributes of the
original data.
}
