% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.bivar}
\alias{data_get.bivar}
\title{Get data from the SQLite database for a \code{bivar} class}
\usage{
\method{data_get}{bivar}(vars, df, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{df_get}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the two variables fresh out of the sqlite db,
binded in the same dataframe with an added \code{group} column for map colouring.
}
\description{
Get data from the SQLite database for a \code{bivar} class
}
