% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.bivar_ldelta_rq3}
\alias{data_get.bivar_ldelta_rq3}
\title{Get data from the SQLite database for a \code{bivar_ldelta_rq3} class}
\usage{
\method{data_get}{bivar_ldelta_rq3}(
  vars,
  df,
  scales_as_DA = c("building", "street"),
  data_path = "data/",
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_scale}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{data_path}{<\code{character}> A string representing the path to the directory
containing the QS files. Default is "data/".}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the percentage change between two
years of one variable with the static one year value of another variable.
\code{q3}for the percent change value is  calculated on the spot and used
to regroup with the \code{q3} column of the second value to create the \code{group}
column for map colouring.
}
\description{
Get data from the SQLite database for a \code{bivar_ldelta_rq3} class
}
