% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.delta}
\alias{data_get.delta}
\title{Get data from the SQLite database for a \code{delta} class}
\usage{
\method{data_get}{delta}(
  vars,
  df,
  scales_as_DA = c("building", "street"),
  data_path = "data/",
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the percentage change between two
years of the same variable. \code{q5} is calculated on the spot with and doubled
to the \code{group} column for map colouring.
}
\description{
Get data from the SQLite database for a \code{delta} class
}
