% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.delta_bivar}
\alias{data_get.delta_bivar}
\title{Get data from the SQLite database for a \code{delta_bivar} class}
\usage{
\method{data_get}{delta_bivar}(
  vars,
  df,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_scale}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{data_path}{<\code{character}> A string representing the path to the directory
containing the QS files. Default is "data/".}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the percentage change between two
years of two variables. \code{q3}s are calculated on the spot along with the
\code{group} column for the map colouring.
}
\description{
Get data from the SQLite database for a \code{delta_bivar} class
}
