% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R, R/data_get_colours.R
\name{data_get.q5}
\alias{data_get.q5}
\title{Get data from the SQLite database for a \code{q5} class}
\usage{
\method{data_get}{q5}(vars, region, ignore_scales = c("building", "street"), ...)

\method{data_get}{q5}(vars, region, ignore_scales = c("building", "street"), ...)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link[curbcut]{vars_build}} function.}

\item{...}{Additional arguments passed to methods.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link[curbcut]{df_get}}.}
}
\value{
A dataframe containing the data fresh out of the sqlite db, with an
added \code{group} column for map colouring.
}
\description{
Get data from the SQLite database for a \code{q5} class
}
