% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get}
\alias{data_get}
\title{Get data from the SQLite database}
\usage{
data_get(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link[curbcut]{vars_build}} function. The class of the vars object is
used to determine how to grab de data and output it.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the data according to the class of \code{vars}
along with a \code{group} column for map colouring.
}
\description{
This function retrieves data from an SQLite database using the appropriate
method based on the class of the input vars object. vars should be a named
list with a class, built using the \code{\link[curbcut]{vars_build}} function.
Depending on the class of vars, different methods will be used to retrieve
and process the data.
}
