% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get_colours.R
\name{data_get_colours}
\alias{data_get_colours}
\alias{data_get_colours.q5}
\alias{data_get_colours.bivar}
\alias{data_get_colours.delta}
\alias{data_get_colours.delta_bivar}
\alias{data_get_colours.bivar_ldelta_rq3}
\alias{data_get_colours.default}
\title{Retrieve colors for all IDs of a specified region}
\usage{
data_get_colours(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)

\method{data_get_colours}{q5}(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)

\method{data_get_colours}{bivar}(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)

\method{data_get_colours}{delta}(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)

\method{data_get_colours}{delta_bivar}(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)

\method{data_get_colours}{bivar_ldelta_rq3}(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)

\method{data_get_colours}{default}(
  vars,
  region,
  colours_table,
  scales_as_DA = c("building", "street"),
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link[curbcut]{vars_build}} function.}

\item{region}{<\code{character}> Region for which to get the \code{map_zoom_levels} and
retrieve all the according colours of all the IDs. One of the regions available
in the \code{regions_dictionary}.}

\item{colours_table}{<\code{character}> Fromn which colour table should the colour
be matched to the \code{group} column of the retrieved data. For \code{q5} class would be
\code{left_5}, for \code{bivar} class would be \code{bivar}, etc. One of the names of the
\code{colours_dfs} list present in the global environment.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales} that
should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A data frame with the columns \code{ID} and \code{fill} to use in
an \code{rdeck} scale function.
}
\description{
This function retrieves the colors of all ID of a specified region by matching
the \code{group} column in the data with the colors in the correct \code{colours}
data frame. The resulting data frame contains only the \code{ID} and \code{fill}
columns, to use in an \code{rdeck} scale function.
}
\seealso{
\code{\link[curbcut]{data_get_colours}}

\code{\link[curbcut]{data_get_colours}}

\code{\link[curbcut]{data_get_colours}}

\code{\link[curbcut]{data_get_colours}}

\code{\link[curbcut]{data_get_colours}}

\code{\link[curbcut]{data_get_colours}}
}
