% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get_colours.R
\name{data_get_colours_helper}
\alias{data_get_colours_helper}
\title{Helper function for \code{data_get_colours}}
\usage{
data_get_colours_helper(
  vars,
  region,
  time,
  zoom_levels,
  colours_table,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{region}{<\code{character}> Region for which to get the \code{map_zoom_levels} and
retrieve all the according colours of all the IDs. One of the regions available
in the \code{regions_dictionary}. Usually one of the output of \code{\link{geography_server}}.}

\item{time}{<\code{numeric}> The \code{time} at which data is displayed. A subset of the list
\code{time}. Only the numeric vector.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{geography_server}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-scale).}

\item{colours_table}{<\code{character}> Fromn which colour table should the colour
be matched to the \code{group} column of the retrieved data. For \code{q5} class would be
\code{left_5}, for \code{bivar} class would be \code{bivar}, etc. One of the names of the
\code{colours_dfs} list present in the global environment.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales} that
should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A data frame with the columns \code{ID} and \code{fill} to use in
an \code{cc.map::map_choropleth_fill_fun}-like scale function.
}
\description{
This function retrieves the colors of all ID of a specified region by matching
the \code{group} column in the data with the colors in the correct \code{colours}
data frame. The resulting data frame contains only the \code{ID} and \code{fill}
columns, to use in an \code{cc.map::map_choropleth_fill_fun}-like  function.
}
