% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_delta}
\alias{data_get_delta}
\title{Calculate the percentage change between two variables over two years}
\usage{
data_get_delta(
  vars,
  time,
  scale,
  vl_vr = "var_left",
  data_path = get_data_path()
)
}
\arguments{
\item{vars}{<\verb{character vector}> A var_code. The variable to get data for.}

\item{time}{<\verb{character vector}> A character vector of length 2. The
two years for which the delta should be calculated.}

\item{scale}{<\code{character}> A string specifying the scale at which to retrieve
data, corresponding to a path on disk, e.g. \code{DA} or \code{CSD}.}

\item{vl_vr}{<\code{character}> Which of var_left or var_right is this delta supposed
to be for. Defaults to var_left.}

\item{data_path}{<\code{character}> A string representing the path to the directory
containing the QS files. Default is "data/".}
}
\value{
A data frame with the following columns: ID, var_1, var_2, and var.
\code{ID} is the ID column from the original data, \code{var_1} and \code{var_2} are the
values of the two variables being compared, and \code{var} is the percentage
change between the two variables.
}
\description{
This function takes two variables representing the same quantity measured two
years apart and calculates the percentage change between the two values.
}
