% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_delta_fun}
\alias{data_get_delta_fun}
\title{Inner function to get data based on the type of \code{vars}}
\usage{
data_get_delta_fun(
  vars,
  select_id,
  df,
  data,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)
}
\description{
This function dispatches the data retrieval based on the class
of the \code{vars} object.
}
\seealso{
\code{\link{data_get.delta}}
}
