% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_delta_fun.scalar}
\alias{data_get_delta_fun.scalar}
\title{Retrieve \code{delta} data for scalar variables}
\usage{
\method{data_get_delta_fun}{scalar}(
  vars,
  df,
  scales_as_DA = c("building", "street"),
  data_path = "data/",
  ...
)
}
\description{
This function retrieves data for scalar variables
and performs additional operations for map coloring.
}
\seealso{
\code{\link{data_get.delta}}
}
