% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_sql}
\alias{data_get_sql}
\title{Retrieve data from a SQLite database table using SQL syntax}
\usage{
data_get_sql(var, df, select = "*")
}
\arguments{
\item{var}{<\code{character}> A string specifying the name of the table to retrieve
data from. Single variable (year) = single table. e.g. \code{housing_tenant_2016}}

\item{df}{<\code{character}> A string specifying the name of the database to retrieve
data from. Combination of the region and the scale, e.g. \code{CMA_DA}.}

\item{select}{A character vector specifying the names of the columns to
select from the table. Default is "*", which means all columns will be
selected. If select does not contain the "ID" column, it will be added to the query.}
}
\value{
A data.frame object with the selected data from the specified table.
}
\description{
This function retrieves data from a SQLite database table using SQL syntax.
It requires an established connection to the database, which should be stored
in the global environment with a specific naming convention: combination of
the region, the scale, and \verb{_conn}, e.g. "CMA_DA_conn" would be a valid
SQLite connection.
}
