% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{db_get}
\alias{db_get}
\title{Retrieve Data from Database}
\usage{
db_get(
  select = "*",
  from,
  where = NULL,
  schema = get_from_globalenv("inst_prefix")
)
}
\arguments{
\item{select}{<\verb{character vector}> The names of columns to select. Defaults
to everything.}

\item{from}{<\code{character}> The name of the table to select from.}

\item{where}{<\code{list}> Optional. A named list where names are column names
and values are the values those columns should match. Supports multiple values
for a single column.}

\item{schema}{<\code{character}> The database schema to use. Defaults to a
global environment variable \code{inst_prefix}.}
}
\value{
Returns a data frame containing the requested data.
}
\description{
Constructs and executes a SELECT SQL query to retrieve data from a specified
database table. Supports filtering through a WHERE clause.
}
