% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{db_get_helper}
\alias{db_get_helper}
\title{Execute a SQL Query Helper Function}
\usage{
db_get_helper(call)
}
\arguments{
\item{call}{\if{html}{\out{<\code{character}>}} The SQL query string to be executed.}
}
\value{
Returns the result set as a data frame.
}
\description{
A helper function that executes a SQL query using a generalized database
operation function. It wraps the \code{db_operation} function for specific use
with \code{DBI::dbGetQuery}.
}
