% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{db_operation}
\alias{db_operation}
\title{Perform a Database Operation}
\usage{
db_operation(what, args)
}
\arguments{
\item{what}{<\code{function}> The database operation to perform. This should be
a function capable of accepting a database connection and additional arguments.
Usually a DBI function, like \code{DBI::dbGetQuery}.}
}
\value{
Returns the output of the database operation.
}
\description{
This function abstracts the execution of database operations by checking out a
connection from a global connection pool, performing the specified operation,
and then returning the connection to the pool.
}
