% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delta_which_colors}
\alias{delta_which_colors}
\title{Determine Delta Colors Based on Data Class}
\usage{
delta_which_colors(data)
}
\arguments{
\item{data}{<\code{ANY}> The data object for which the color set needs to be
determined. The function checks for specific classes ('normal', 'negative',
'positive') in the data object to decide which color set to return.}
}
\value{
A color set from \code{colours_dfs}. If the class of \code{data} is 'normal',
\code{colours_dfs$delta} is returned. If the class is 'negative',
\code{colours_dfs$delta_neg} is returned. If the class is 'positive',
\code{colours_dfs$delta_pos} is returned. The return type is dependent on the
structure of \code{colours_dfs}.
}
\description{
This function selects the appropriate color set from a predefined set of
colors (\code{colours_dfs}) based on the class of the input data. It supports
three classes: 'normal', 'negative', and 'positive'. Depending on the class
of the data, it returns a corresponding set of delta colors.
}
