% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_make.R
\name{dropdown_make}
\alias{dropdown_make}
\title{Generate a dropdown list of variables (for use by the widget_picker)}
\usage{
dropdown_make(vars, compare = FALSE)
}
\arguments{
\item{vars}{<\verb{character vector}> The codes of variables to include.}

\item{compare}{<\code{logical}> Indicates if the dropdown is for use by the
compare panel. In that case, the first option will be an empty selection
meaning nothing (no comparison to display).}
}
\value{
A named list of variable codes grouped by theme for each category in
the dropdown.
}
\description{
This function generates a dropdown list of variables based on the input. The
output is a named list of variable codes grouped by theme for each category in
the dropdown.
}
