% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk_get.R
\name{dyk_get}
\alias{dyk_get}
\title{Get Did You Know Contents}
\usage{
dyk_get(id, vars, df, select_id, poi, lang = NULL)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the module will appear,
e.g. \code{alp}.}

\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{poi}{<\code{reactive}> (Optional) Point of interests. The output of
\code{\link{update_df}}. Default is NULL.}

\item{lang}{<\code{character}> A character string specifying the language to
translate the content. Defaults to NULL for no translation.}
}
\value{
An HTML list of DYK content with the "links" attribute containing the
necessary arguments for the link function (except r, which is added subsequently).
}
\description{
This function retrieves the content for the "Did You Know" (DYK) section of
a map page.
}
