% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk_functions.R
\name{dyk_link}
\alias{dyk_link}
\title{Create an HTML list element with a "Learn More" link and link attributes}
\usage{
dyk_link(id, element_id, page = id, text, lang = NULL, ...)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the DYK will appear,
e.g. \code{alp}.}

\item{element_id}{<\code{numeric}> This is used to create the HTML \verb{<a>} tag of
the \verb{[LEARN MORE]} button. It must be numeric, either 1 or 2. A click on
\code{dyk_1} and \code{dyk_2} only will trigger a reaction.}

\item{page}{<\code{character}> The target page for the link's action. Defaults
to \code{id}, stay on the same page.}

\item{text}{<\code{character}> Text to be displayed in the DYK.}

\item{lang}{<\code{character}> A character string specifying the language to
translate the content. Defaults to NULL for no translation.}

\item{...}{Additional attributes to pass to the \code{link} function. Any of
\code{select_id}, \code{var_right}, \code{date}, ...}
}
\value{
An HTML list item (\verb{<li>}) with an embedded action link. The list item
will also have attributes specifying that these are links.
}
\description{
This function generates an HTML list item (\verb{<li>}) containing the provided text
and an action link labeled "Learn More". The link appears as an action button.
The function also attaches attributes that specify that these are links.
}
