% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk_functions.R
\name{dyk_poi}
\alias{dyk_poi}
\title{Get Did You Know Contents for Points of Interest}
\usage{
dyk_poi(id, poi, lang)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the module will appear,
e.g. \code{alp}.}

\item{poi}{<\code{character}> TKTK}

\item{lang}{<\code{character}> A character string specifying the language to
translate the content. Defaults to NULL for no translation.}
}
\value{
An HTML list of DYK content with the "links" attribute containing the
necessary arguments for the link function (except r, which is added
subsequently).
}
\description{
This function retrieves POI content for the "Did You Know" (DYK) section of
a map page.
}
