% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk_functions.R
\name{dyk_text}
\alias{dyk_text}
\alias{dyk_text.default}
\alias{dyk_text.q5}
\title{Generate DYK text for the given variables and region}
\usage{
dyk_text(vars, df, select_id, lang, region, zoom_levels, scales_as_DA, ...)

\method{dyk_text}{default}(vars, df, select_id, lang, region, zoom_levels, scales_as_DA, ...)

\method{dyk_text}{q5}(vars, df, select_id, lang, region, zoom_levels, scales_as_DA, ...)
}
\arguments{
\item{vars}{<\code{character}> A list containing the variable names for which the
text needs to be generated. Usually the output of \code{\link{vars_build}}.}

\item{df}{<\code{character}> The combination of the region under study and the
scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{select_id}{<\code{character}> A string indicating the ID of the currently
selected region (if any). Usually \code{r[[id]]$select_id()}}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL. Usually is \code{r$lang()}.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{...}{Additional arguments passed to the dispatched function.}
}
\value{
The resulting text.
}
\description{
This function dispatches to the appropriate text-generating function based on
the type of \code{vars} and returns the resulting text.
}
