% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph.R
\name{explore_graph}
\alias{explore_graph}
\alias{explore_graph.q5_ind}
\alias{explore_graph.q5}
\alias{explore_graph.bivar}
\alias{explore_graph.delta_ind}
\alias{explore_graph.delta}
\alias{explore_graph.delta_bivar}
\alias{explore_graph.bivar_ind}
\title{Explore Graph Function}
\usage{
explore_graph(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{q5_ind}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{q5}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{bivar}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{delta_ind}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{delta}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{delta_bivar}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph}{bivar_ind}(
  vars,
  select_id,
  scale,
  data,
  time,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of variable codes with classes. The
output of \code{\link{vars_build}}.}

\item{select_id}{<\code{character}> A string indicating the ID of the currently
selected region (if any). Usually \code{r[[id]]$select_id()}}

\item{scale}{<\code{character}> Current scale. The output of
\code{\link{update_scale}}.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations to be compared. The output of \code{\link{data_get}}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their graph will be the one of their DA.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{font_family}{<\code{character}> A string specifying the font family for the
plot, default is "acidgrotesk-book".}

\item{...}{Additional arguments passed to the specific method.}
}
\value{
A ggplot2 object representing the plot.
}
\description{
This function creates a ggplot for the explore panel of Curbcut, based on the
input parameters.
}
