% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_helpers.R
\name{explore_graph_info}
\alias{explore_graph_info}
\title{Explore Graph Info Function}
\usage{
explore_graph_info(
  vars,
  font_family = "SourceSansPro",
  lang = NULL,
  df,
  scales_as_DA,
  select_id,
  data,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of variable codes with classes. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> A string specifying the font family for the
plot, default is "SourceSansPro".}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{df}{<\code{character}> The combination of the region under study and the
scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their graph will be the one of their DA.}

\item{select_id}{<\code{character}> A string indicating the ID of the currently
selected region (if any). Usually \code{r[[id]]$select_id()}}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations to be compared. The output of \code{\link{data_get}}.}

\item{...}{Additional arguments passed to the specific method.}
}
\value{
A list containing the default theme and color data frames.
}
\description{
This function returns shared information between the graphs, such as the
default theme and colors.
}
