% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_labels.R
\name{explore_graph_labels}
\alias{explore_graph_labels}
\alias{explore_graph_labels.delta}
\alias{explore_graph_labels.default}
\title{Generic legend label function for Curbcut legends}
\usage{
explore_graph_labels(vars, lang = NULL, ...)

\method{explore_graph_labels}{delta}(vars, lang = NULL, time, ...)

\method{explore_graph_labels}{default}(vars, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link{vars_build}}.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}

\item{time}{<\verb{named list}> A named list with the time variables.}
}
\value{
It returns an output of \code{\link[ggplot2]{labs}}.
}
\description{
\code{explore_graph_labels} is a generic function used to produce a list of
\code{\link[ggplot2]{labs}} for the Curbcut explore graph. The function invokes
particular methods which depend on the class of the \code{vars} argument.
}
