% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_helpers.R
\name{explore_graph_point_jitter}
\alias{explore_graph_point_jitter}
\title{Custom ggplot2 jitter function}
\usage{
explore_graph_point_jitter(dat, cols, ...)
}
\arguments{
\item{dat}{<\code{data.frame}> A data frame or tibble object that contains the data.}

\item{cols}{<\verb{character vector}> A character vector of length 2 specifying the
column names of 'dat' that will be used for the x and y coordinates, respectively.}

\item{...}{Additional arguments passed to the ggplot2::geom_jitter function.
Examples include 'alpha' for transparency, 'size' for point size, or 'color'
for point color. \code{ggplot2::aes()} also works here.}
}
\value{
A ggplot2::geom_jitter layer with customized width and height.
If the proportion of unique values to total observations in a given column
is less than 0.1, jitter will be applied along that axis (width for x,
height for y) to improve visualization. Otherwise, no jitter will be applied
along that axis.
}
\description{
This function applies a jitter effect to a ggplot2 scatter plot to better
visualize overlapping points. The jitter effect is dependent on the
proportion of unique values to total observations in the data.
}
