% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_scales.R
\name{explore_graph_scale}
\alias{explore_graph_scale}
\alias{explore_graph_scale.pct}
\alias{explore_graph_scale.dollar}
\alias{explore_graph_scale.default}
\title{Explore graph scale based on the variable type}
\usage{
explore_graph_scale(var, x_y, ...)

\method{explore_graph_scale}{pct}(var, x_y, ...)

\method{explore_graph_scale}{dollar}(var, x_y, data_vals, ...)

\method{explore_graph_scale}{default}(var, x_y, ...)
}
\arguments{
\item{var}{<\code{character}> The variable with its type as a class, either "pct",
"dollar", or any other type for the default scale.}

\item{x_y}{<\code{character}> x or y axis?}

\item{...}{Additional arguments passed to specific methods.}

\item{data_vals}{A numeric vector containing data values, required only for the
"dollar" method.}
}
\value{
A list containing ggplot2 x and y-axis scale functions.
}
\description{
This function provides a set of ggplot2 x-axis scales based on the variable type,
simplifying the process of creating axis scales for different types of data.
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_graph_scale(pct)}: The method for percentage values. Applies a
percentage scale to the x-axis.

\item \code{explore_graph_scale(dollar)}: The method for dollar values. Applies a
dollar scale to the x-axis with appropriate suffixes based on the magnitude
of the data.

\item \code{explore_graph_scale(default)}: The default method for unspecified variable
types. Applies a comma scale to the x-axis.

}}
