% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_scales.R
\name{explore_graph_scale_ind}
\alias{explore_graph_scale_ind}
\alias{explore_graph_scale_ind.scalar}
\alias{explore_graph_scale_ind.ordinal}
\title{Explore Graph Scale Indice}
\usage{
explore_graph_scale_ind(var, x_y, df, lang = NULL, ...)

\method{explore_graph_scale_ind}{scalar}(var, x_y, df, lang = NULL, data_vals, ...)

\method{explore_graph_scale_ind}{ordinal}(var, x_y, df, lang = NULL, ...)
}
\arguments{
\item{var}{<\code{character}> The variable with its type as a class, either "scale",
"ordinal".}

\item{x_y}{<\code{character}> x or y axis?}

\item{df}{<\code{character}> The combination of the region under study and the
scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{lang}{<\code{character} Which language should the label be translated to.
Defaults to NULL for no translation.}

\item{...}{Additional arguments passed to the underlying ggplot2 functions.}

\item{data_vals}{<\verb{numeric vector}> containing data values}
}
\value{
A list containing the ggplot2 scale object generated by the specialized
method.
}
\description{
This function generates a scale for a ggplot2 graph based on the variable \code{var}
and the coordinate axis specified in \code{x_y}. The function dispatches to
specialized methods based on the type of \code{var}, either scalar or ordinal.
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_graph_scale_ind(scalar)}: scalar method for index values

\item \code{explore_graph_scale_ind(ordinal)}: ordinal method for index values

}}
