% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_module.R
\name{explore_server}
\alias{explore_server}
\alias{explore_UI}
\title{Explore Server}
\usage{
explore_server(
  id,
  r,
  data,
  vars,
  region,
  df,
  select_id,
  scales_as_DA = shiny::reactive(c("building", "street")),
  graph = shiny::reactive(explore_graph),
  graph_args = shiny::reactive(list(r = r, data = data(), vars = vars(), df = df(),
    select_id = select_id(), region = region(), scales_as_DA = scales_as_DA(), lang =
    r$lang())),
  table = shiny::reactive(explore_text),
  table_args = shiny::reactive(list(r = r, data = data(), vars = vars(), select_id =
    select_id(), region = region(), scales_as_DA = scales_as_DA(), df = df(), lang =
    r$lang()))
)

explore_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{data}{<\verb{reactive data.frame}> Data frame containing all the scale and
the \code{var_left} and \code{var_right}. The output of \code{\link{data_get}}.}

\item{vars}{<\verb{reactive named list}> Named list with a class. Object built
using the \code{\link{vars_build}} function. The class of the vars object is
used to determine which type of legend to draw.}

\item{region}{<\verb{reactive character}> A string or numeric value representing the
selected region.}

\item{df}{<\verb{reactive character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{select_id}{<\verb{reactive character}> the current selected ID, usually
\code{r[[id]]$select_id()}.}

\item{scales_as_DA}{<\verb{reactive character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{graph}{<\verb{reactive function}> A reactive expression for the function used
to generate the explore graph in the explore module. The default value is
\code{shiny::reactive(explore_graph)}, which is a reactive expression that
evaluates to the \code{explore_graph} function. If you want to use a different
function to generate the explore graph, you can provide a custom reactive
expression for this argument. The function you provide should have the same
arguments as graph_args.}

\item{graph_args}{<\verb{reactive list}> A reactive expression for the arguments
to be passed to the explore graph function. If you want to customize the arguments
passed to the explore graph function, you can provide a different reactive
expression for this argument. The reactive expression should evaluate to a
list of arguments to be passed to the explore graph function, including any
additional arguments required by your custom graph function.}

\item{table}{<\verb{reactive function}> A reactive expression for the function used
to generate the info table in the explore module. The default value is
\code{shiny::reactive(explore_text)}, which is a reactive expression that
evaluates to the \code{explore_text} function. If you want to use a different
function to generate the info table, you can provide a custom reactive
expression for this argument. The function you provide should have the same
arguments as table_args.}

\item{table_args}{<\verb{reactive list}> A reactive expression for the arguments
to be passed to the info table function. If you want to customize the arguments
passed to the info table function, you can provide a different reactive
expression for this argument. The reactive expression should evaluate to a
list of arguments to be passed to the info table function, including any
additional arguments required by your custom table function.}
}
\value{
The explore Shiny UI and server module functions
}
\description{
This function creates the server for the explore module. The module displays
an info table and a graph of the selected variables and geographic region.
The explore_server function has an argument named table, which is a reactive
expression for the function used to generate the info table in the explore
module. By default, this argument is set to \code{shiny::reactive(explore_text)},
which is a reactive expression that evaluates to the explore_text function.
If you want to substitute the table argument with another function, you can
simply pass a different reactive expression for this argument. The function
you provide should have the same arguments as table_args, which is another
argument of the explore_server function that is a reactive expression for
the arguments to be passed to the info table function. Same goes for the \code{graph}
and \code{graph_args} arguments.
}
\section{Functions}{
\itemize{
\item \code{explore_UI()}: Create the UI for the explore module

}}
