% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text}
\alias{explore_text}
\title{Generate text for the given variables and region}
\usage{
explore_text(vars, ...)
}
\arguments{
\item{vars}{<\code{character}> A list containing the variable names for which the
text needs to be generated. Usually the output of \code{\link{vars_build}}.}

\item{...}{Additional arguments passed to the dispatched function.}
}
\value{
The resulting text.
}
\description{
This function dispatches to the appropriate text-generating function based on
the variable type and returns the resulting text.
}
