% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_bivar_helpers.R
\name{explore_text_bivar_correlation}
\alias{explore_text_bivar_correlation}
\title{Function for exploring bivariate correlation between two variables}
\usage{
explore_text_bivar_correlation(vars, data, time, lang = NULL)
}
\arguments{
\item{vars}{<\code{character}> A list containing the variable names for which the
text needs to be generated. Usually the output of \code{\link{vars_build}}.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations. The output of \code{\link{data_get}}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL.}
}
\value{
A list containing the correlation coefficient, a boolean indicating
whether the correlation is positive or negative, a text string describing
the strength and direction of the correlation, and a text string describing
the relationship between the variables.
}
\description{
This function calculates bivariate correlation between two variables and
returns a list containing the correlation coefficient, a boolean indicating
whether the correlation is positive or negative, a text string describing the
strength and direction of the correlation, and a text string describing the
relationship between the variables.
}
