% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_bivar_correlation_helper}
\alias{explore_text_bivar_correlation_helper}
\alias{explore_text_bivar_correlation_helper.scalar}
\alias{explore_text_bivar_correlation_helper.ordinal}
\title{Helper function for exploring bivariate correlation between variables}
\usage{
explore_text_bivar_correlation_helper(vars, data, time, lang = NULL, ...)

\method{explore_text_bivar_correlation_helper}{scalar}(vars, data, time, lang = NULL, ...)

\method{explore_text_bivar_correlation_helper}{ordinal}(vars, data, time, lang = NULL, ...)
}
\arguments{
\item{vars}{<\code{character}> A list containing the variable names for which the
text needs to be generated. Usually the output of \code{\link{vars_build}}.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations. The output of \code{\link{data_get}}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL.}

\item{...}{Additional arguments to be passed.}
}
\value{
A list containing the correlation coefficient and a formatted string.
}
\description{
This function calculates bivariate correlation between two text variables
using either Pearson's correlation coefficient or Spearman's rank correlation
coefficient based on the type of variables passed as arguments. The result is
returned as a list containing the correlation coefficient and a formatted
string.
}
