% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_color}
\alias{explore_text_color}
\title{Explore Text Color Function}
\usage{
explore_text_color(x, meaning)
}
\arguments{
\item{x}{<\code{character}> A character string to which a color will be assigned.}

\item{meaning}{<\code{character}> A character string indicating the color meaning.
Can be one of the following: "left", "right", "increase", or "decrease".
Default is "left".}
}
\value{
A character string with an HTML span tag containing the input text x and its
assigned color as an inline style.
}
\description{
This function takes a text input x and assigns it a color based on the meaning
argument. The colors are determined using the global environment's color
schemes for bivariate and delta categories.
}
