% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text_delta_bivar_adjective}
\alias{explore_text_delta_bivar_adjective}
\alias{explore_text_delta_bivar_adjective.default}
\title{Helper function for generating adjective to describe delta/bivariate
relationship between text variables}
\usage{
explore_text_delta_bivar_adjective(var, left, positive, lang = NULL, ...)

\method{explore_text_delta_bivar_adjective}{default}(var, left, positive, lang, ...)
}
\arguments{
\item{var}{<\code{character}> The variable code for which the text needs to be
generated. \code{vars$var_left} or \code{vars$var_right}}

\item{left}{<\verb{logical>} Whether the \code{var} supplied is the var_left
or the \code{var_right}. If \code{var_left}, TRUE.}

\item{positive}{<\code{logical}> Wheter the bivariate relationship is positive
or negative. One of the output of \code{\link{explore_text_bivar_correlation}}}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL.}

\item{...}{Additional arguments to be passed.}
}
\value{
A text string containing an adjective to describe the bivariate
relationship.
}
\description{
This function generates a text string containing an adjective to describe the
delta/bivariate relationship between two variables based on whether the relationship
is positive or negative.
}
