% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_delta_helpers.R
\name{explore_text_delta_exp}
\alias{explore_text_delta_exp}
\alias{explore_text_delta_exp.ind}
\alias{explore_text_delta_exp.default}
\title{Explore text \code{delta} explanation and values for}
\usage{
explore_text_delta_exp(
  var,
  region,
  select_id,
  left_right = "left",
  time,
  scale,
  data,
  schemas,
  lang,
  ...
)

\method{explore_text_delta_exp}{ind}(
  var,
  region,
  select_id,
  left_right = "left",
  time,
  scale,
  data,
  schemas,
  lang,
  val = NULL,
  ...
)

\method{explore_text_delta_exp}{default}(
  var,
  region,
  select_id,
  left_right = "left",
  time,
  scale,
  data,
  schemas,
  lang,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The variable code for which the text and values need
to be generated. Usually \code{vars$var_left}.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{select_id}{A string indicating the ID of the currently selected region
(if any). Usually \code{r[[id]]$select_id()}}

\item{left_right}{<\code{character}> Is it a left or right variable? Possible
options are "left" or "right".}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{scale}{<\code{character}> The scale under study. The output of
\code{\link{update_scale}}.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations. The data frame must have columns named var_left
and ID. The output of \code{\link{data_get}}.}

\item{schemas}{<\verb{named list}> Current schema information. The additional widget
values that have an impact on which data column to pick. Usually \code{r[[id]]$schema()}.}

\item{lang}{<\code{character}> Language for translation.}

\item{...}{Additional arguments passed to the dispatched method.}

\item{val}{<\code{numeric}> If the value is not part of \code{data}. It happens on raster
data where we show region values for the highest resolution possible, but we still
want to allow user to select grid cells of lower resolutions. Defaults to NULL
for normal operations.}
}
\value{
A list with the following elements:
\item{exp}{A character string describing the explanation of the variable}
\item{region_vals}{A numeric vector containing the variable values in the
current region and time period.}
\item{region_vals_strings}{A character vector with the same length as
\code{region_vals}, containing the variable values formatted according
to the class of \code{var}}
}
\description{
This function delivers the delta text and values of a given variable when
compared between two yeas. It dispatches to specific methods depending on the
structure of the variable.
}
