% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_delta_helpers.R
\name{explore_text_delta_first_p}
\alias{explore_text_delta_first_p}
\alias{explore_text_delta_first_p.ind}
\alias{explore_text_delta_first_p.default}
\title{Generate Text for the first paragraph of Delta}
\usage{
explore_text_delta_first_p(
  var,
  context,
  exp_vals,
  lang = NULL,
  change_string,
  ...
)

\method{explore_text_delta_first_p}{ind}(
  var,
  context,
  exp_vals,
  lang,
  change_string,
  select_id,
  ...
)

\method{explore_text_delta_first_p}{default}(var, context, exp_vals, lang, change_string, ...)
}
\arguments{
\item{var}{<\code{character}> The variable code for which the text and values need
to be generated. Usually \code{vars$var_left}.}

\item{context}{<\code{list}> The output of \code{\link{explore_context}}}

\item{exp_vals}{<\code{list}> The output of \code{\link{explore_text_delta_exp}}}

\item{lang}{<\code{character}> Language specifying the language of the generated text.
Either \code{fr} or \code{en}. Defaults to NULL for no translation.}

\item{change_string}{<\code{list}> The output of \code{\link{explore_text_delta_change}}}

\item{...}{Additional arguments passed to the methods.}

\item{select_id}{A string indicating the ID of the currently selected region
(if any). Usually \code{r[[id]]$select_id()}}
}
\value{
A character string containing the generated text.
}
\description{
This function generates text for the first paragraph of delta of a given variable,
comparing values at two different points in time.
}
