% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_parent_title}
\alias{explore_text_parent_title}
\title{Get parent title of a variable}
\usage{
explore_text_parent_title(var, lang = NULL)
}
\arguments{
\item{var}{<\code{character}> The variable name for which the parent title needs
to be retrieved.}

\item{lang}{<\code{character}> Active language. "en" or "fr". Defaults to NULL
for no translation.}
}
\value{
The title of the parent variable in lowercase.
}
\description{
This function retrieves the title of the parent variable given the variable
name and returns it in lowercase.
}
