% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_region_val_df}
\alias{explore_text_region_val_df}
\title{Get region values data frame}
\usage{
explore_text_region_val_df(var, region, select_id, left = TRUE, ...)
}
\arguments{
\item{var}{<\code{character}> The variable name for which the region values data
frame needs to be retrieved.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{select_id}{<\code{character}> the current selected ID, usually
\code{r[[id]]$select_id()}.}

\item{left}{<\code{logical}> Whether the variable to grab data for is the \code{var_left}
or the \code{var_right}. It will impact on which column of \code{data} is selected
to grab the information.}

\item{...}{Additional arguments for the \code{\link{explore_text_select_val}}
function: \itemize{
\item{data <\code{data.frame}>}{The output of \code{\link{data_get}}.}
\item{df <\code{character}>}{The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}
}}
}
\value{
The resulting data frame after subsetting or list when there is a
selection.
}
\description{
This function retrieves the region values data frame given the variable name
and region and subsets the data frame based on the region name. If the variable
is a year, then it filters the row based on the year. It then returns the
resulting data frame.
}
\details{
If there is a selection, then
}
