% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_select_val_ind}
\alias{explore_text_select_val_ind}
\alias{explore_text_select_val_ind.scalar}
\alias{explore_text_select_val_ind.ordinal}
\title{Generate values for the given \code{ind} variable and selection}
\usage{
explore_text_select_val_ind(
  var,
  data,
  select_id,
  col = "var_left",
  time,
  lang,
  schemas = NULL,
  val = NULL,
  ...
)

\method{explore_text_select_val_ind}{scalar}(
  var,
  data,
  select_id,
  col = "var_left",
  time,
  lang,
  schemas = NULL,
  val = NULL,
  ...
)

\method{explore_text_select_val_ind}{ordinal}(
  var,
  data,
  select_id,
  col = "var_left",
  time,
  lang,
  schemas = NULL,
  val = NULL,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The variable code of the variable for which the
values need to be generated. Usually one element of the output of
\code{\link{vars_build}}.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations. The output of \code{\link{data_get}}.}

\item{select_id}{<\code{character}> The ID of the selected zone.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{lang}{<\code{character}> Language the ranking character should be translated
to. Defaults to NULL for no translation.}

\item{schemas}{<\verb{named list}> Current schema information. The additional widget
values that have an impact on which data column to pick. Usually \code{r[[id]]$schema()}.}

\item{val}{<\code{numeric}> If the value is not part of \code{data}. It happens on raster
data where we show region values for the highest resolution possible, but we still
want to allow user to select grid cells of lower resolutions. Defaults to NULL
for normal operations.}

\item{...}{Additional arguments passed to the dispatched function.}
}
\value{
The resulting values
}
\description{
Generate values for the given \code{ind} variable and selection
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_text_select_val_ind(scalar)}: Method for \code{scalar}

\item \code{explore_text_select_val_ind(ordinal)}: Method for \code{ordinal}

}}
