% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_select_val.R
\name{explore_text_select_val_postgresql}
\alias{explore_text_select_val_postgresql}
\alias{explore_text_select_val_postgresql.pct}
\alias{explore_text_select_val_postgresql.ind}
\alias{explore_text_select_val_postgresql.default}
\title{Method to generate values for the given variable and selection (for data
on the postgresql server)}
\usage{
explore_text_select_val_postgresql(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  ...
)

\method{explore_text_select_val_postgresql}{pct}(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  ...
)

\method{explore_text_select_val_postgresql}{ind}(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  data,
  lang = NULL,
  ...
)

\method{explore_text_select_val_postgresql}{default}(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{scale}{<\code{character}> The scale of interest.}

\item{parent_string}{<\code{character}> The parent string as character ex. c_population}

\item{var_schema}{<\code{character}> With the schema on ex. alp_2021}

\item{parent_string_schema}{<\code{character}> With the schema on ex. c_population_2021}

\item{select_id}{<\code{character}> The selected identifier.}

\item{...}{Additional arguments (e.g., data, time) used in specific methods.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{lang}{<\code{character}> Active language. \code{"en"} or \code{"fr"}}
}
\value{
The resulting values
}
\description{
This function dispatches to the appropriate value-generating function based on
the variable type and returns the resulting values. It is only used when
there is a selection, and replaces the value of \code{\link{explore_text_region_val_df}}
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_text_select_val_postgresql(pct)}: Method for pct

\item \code{explore_text_select_val_postgresql(ind)}: Method for \code{ind}

\item \code{explore_text_select_val_postgresql(default)}: Default method

}}
