% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_select_val.R
\name{explore_text_select_val_postgresql_ind}
\alias{explore_text_select_val_postgresql_ind}
\alias{explore_text_select_val_postgresql_ind.scalar}
\alias{explore_text_select_val_postgresql_ind.ordinal}
\title{Generate values for the given \code{ind} variable and selection (for posgresql)}
\usage{
explore_text_select_val_postgresql_ind(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  data,
  lang = NULL,
  ...
)

\method{explore_text_select_val_postgresql_ind}{scalar}(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  data,
  lang = NULL,
  ...
)

\method{explore_text_select_val_postgresql_ind}{ordinal}(
  var,
  scale,
  parent_string,
  var_schema,
  parent_string_schema,
  select_id,
  data,
  lang = NULL,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{scale}{<\code{character}> The scale of interest.}

\item{parent_string}{<\code{character}> The parent string as character ex. c_population}

\item{var_schema}{<\code{character}> With the schema on ex. alp_2021}

\item{parent_string_schema}{<\code{character}> With the schema on ex. c_population_2021}

\item{select_id}{<\code{character}> The selected identifier.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{lang}{<\code{character}> Active language. \code{"en"} or \code{"fr"}}

\item{...}{Additional arguments passed to the dispatched function.}
}
\value{
The resulting values
}
\description{
Generate values for the given \code{ind} variable and selection (for posgresql)
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_text_select_val_postgresql_ind(scalar)}: Method for scalar

\item \code{explore_text_select_val_postgresql_ind(ordinal)}: Method for ordinal

}}
