% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_selection_comparison}
\alias{explore_text_selection_comparison}
\title{Explore Text Selection Comparison}
\usage{
explore_text_selection_comparison(var, data, select_id, left = TRUE)
}
\arguments{
\item{var}{<\code{character}> A variable code specifying the variable of interest. This
variable will be compared across observations.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations to be compared. The data frame must have columns named var_left
and ID. The output of \code{\link{data_get}}.}

\item{select_id}{<\code{character}> The ID of the selected zone for which to
retrieve the ranking.}

\item{left}{<\code{logical}> Whether the variable to grab data for is the \code{var_left}
or the \code{var_right}. It will impact on which column of \code{data} is selected
to grab the information.}
}
\value{
A named list with two elements:
\itemize{
\item \code{higher_than}: A character string representing the proportion of
other observations with a lower value for the specified variable than the
selected observation, formatted as a percentage with one decimal place.
\item \code{rank_chr}: A character string representing the ranking category
of the selected observation in comparison to the other observations for the
specified variable.
}
}
\description{
This function calculates the percentage of observations with a lower value
than the selected observation for a given variable and ranks the selected
observation within predefined groups (as a character extracted from the
\code{variables} table.
}
