% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text_values_q5.pct}
\alias{explore_text_values_q5.pct}
\title{Generate text for the given variables and region - q5 version using percentage}
\usage{
\method{explore_text_values_q5}{pct}(
  var,
  region,
  data,
  df,
  select_id,
  col = "var_left",
  lang,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The variable name for which the text needs to be
generated. Usually \code{vars$var_left}}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{data}{<\code{data.frame}> The output of \code{\link{data_get}}.}

\item{df}{<\code{character}>The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{select_id}{<\code{character}> the current selected ID, usually
\code{r[[id]]$select_id()}.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL. Usually is \code{r$lang()}.}

\item{...}{Additional arguments passed to the function.}
}
\value{
The resulting text.
}
\description{
This function generates text for the given variables and region using the
q5 version and percentage. It returns the resulting text.
}
