% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_values_q5.R
\name{explore_text_values_q5}
\alias{explore_text_values_q5}
\alias{explore_text_values_q5.pct}
\alias{explore_text_values_q5.count}
\alias{explore_text_values_q5.dollar}
\alias{explore_text_values_q5.ind}
\alias{explore_text_values_q5.avg}
\alias{explore_text_values_q5.sqkm}
\alias{explore_text_values_q5.per1k}
\alias{explore_text_values_q5.ppo}
\title{Generate text for the given variables and region - q5 version}
\usage{
explore_text_values_q5(var, region, ...)

\method{explore_text_values_q5}{pct}(
  var,
  region,
  data,
  scale,
  select_id,
  col = "var_left",
  lang,
  time,
  ...
)

\method{explore_text_values_q5}{count}(
  var,
  region,
  data,
  scale,
  select_id,
  col = "var_left",
  lang,
  time,
  ...
)

\method{explore_text_values_q5}{dollar}(
  var,
  region,
  data,
  scale,
  select_id,
  col = "var_left",
  lang,
  time,
  ...
)

\method{explore_text_values_q5}{ind}(
  var,
  region,
  select_id,
  data,
  scale,
  col = "var_left",
  lang,
  time,
  ...
)

\method{explore_text_values_q5}{avg}(
  var,
  region,
  select_id,
  data,
  scale,
  col = "var_left",
  lang,
  time,
  ...
)

\method{explore_text_values_q5}{sqkm}(
  var,
  region,
  select_id,
  data,
  scale,
  col = "var_left",
  lang,
  time,
  ...
)

\method{explore_text_values_q5}{per1k}(
  var,
  region,
  select_id,
  data,
  scale,
  col = "var_left",
  lang = lang,
  time,
  ...
)

\method{explore_text_values_q5}{ppo}(
  var,
  region,
  select_id,
  data,
  scale,
  col = "var_left",
  lang,
  time,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The variable name for which the text needs to be
generated. Usually \code{vars$var_left}}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{...}{Additional arguments passed to the function..}

\item{data}{<\code{data.frame}> The output of \code{\link{data_get}}.}

\item{scale}{<\code{character}> Current scale. The output of
\code{\link{update_scale}}.}

\item{select_id}{<\code{character}> the current selected ID, usually
\code{r[[id]]$select_id()}.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL. Usually is \code{r$lang()}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}
}
\value{
The resulting text.
}
\description{
This function generates text for the given variables and region using the
Q5 version. It dispatches to the appropriate text-generating function based on
the variable type and returns the resulting text.
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_text_values_q5(pct)}: The method for percentage values.

\item \code{explore_text_values_q5(count)}: The method for count values.

\item \code{explore_text_values_q5(dollar)}: The method for dollar values.

\item \code{explore_text_values_q5(ind)}: The method for \code{ind} values.

\item \code{explore_text_values_q5(avg)}: The method for average values.

\item \code{explore_text_values_q5(sqkm)}: The method for square kilometers values.

\item \code{explore_text_values_q5(per1k)}: The method for per1k values.

\item \code{explore_text_values_q5(ppo)}: The method for people per object values.

}}
