% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_name_2}
\alias{fill_name_2}
\title{Get name_2 based on biven ID scale, scale, and top scale}
\usage{
fill_name_2(ID_scale, scale, top_scale)
}
\arguments{
\item{ID_scale}{<\verb{character vector}> A vector of IDs from the scale table
to keep.}

\item{scale}{<\code{character}> The name of the scale table in the global environment.}

\item{top_scale}{<\code{character}> The name of the top scale table in the global
environment (from which the column \code{name} will be used as the output).}
}
\value{
A character vector of names in the order of 'ID_scale'.
}
\description{
This function retrieves the name from a top scale table and attaches
it to a scale table based on given ID scales. It uses efficient
matching instead of the slower merge operation.
}
