% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{filter_region}
\alias{filter_region}
\title{Filter data by region}
\usage{
filter_region(data, scale, region)
}
\arguments{
\item{data}{<\code{data.frame}> Data.frame containing the data to be filtered.
It must have an \code{ID} column that matches the IDs in \code{regions_dictionary}.}

\item{scale}{<\code{character}> The scale at which to filter the data (e.g., 'CSD', 'CT').}

\item{region}{<\code{character}> The code of the region for filtering (e.g., 'CMA').}
}
\value{
Returns a filtered data frame containing only the rows corresponding
to the specified region and scale.
}
\description{
Filters a given data frame by the region and scale specified. The function
grabs a global variable \code{regions_dictionary} to identify the IDs
associated with the given region and scale.
}
