% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get_helpers.R
\name{find_breaks_quintiles}
\alias{find_breaks_quintiles}
\title{Find quintile breaks}
\usage{
find_breaks_quintiles(dist, q3_q5 = "q5")
}
\arguments{
\item{dist}{<\code{numeric}> Distribution (numerics) with no NAs.}

\item{q3_q5}{<\code{character}> How many bins? \code{q3} or \code{q5}.}
}
\value{
Returns a numeric vector with quintile breaks.
}
\description{
Find quintile breaks
}
