% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_closest_scale_not_in_db}
\alias{find_closest_scale_not_in_db}
\title{Find the scale closer to current scale that is available ondisk}
\usage{
find_closest_scale_not_in_db(
  scale,
  zoom_levels,
  db_scales = get_from_globalenv("db_scales")
)
}
\arguments{
\item{scale}{<\code{character}> The scale at which the search for the previous
element starts. It must exist within 'zoom_levels'.}

\item{zoom_levels}{<\verb{character vector}> A vector containing different zoom
levels. It should include the 'scale' value.}

\item{db_scales}{<\verb{character vector}> A vector specifying which scales
are in the postgresql database}
}
\value{
<\code{character}> The element preceding 'scale' that is not in 'db_scales'.
}
\description{
This function searches through the names of the'zoom_levels' vector to find the first
element immediately preceding a specified 'scale' that is not present in
the 'db_scales' vector (scales in the database).
}
