% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_outliers}
\alias{find_outliers}
\alias{remove_outliers_df}
\title{Find and remove outliers in specified columns of a data frame.}
\usage{
find_outliers(
  x,
  lower_bracket = 0.02,
  higher_bracket = 0.98,
  iqr_multiplier = 1.5
)

remove_outliers_df(df, cols)
}
\arguments{
\item{x}{<\code{numeric}> A numeric vector for which to find outliers.}

\item{lower_bracket}{<\code{numeric}> The lower percentile to use for the
IQR calculation. Defaults to 0.02.}

\item{higher_bracket}{<\code{numeric}> The higher percentile to use for the
IQR calculation. Defaults to 0.98.}

\item{iqr_multiplier}{<\code{numeric}> Multiplier for the IQR to determine
outlier thresholds. Defaults to 1.5.}

\item{df}{<\code{data.frame}> A data frame from which to remove outliers.}

\item{cols}{<\verb{character vector}> Column names in df for which to remove o
utliers.}
}
\value{
For find_outliers(): A vector of indices of the outliers in x.
For remove_outliers_df(): A data frame with outliers removed from specified
columns.
}
\description{
This function package consists of two functions: find_outliers() and
remove_outliers_df(). find_outliers() identifies outliers in a numeric vector
using the 1.5 * IQR rule. remove_outliers_df() removes outliers in specified columns
of a data frame, preserving the original row indices.
}
\examples{
x <- c(1, 2, 3, 4, 5, 100)
find_outliers(x)

df <- data.frame(a = c(1, 2, 3, 4, 5, 100), b = c(1, 2, 3, 4, 5, 6))
remove_outliers_df(df, cols = "a")

}
