% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_module.R
\name{geography_server}
\alias{geography_server}
\alias{geography_UI}
\title{Geography module for CUrbcut}
\usage{
geography_server(id, r = r, regions, avail_scale_combinations)

geography_UI(id, regions, avail_scale_combinations)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{regions}{<\verb{character vector}> Possible available regions for selection.}

\item{avail_scale_combinations}{<\verb{character vector}> Available scale combinations
for the top scale.}
}
\value{
A reactive list containing selected region and calculated zoom levels.
}
\description{
This function creates a shiny module server to handle geography settings.
It initializes two dropdown menus for selecting regions and top scale.
A calculated map zoom level is also provided.
}
\section{Functions}{
\itemize{
\item \code{geography_UI()}: Create the UI for the zoom module

}}
