% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_events.R
\name{get_click}
\alias{get_click}
\title{Get the Map Click as a Reactive}
\usage{
get_click(map_ID)
}
\arguments{
\item{map_ID}{<\code{character}> The unique identifier of the map whose view state
is to be retrieved. This ID is used to monitor changes in the map's view state.}
}
\value{
A reactive value containing the most recent click event on the map. This reactive value
updates whenever the \code{cc.map::get_map_click} function produces a non-null output.
}
\description{
This function retrieves the map click by observing changes in the
\code{cc.map::get_map_click} function. The map view state is then stored and
returned as a reactive value.
}
