% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_path}
\alias{get_data_path}
\title{Get the data path for Curbcut}
\usage{
get_data_path()
}
\value{
A character string representing the path where data is stored.
}
\description{
This function retrieves the data path for testing purposes. It looks for the
environment variable \code{CURBCUT_DATA} and should return a path to a Curbcut
data folder if it exists. If the environment variable is not set, which is
what happened in production, it returns "data/" as a fallback.
}
