% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dist}
\alias{get_dist}
\title{Calculate the distance between two points on the Earth's surface using the
Haversine formula}
\usage{
get_dist(x, y)
}
\arguments{
\item{x}{A vector or matrix/data frame containing the longitude and latitude
of one or more points}

\item{y}{A vector containing the longitude and latitude of a single point}
}
\value{
The distance between the two points in meters
}
\description{
The function calculates the distance between two points on the Earth's
surface using the Haversine formula. If \code{x} is a matrix or data frame,
the function extracts the longitude and latitude vectors for each point.
The latitude values are converted from degrees to radians, and the function
calculates the central angle between the two points using the Haversine formula.
The distance between the points in meters is then calculated and returned.
}
