% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_from_globalenv}
\alias{get_from_globalenv}
\title{Get object from global environment}
\usage{
get_from_globalenv(x)
}
\arguments{
\item{x}{<\code{character}> The name of the object to retrieve from the global
environment.}
}
\value{
The requested object from the global environment.
}
\description{
This function retrieves an object from the global environment by the name of
the object. If the object is not found in the global environment, an error is
thrown.
}
