% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab_df_from_bslike}
\alias{grab_df_from_bslike}
\title{Grab full dataframe from a building-street-like dataframe (large!)}
\usage{
grab_df_from_bslike(scale)
}
\arguments{
\item{scale}{<\code{character}>  A string, the name of the dataframe in which to look
for full dataframe. The dataframe should be in the global environment, and if it isn't,
there must be an established sqlite connection to it.}
}
\value{
The full datraframe corresponding to the given \code{scale}. If
\code{select_id} is found in the \code{scale} in the global environment, the corresponding
table is returned. If \code{scale} is not in the global
environment, the function fetches the dataframe from the connected database.
}
\description{
This function fetches the full dataframe matching with \code{df}. It can also fetch
the full dataframe from a connected database when the \code{df} is not found in the
global environment.
}
