% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grab_row_from_bslike}
\alias{grab_row_from_bslike}
\title{Grab row from a building-street-like dataframe (large!)}
\usage{
grab_row_from_bslike(scale, select_id, cols = "*")
}
\arguments{
\item{scale}{<\code{character}> A string, the name of the dataframe in which to look
for row. The dataframe should be in the global environment, and if it isn't,
there must be an established sqlite connection to it.}

\item{select_id}{<\code{character}> A value representing the ID that needs to be
selected in order to fetch the corresponding row.}

\item{cols}{<\verb{character vector}> Which columns to retrieve. Defaults to \code{'*'} for
all columns.}
}
\value{
The row corresponding to the given \code{select_id}. If
\code{select_id} is found in the \code{df} in the global environment, the corresponding
row is returned. If \code{df} is a 'scales_as_DA' and not in the global
environment, the function fetches row from the connected database.
}
\description{
This function fetches the row matching with the selected ID from a given
dataframe \code{scale} based on a provided \code{select_id}. It can also fetch the row
from a connected database when the \code{df} is not found in the global environment
but is identified as a scales_as_DA'. In such cases, it uses the established
connection and fetches the row via a SQL query.
}
