% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat}
\alias{heartbeat}
\title{Heartbeat Function}
\usage{
heartbeat(input)
}
\arguments{
\item{input}{<\code{input}> A reactive input object that contains all of the input
values from the user interface. As this function will be placed in \code{server.R}
in the \code{\link[shiny]{shinyServer}} function, this will usually be
\code{input = input}.}
}
\value{
The function returns a Shiny observer that triggers every 10 seconds
when \code{input} haven't changed to keep the app alive. After 2 hours, the
heartbeat will stop and the app can disconnect with usual Shiny timeout.
}
\description{
This function is used to create a "heartbeat" that keeps the Curbcut app
running in the background even if the user is not interacting with it.
}
\details{
The function first creates an eventReactive object that listens for
changes to the input values. This is used to update the reactive input object
every time an input changes. he observer function is then created using the
\code{\link[shiny]{observe}} function. This function triggers every 10 seconds
using \code{\link[shiny]{invalidateLater}}. The observer is invalidated and
re-triggered after 10 seconds have elapsed. This creates a "heartbeat" that
keeps the app running even if the user is not interacting with it. After 2
hours (7200 seconds), the heartbeat will stop and the app can disconnect with
the usual Shiny timeout. This prevents the app from running indefinitely if
the user forgets to close it.
}
