% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex6_to_rgb}
\alias{hex6_to_rgb}
\title{Convert a HEX6 color code to an RGB color string}
\usage{
hex6_to_rgb(hex6)
}
\arguments{
\item{hex6}{<\code{character}> A string representing the HEX6 color code (e.g.
"#FF5733").}
}
\value{
A string representing the RGB color (e.g. "rgb(255, 87, 51)").
}
\description{
This function takes a HEX6 color code and converts it to an RGB string. The
HEX6 code must include the hash symbol (#) at the beginning.
}
