% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex8_to_rgba}
\alias{hex8_to_rgba}
\title{Convert a HEX8 color code to an RGBA color string}
\usage{
hex8_to_rgba(hex8)
}
\arguments{
\item{hex8}{<\code{character}> A string representing the HEX8 color code (e.g.
"#FF5733FF").}
}
\value{
A string representing the RGBA color (e.g. "rgba(255, 87, 51, 1)").
}
\description{
This function takes a HEX8 color code (including alpha channel) and converts
it to an RGBA string. The HEX8 code must include the hash symbol (#) at the
beginning.
}
