% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex_to_rgb_or_rgba}
\alias{hex_to_rgb_or_rgba}
\title{Convert a HEX color code to an RGB or RGBA color string}
\usage{
hex_to_rgb_or_rgba(hex)
}
\arguments{
\item{hex}{<\code{character}> A string representing the HEX color code. It can be
either HEX6 (e.g. "#FF5733") or HEX8 (e.g. "#FF5733FF").}
}
\value{
A string representing the RGB or RGBA color.
}
\description{
This function takes a HEX color code and converts it to either an RGB or
RGBA string based on its length. HEX6 codes are converted to RGB, while
HEX8 codes are converted to RGBA.
}
