% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_functions.R
\name{highlight_dropdown}
\alias{highlight_dropdown}
\title{Highlight options in a dropdown menu}
\usage{
highlight_dropdown(dropdown_identifier, options, color = NULL)
}
\arguments{
\item{dropdown_identifier}{<\code{character}> The dropdown identifier. Must have been
previously added using the options argument of picker widget, e.g.:
\code{options = list(identifier = "dropdown_id")}. In the latter case, the value of
this argument should be 'dropdown_id'.}

\item{options}{<\verb{character vector}> A character vector of option labels to be
highlighted in the dropdown. It's the user-facing option, e.g. 'housing_tenant'
is 'Tenant-occupied (\%)'. The latter must be used.}

\item{color}{<\code{character}> The color (as a hex code) for highlighting. Defaults
to NULL which mean the light blue of colours_dfs$bivar will be used.}
}
\value{
Invisible NULL. The function is called for its side effects.
}
\description{
This function highlights specified options in a dropdown menu of a shinyWidgets
picker input. It requires the dropdown options to highlight and the desired color.
}
