% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\name{home_server}
\alias{home_server}
\title{Create the home server}
\usage{
home_server(id = "home", r)
}
\arguments{
\item{id}{<\code{character}> A unique identifier for the home server. Normally \code{home}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}
}
\value{
A Shiny module server for the home page.
}
\description{
This function generates the server for the home page.
It handles page clicks, active page monitoring, language switch,
and bookmarks.
}
