% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\name{home_UI}
\alias{home_UI}
\title{Create the home user interface}
\usage{
home_UI(id = "home")
}
\arguments{
\item{id}{<\code{character}> A unique identifier for the home UI. Normally \code{home}.}
}
\value{
A Shiny UI object for the home page.
}
\description{
This function generates the user interface for the home page,
retrieving necessary data from the global environment and constructing
the landing page with it.
}
