% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{icon_custom}
\alias{icon_custom}
\title{Create a custom icon element}
\usage{
icon_custom(icon)
}
\arguments{
\item{icon}{<\code{character}> The name or identifier of the custom icon to be
used, location in the www/icons/custom folder.}
}
\value{
An HTML \if{html}{\out{<span>}} tag representing the custom icon element.
}
\description{
This function creates a custom icon element styled in the same way as the
Google Material icons created by the \code{\link{icon_material}} function.
The icon parameter specifies the name or identifier of the custom icon to
be used. To add a custom icon, add it in www/icons/custom
}
