% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{icon_material}
\alias{icon_material}
\title{Create a Material Design icon.}
\usage{
icon_material(icon)
}
\arguments{
\item{icon}{<\code{character}> The name of the icon to be displayed.}
}
\value{
A "span" element with the "material-icons" class and the specified icon.
}
\description{
This function creates a Material Design icon using the "material-icons" class
from the Google Material Design system. The "material-icons" class is a
predefined class that is used to style HTML elements as icons, and is in this case
applied to a "span" element. The Material Design system provides a set of predefined
icons that can be used with the "material-icons" class, and these icons can
be customized using CSS.
}
\details{
To view the available icons that can be used with the "material-icons" class,
please visit \href{https://material.io/resources/icons/}{Material Symbols and Icons - Google Fonts}.
}
