% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{icon_material_button}
\alias{icon_material_button}
\title{Create a Material Design icon button.}
\usage{
icon_material_button(tag, icon)
}
\arguments{
\item{tag}{<\verb{HTML element}>An HTML tag object (a list that represents an HTML
element, e.g. \code{\link[shinyWidgets]{dropdownButton}}) to be modified into
a button with a Material Design icon.}

\item{icon}{<\code{character}> A string indicating the name of the icon to be displayed.}
}
\value{
The modified "tag" object with a button and a Material Design icon.
}
\description{
This function is a wrapper around the \code{\link{icon_material}} function,
which creates the actual icon element. The returned "tag" object can be used
directly in a Shiny UI to create a button with a Material Design icon.
}
\details{
To view the available icons that can be used with the "material-icons" class,
please visit \href{https://material.io/resources/icons/}{Material Symbols and Icons - Google Fonts}.
}
