% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{icon_material_title}
\alias{icon_material_title}
\title{Create a Material Design icon for a title in the right panel.}
\usage{
icon_material_title(icon)
}
\arguments{
\item{icon}{<\code{character}> The name of the icon to be displayed.}
}
\value{
A "span" element with the "material-icons" class and the specified
icon, styled for a title.
}
\description{
This function creates a Material Design icon for a title, using the
\code{\link{icon_material}} function. It applies a predefined set of CSS
styles to the icon, such as font size, color, vertical alignment, and margin.
}
