% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_prod}
\alias{in_prod}
\title{Are we currently in production environment?}
\usage{
in_prod()
}
\value{
A boolean telling if we are or not in a production environment
}
\description{
Are we currently in production environment?
}
