% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric}
\alias{is_numeric}
\title{Whether a value can be transformed to numeric}
\usage{
is_numeric(x)
}
\arguments{
\item{x}{<\code{character}> String to detect if a transformation to numeric
is possible.}
}
\value{
Logical indicating whether the value can be numeric or not.
}
\description{
Purpose mostly for the bookmark. To detect if "2" can be numeric.
}
