% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_scale_in}
\alias{is_scale_in}
\title{Test if x scale is under study}
\usage{
is_scale_in(scales, scale, vectorized = FALSE)
}
\arguments{
\item{scales}{<\verb{character vector}> All scales to test if \code{scale} is in them..}

\item{scale}{<\code{character}> Scale at which the user is on.}

\item{vectorized}{<\code{logical}> Should all elements of \code{scales} be evaluated
and return a logical vector the same length of \code{scales}?}
}
\value{
Returns TRUE or FALSE
}
\description{
Test if x scale is under study
}
