% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_functions.R
\name{js_function_helper}
\alias{js_function_helper}
\title{Helper to use JS functions}
\usage{
js_function_helper(js_fun)
}
\arguments{
\item{js_fun}{<\code{character}> The name of the function to be used in the helper.
Must be one of the \code{functions} argument of the declared functions in
\code{\link{use_curbcut_js}}.}
}
\value{
If in the I, returns the function as a character (e.g. for use in an \code{a}
tag as the \code{onclick} argument). Inside a reactive context, it returns the
\code{shinyjs} framework.
}
\description{
Helper to use javascript functions previously declared by \code{\link{use_curbcut_js}}.
It looks the the use of the script is in the UI or in the server. If in the
UI, then the js function is outputed as a character (e.g. for use in an \code{a}
tag as the \code{onclick} argument). Inside a reactive context, it uses the
\code{shinyjs} framework.
}
