% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovars_helpers.R
\name{label_indicators}
\alias{label_indicators}
\title{Creates a UI label element with an icon and a title}
\usage{
label_indicators(id)
}
\arguments{
\item{id}{A character string specifying the id attribute for the outer div.}
}
\value{
A shiny::div object that includes the specified icon and title.
}
\description{
This function generates a UI element for a Shiny application. The output
is a div with two child divs. One contains a Material Design icon, the
other contains the 'Select an indicator' title.
}
