% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_label_server.R
\name{label_server}
\alias{label_server}
\title{Label and texture server for a Curbcut map}
\usage{
label_server(
  id,
  tile,
  zoom,
  zoom_levels,
  region,
  show_buildings = shiny::reactive(TRUE),
  show_streets = shiny::reactive(TRUE),
  show_stories = shiny::reactive(TRUE),
  mapbox_username = get0("mapbox_username", envir = .GlobalEnv),
  tileset_prefix = get0("tileset_prefix", envir = .GlobalEnv)
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{alp}.}

\item{tile}{<\verb{reactive character}> A reactive string with the map tile to
be used. Either a combination of a region with auto-scale (e.g. \code{city_auto_zoom})
or a combination of a region and a scale (e.g. \code{city_DA}). The output of
\code{\link{zoom_server}}.}

\item{zoom}{<\verb{reactive numeric}> Value representing the current zoom level. Usually
is \code{r[[id]]$zoom()}.}

\item{zoom_levels}{<\verb{reactive named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-scale).}

\item{region}{<\verb{reactive character}> representing the current region displayed
on the map. Usually one of the output of \code{\link{zoom_get_levels}}.}

\item{show_buildings}{<\verb{reactive logical}> Should there be buildings texture?}

\item{show_streets}{<\verb{reactive logical}> Should there be streets texture?}

\item{show_stories}{<\verb{reactive logical}> Should the stories appear on the map?}

\item{mapbox_username}{<\code{character}> Mapbox account username. Defaults to
grabbing the \code{mapbox_username} object from the global environment.}

\item{tileset_prefix}{<\code{character}> Prefix attached to every tileset. Should
correspond to the Curbcut city, e.g. \code{mtl}. Defaults to grabbing the
\code{tileset_prefix} object from the global environment.}
}
\value{
A module server for the label and texture server.
}
\description{
This module server shows labels, for a given region on the map and also loads
and shows textures for streets, buildings, and parks depending on the zoom
level and tile name.
}
