% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{lang_button_label}
\alias{lang_button_label}
\title{Create a language button label}
\usage{
lang_button_label(text)
}
\arguments{
\item{text}{<\code{character}> A character string containing the text to be
displayed next to the icon.}
}
\value{
A character string representing the language button label.
}
\description{
This function creates a label for a language button, which includes an icon
and a text label.
}
