% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_server}
\alias{language_server}
\title{Language Switching Module}
\usage{
language_server(id = "language", r)
}
\arguments{
\item{id}{<\code{character}> A unique id for the module. Default is "language".}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}
}
\value{
A module server function that controls the behavior of the module.
}
\description{
This module allows users to switch the language of Curbcut.
}
