% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_UI}
\alias{language_UI}
\title{Create a UI element for selecting a language}
\usage{
language_UI(id = "language")
}
\arguments{
\item{id}{<\code{character}> The ID of the language selector, defaults to "language".}
}
\value{
An action link that can be clicked to activate the language selector.
}
\description{
This function creates a UI element for selecting a language. It returns an
action link that can be clicked to activate the language selector.
}
