% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.bivar}
\alias{legend_breaks.bivar}
\title{Compute legend breaks for a comparison of two variables}
\usage{
\method{legend_breaks}{bivar}(vars, region, scale, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{delta}. The necessary
named objects in the list are \code{var_left} and \code{var_right}.}

\item{region}{<\code{character}> Indicates the region of interest, e.g. \code{"CMA"}}

\item{scale}{<\code{character}> Indicates the scale of interest, e.g. \code{"CSD"}}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
The \code{delta} class is used when we compare two scores of the same left-hand
variables but in different years, e.g. "housing_tenant_2016" and
"housing_tenant_2021".
}
