% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.bivar_ldelta_rq3}
\alias{legend_breaks.bivar_ldelta_rq3}
\title{Compute legend breaks for a comparison of a single variable of two date times
with another single year variable.}
\usage{
\method{legend_breaks}{bivar_ldelta_rq3}(vars, data, df, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{bivariate_xdelta_yq3}. The
only necessary named object in the list is var_left.}

\item{data}{<\code{data.frame}> Must contains the \code{var_left} and \code{var_left_q3}
columns, which are used to extract the \code{q3} breaks (variation) for the single
variable of two date times.}

\item{df}{<\code{character}> Indicates the combination of the region and scale
of interest, e.g. \code{"CMA_DA"}}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
The \code{bivariate_xdelta_yq3} class is used when there are multiple year selected
for the left-hand variables, but it can only be matched to one year of the
right-hand variable. We show a variation on the left-hand, and a static
year score on the right-hand.
}
