% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.q100}
\alias{legend_breaks.q100}
\title{Compute legend breaks for a variable displaying 100 breaks}
\usage{
\method{legend_breaks}{q100}(vars, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{q100}.}

\item{lang}{<\code{character}> String indicating the language to translate the
\code{Low} and \code{High} breaks to. Defaults to \code{NULL}, which is no translation.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
Compute legend breaks for a variable displaying 100 breaks
}
