% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.q5}
\alias{legend_breaks.q5}
\title{Compute legend breaks for a single \code{var_left}}
\usage{
\method{legend_breaks}{q5}(vars, region, scale, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{q5}. The only necessary
named object in the list is var_left.}

\item{region}{<\code{character}> Indicates the region of interest, e.g. \code{"CMA"}}

\item{scale}{<\code{character}> Indicates the scale of interest, e.g. \code{"CSD"}}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
Compute legend breaks for a single \code{var_left}
}
