% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.qual}
\alias{legend_breaks.qual}
\title{Compute legend breaks for a qualitative variable}
\usage{
\method{legend_breaks}{qual}(vars, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{qual}. The output of
\code{\link[curbcut]{vars_build}}.}

\item{lang}{<\code{character}> String indicating the language to translate the
breaks to. Defaults to \code{NULL}, which is no translation.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
Compute legend breaks for a qualitative variable
}
