% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks}
\alias{legend_breaks}
\title{Generic legend breaks function for Curbcut legends}
\usage{
legend_breaks(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. Named
objects in the list are \code{var_left} and \code{var_right}.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}
}
\value{
It returns an output of \code{\link[ggplot2]{labs}}.
}
\description{
\code{legend_labels} is a generic function used to produce a vector of break
labels for the Curbcut legends. The function invokes
particular methods which depend on the class of the \code{vars} argument.
}
