% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_get_info}
\alias{legend_get_info}
\title{Get information (labels, breaks, themes, ...) for legend creation}
\usage{
legend_get_info(vars, font_family = "SourceSansPro", ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link[curbcut]{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{SourceSansPro}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{...}{Additional arguments to be passed to \code{\link[curbcut]{legend_labels}}
and \code{\link[curbcut]{legend_breaks}}, such as \code{lang}, \code{df}, ...}
}
\value{
A list with the legend labels and breaks computed by
\code{legend_labels()} and \code{legend_breaks()}, and a default theme for
the legend.
}
\description{
Given variables and optional arguments, this function computes the legend
labels and breaks, and returns them with a default theme for the legend.
}
