% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels.bivar_ldelta_rq3}
\alias{legend_labels.bivar_ldelta_rq3}
\title{Get a list of label titles for bivariate xdelta yq3 Curbcut legends}
\usage{
\method{legend_labels}{bivar_ldelta_rq3}(vars, lang, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{bivar_ldelta_rq3}. The
necessary named objects in the list are var_left and var_right.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of label titles returned from \code{\link[ggplot2]{labs}}.
The list contains the x-axis label, the y-axis label, and the short titles for
both axes.
}
\description{
The \code{bivariate_xdelta_yq3} class is used when there are multiple year selected
for the left-hand variables, but it can only be matched to one year of the
right-hand variable. We show a variation on the left-hand, and a static
year score on the right-hand.
}
\details{
\code{legend_labels.bivar_ldelta_rq3}is a method for \code{\link[curbcut]{legend_labels}},
takes a named list of variables and an optional language argument, which it
passes to helper functions to extract the titles of the left and right variables.
It also retrieves the short titles of each variable. The resulting list
contains a label for the x-axis, a label for the y-axis, and the short titles
for each axis.
}
