% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels.delta}
\alias{legend_labels.delta}
\title{Get a list of label titles for delta Curbcut legends}
\usage{
\method{legend_labels}{delta}(vars, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{delta}. The
output of \code{\link{vars_build}}.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of label titles returned from \code{\link[ggplot2]{labs}}.
}
\description{
The \code{delta} class is used when we compare two scores of the same left-hand
variables but in different years, e.g. "housing_tenant_2016" and
"housing_tenant_2021".
}
\details{
\code{legend_labels.delta} takes a named list of variables and an optional language
argument, which it passes to a helper function \code{\link{var_get_title}}
to extract the title of the left variable. The resulting list contains a
label for the x-axis, with a NULL value for the y-axis.
}
