% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels.delta_bivar}
\alias{legend_labels.delta_bivar}
\title{Get a list of label titles for delta bivariate Curbcut legends}
\usage{
\method{legend_labels}{delta_bivar}(vars, lang, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{delta_bivar}. The necessary
named objects in the list are \code{var_left} and \code{var_right}. They both need
to be of length 2.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of label titles returned from \code{\link[ggplot2]{labs}}.
}
\description{
The \code{delta_bivar} class is used when we compare two variations together.
}
\details{
\code{legend_labels.delta_bivar} takes a named list of variables and an optional
language argument, which it passes to a helper function \code{\link[curbcut]{var_get_info}}
to extract the title of the left and right variables, as well as their respective dates.
The resulting list contains a label for the x-axis, with the title of the right variable
and a label for the y-axis with the title of the left variable along with the date of each
variable.
}
