% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels.q5}
\alias{legend_labels.q5}
\title{Get a list of label titles for Curbcut legends containing 5 breaks}
\usage{
\method{legend_labels}{q5}(vars, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{q5}. The only necessary
named objects in the list is \code{var_left}.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of label titles returned from \code{\link[ggplot2]{labs}}.
}
\description{
\code{legend_labels.q5} is a method for \code{\link[curbcut]{legend_labels}},
which takes a named list of variables and an optional language argument, which
it passes to a helper function \code{\link[curbcut]{var_get_info}}
to extract the title of the left variable. The resulting list contains a
label for the x-axis, with a NULL value for the y-axis.
}
