% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels.qual}
\alias{legend_labels.qual}
\title{Get a list of label titles for qualitative Curbcut legends}
\usage{
\method{legend_labels}{qual}(vars, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A named list object of class \code{qual}. The
output of \code{\link{vars_build}}.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list of label titles returned from \code{\link[ggplot2]{labs}}.
}
\description{
\code{legend_labels.qual} is a method for \code{\link{legend_labels}},
which takes a named list of variables with class \code{qual} and an optional
language argument. It passes the \code{var_left} object in \code{vars} to a helper
function \code{\link{var_get_info}} to extract the title of the
variable, and returns a list containing a label for the x-axis and NULL for
the y-axis.
}
