% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels_delta_text}
\alias{legend_labels_delta_text}
\title{Generate a Legend Label with Delta Time}
\usage{
legend_labels_delta_text(title, time_vrvl)
}
\arguments{
\item{title}{<\code{character}> The title for the legend label.}

\item{time_vrvl}{<\verb{numeric vector}> A numeric vector representing time.
This vector can be named or unnamed. If unnamed, the function will
concatenate the time values with a dash. If named, the specific time
values are disregarded, and only a delta symbol is added to the title.}
}
\value{
<\code{character}> A character string representing the formatted legend label.
}
\description{
This function creates a legend label that includes a delta time notation.
It handles both named and unnamed time variables. For unnamed time variables,
the label will include the delta time interval. For named time variables, it
will simply denote a delta without specific time.
}
