% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render.default}
\alias{legend_render.default}
\title{Default legend rendering method}
\usage{
\method{legend_render}{default}(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with an unknown class.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
\code{NULL}
}
\description{
This is the default legend rendering method, which returns a \code{NULL}
value. It is intended to be used when no specific legend rendering method
is available or necessary. In this case, there will be no legend.
}
