% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render.delta}
\alias{legend_render.delta}
\title{Render the legend for a \code{delta} class (variation between two years).}
\usage{
\method{legend_render}{delta}(vars, font_family = "SourceSansPro", ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a \code{delta} class. The only
necessary named object is \code{var_left}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{SourceSansPro}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A ggplot object that represents the \code{delta} legend.
}
\description{
The legend displays a color scale going from red to blue, representing the
changes in the data between two years. The function generates a plot with a
rectangle for each color and a label indicating the percentage change
associated with the color. The first rectangle is a neutral gray color, with
a label indicating that it represents missing data.
}
