% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render.q100}
\alias{legend_render.q100}
\title{Render the legend for a \code{q100} class}
\usage{
\method{legend_render}{q100}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a \code{q100} class. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{acidgrotesk-book}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{scale}{<\code{character}> Scale under study. The output of
\code{\link{update_scale}}.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A ggplot object that represents the \code{q100} legend.
}
\description{
This function generates a legend displaying a variation in viridis colour
scales with labels going from \code{Low} to \code{High}.
}
