% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render.q5}
\alias{legend_render.q5}
\title{Generic legend render function for Curbcut legends}
\usage{
\method{legend_render}{q5}(vars, labs_xy, break_labs, theme_default, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. Named
objects in the list are \code{var_left} and \code{var_right}.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}
}
\value{
It returns a ggplot object
}
\description{
\code{legend_render} is a generic function used to produce the legend for the
Curbcut map pages. The function invokes particular methods which depend on
the class of the \code{vars} argument.
}
