% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render.qual}
\alias{legend_render.qual}
\title{Render the legend for the qualitative class}
\usage{
\method{legend_render}{qual}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a \code{qual} class. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{acidgrotesk-book}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{scale}{<\code{character}> Scale under study. The output of
\code{\link{update_scale}}.}

\item{...}{Additional arguments to be passed to \code{\link{legend_get_info}}.}
}
\value{
A plot generated using ggplot2.
}
\description{
This function generates a plot with a qualitative color legend using the
ggplot2 package. It uses the \code{\link{legend_get_info}} function
to obtain the necessary information about the legend and cuts the qualitative
color table (\code{colours$qual}) to the number of breaks. The function then
generates the plot using the obtained information.
}
