% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render}
\alias{legend_render}
\title{Generic legend render function for Curbcut legends}
\usage{
legend_render(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{acidgrotesk-book}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{scale}{<\code{character}> Scale under study. The output of
\code{\link{update_scale}}.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}
}
\value{
It returns a ggplot object
}
\description{
\code{legend_render} is a generic function used to produce the legend for the
Curbcut map pages. The function invokes particular methods which depend on
the class of the \code{vars} argument.
}
