% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_legend.R
\name{legend_server}
\alias{legend_server}
\title{Create the server logic for the legend module}
\usage{
legend_server(
  id,
  r,
  data,
  var_left,
  var_right,
  region,
  df,
  hide = shiny::reactive(FALSE),
  build_str_as_DA = shiny::reactive(TRUE),
  breaks = shiny::reactive(NULL)
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file.}

\item{data}{<\verb{reactive data.frame}> Data frame containing all the scale and
the \code{var_left} and \code{var_right}. The output of \code{\link[curbcut]{get_data}}.}

\item{var_left}{<\verb{reactive character}> Character string of the selected
variable, e.g. \code{canale_2016}.}

\item{var_right}{<\verb{reactive character}> Character string of the selected
compared variable, e.g. \code{housing_tenant_2016}.}

\item{region}{<\verb{reactive character}> The region under study, e.g. \code{CMA}.
Normally \code{r$region}, reactive created in the \code{server.R} file.}

\item{df}{<\verb{reactive character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}.}

\item{hide}{<\verb{reactive logical}> Should the legend be hidden? Defaults to
\code{shiny::reactive(FALSE)}}

\item{build_str_as_DA}{<\verb{reactive logical}> Should the buildings and streets
be treated as DAs? Defaults to \code{shiny::reactive(TRUE)}.}

\item{breaks}{<\verb{reactive numeric vector}> Breaks if they need to be manually
supplied to the legend module.}
}
\value{
The legend Shiny server function
}
\description{
Create the server logic for the legend module
}
