% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_legend.R
\name{legend_UI}
\alias{legend_UI}
\title{Create the UI for the legend module}
\usage{
legend_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{canale}.}
}
\value{
The legend Shiny UI object
}
\description{
Create the UI for the legend module
}
