% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_get_zoom}
\alias{link_get_zoom}
\title{Retrieve the Appropriate Zoom Level}
\usage{
link_get_zoom(zoom_levels, df)
}
\arguments{
\item{zoom_levels}{<\verb{named vector}> A named vector of potential zoom levels.}

\item{df}{<\code{character}> df used for scale matching.}
}
\value{
Returns the appropriate zoom level.
}
\description{
This function obtains the zoom level corresponding to the scale of a data
frame. If the zoom level is 0, it defaults to 9.
}
