% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_helpers.R
\name{longest_scale_combination}
\alias{longest_scale_combination}
\title{Find the longest scale combination}
\usage{
longest_scale_combination(top_scale, avail_scale_combinations)
}
\arguments{
\item{top_scale}{<\code{character}> The top scale to search for in the combinations.}

\item{avail_scale_combinations}{<\verb{character vector}> A vector of available
scale combinations.}
}
\value{
<\code{character}> The longest scale combination string for the specified
top scale. If there's only one matching scale, it is returned. If
there are multiple, the one with the most scales is returned.
}
\description{
This function identifies the longest scale combination string for a given
top scale within available scale combinations. It selects the combination
that contains the highest number of scales in its string.
}
