% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_label_helpers.R
\name{map_label_extrude}
\alias{map_label_extrude}
\title{Extrude polygons on map}
\usage{
map_label_extrude(map_view_state, zoom, zoom_levels, tile, extrude)
}
\arguments{
\item{map_view_state}{<\code{list}> A list containing the current state of the map view.
Usually rdeck::get_view_state("map")}

\item{zoom}{<\code{numeric}> Value representing the current zoom level. Usually
is \code{r[[id]]$zoom()}.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-zoom).}

\item{tile}{<\code{character}> String representing the tile name.}

\item{extrude}{<\code{logical}> Indicates whether to extrude or not.
If FALSE, the function will immediately return FALSE.}
}
\value{
Returns TRUE if the polygons should be extruded on the map. Otherwise,
returns FALSE.
}
\description{
This function determines if polygons should be extruded on a map. It
considers parameters like zoom level, tile, and a boolean 'extrude'.
}
