% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_label_helpers.R
\name{map_label_show_texture}
\alias{map_label_show_texture}
\title{Determine whether to show texture for a given zoom level and tile name}
\usage{
map_label_show_texture(zoom, tile, zoom_levels, map_module = FALSE)
}
\arguments{
\item{zoom}{<\code{numeric}> Value representing the current zoom level. Usually
is \code{r[[id]]$zoom()}.}

\item{tile}{<\code{character}> String representing the tile name.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-zoom).}

\item{map_module}{<\code{logical}> Is the function placed inside the map module
to informe extrusion? If so, zoom should have no effect on extrusion.}
}
\value{
A logical value indicating whether or not to show texture for the given
zoom level and tile name
}
\description{
This function determines whether to show texture for a given zoom level and
tile based on a set of conditions.
}
