% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scales.R
\name{map_scale_colour}
\alias{map_scale_colour}
\title{Map scale colour for borders}
\usage{
map_scale_colour(select_id, data_colours, tileset_ID_color = "ID_color")
}
\arguments{
\item{select_id}{<\code{character}> The ID of the feature to highlight. The specified
colour will be applied to this feature.}

\item{data_colours}{<\code{data.frame}> The output of \code{\link{data_get_colours}}.
Two columns: \code{ID} and \code{fill}. Every \code{ID} will be coloured using its according
\code{fill} colour.}

\item{tileset_ID_color}{<\code{character}> The name of the column on the tileset
used to colour every polygon. Defaults to \code{ID_color}, where the buildings
scales have in their \code{ID_color} column the \code{ID} of their DA.}
}
\value{
A color scale for polygons borders used in an rdeck map.
}
\description{
This function assigns a specific colour to the selected ID and a different
colour to all other IDs.
}
