% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scales.R
\name{map_scale_fill}
\alias{map_scale_fill}
\title{Map scale fill}
\usage{
map_scale_fill(data_colours, tileset_ID_color = "ID_color")
}
\arguments{
\item{data_colours}{<\code{data.frame}> The output of \code{\link{data_get_colours}}.
Two columns: \code{ID} and \code{fill}. Every \code{ID} will be coloured using its according
\code{fill} colour.}

\item{tileset_ID_color}{<\code{character}> The name of the column on the tileset
used to colour every polygon. Defaults to \code{ID_color}, where the buildings
scales have in their \code{ID_color} column the \code{ID} of their DA.}
}
\value{
A color scale for polygons used in an rdeck map.
}
\description{
This function creates a color scale for the fill colours of the maps. It
takes the output of \code{\link{data_get_colours}} and feeds it
to the \code{\link[rdeck]{scale_color_category}} function.
}
