% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scales.R
\name{map_scale_lwd}
\alias{map_scale_lwd}
\title{Map scale line width}
\usage{
map_scale_lwd(select_id, tile = NULL, zoom = NULL, zoom_levels = NULL)
}
\arguments{
\item{select_id}{<\code{character}> A string specifying the ID of the selected polygon.
The selected polygon will have an increased in the width of its border.}

\item{tile}{<\code{character}> The name of the tileset currently present on the map.}

\item{zoom}{<\code{numeric}> The current zoom level of the map.}

\item{zoom_levels}{<\verb{named vector}> Zoom levels under study. One of the
map_zoom_levels_x in the global environment. It contains the zoom at which
a scale should switch on an autozoom, e.g. \code{c(CMA = 0, CT = 10.5, DA = 12.5, ...)}.}
}
\value{
A line width scale for polygons in an rdeck map.
}
\description{
This function creates a line width scale for polygons in an rdeck map based
on the zoom level of the map. If \code{tile}, \code{zoom} and \code{zoom_levels} are given,
the borders will disappear at one zoom digit under the zoom at which they
would have appeared on an auto-zoom.
}
