% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_schema_to_z_col}
\alias{match_schema_to_z_col}
\title{Match schema list to a specific column (\code{col}) in \code{data} based on \code{vl_vr} parameter}
\usage{
match_schema_to_z_col(
  data,
  time,
  col,
  vl_vr,
  data_schema = attr(data, sprintf("schema_\%s", vl_vr)),
  schemas
)
}
\arguments{
\item{data}{<\code{data.frame}> A data frame containing the data.}

\item{time}{<\verb{numeric named list}> A named list specifying the time at which
data is displayed. The names should correspond to \code{var_left} and \code{var_right}}

\item{col}{<\code{character}> The specific column to be extracted, e.g. \code{group}.}

\item{vl_vr}{<\code{character}> Determines which schema to use: 'var_left' or 'var_right'.}

\item{data_schema}{<\verb{named list}> The schema information for the data, typically
an attribute of the data frame. Defaults to taking the attributes of data (of vl_vr).}

\item{schemas}{<\verb{named list}> Current schema information, which can impact the
choice of data column. If \code{NULL}, no additional schemas are considered.}
}
\value{
The name of the variable corresponding to the given schema and \code{vl_vr}
selection as a character string.
}
\description{
This function extends the functionality of \code{match_schema_to_col} by adding support
for dynamic selection of schema based on the \code{vl_vr} parameter. It identifies the
column in a \code{data} frame that matches a given schema, considering whether to use
the \code{var_left} or \code{var_right} schema. This is useful in scenarios where the column
itself to select is not \code{var_left} nor \code{var_right}, but for example \code{group}.
}
