% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{min_sig_digits}
\alias{min_sig_digits}
\title{Get the minimum number of significant digits before the decimal point in a
vector of numbers}
\usage{
min_sig_digits(x)
}
\arguments{
\item{x}{<\code{numeric}> A numeric vector}
}
\value{
An integer representing the minimum number of significant digits
before the decimal point in x
}
\description{
This function takes a vector of numbers and returns the minimum number of
significant digits before the decimal point in that vector. For example, if
the vector contains the numbers 123.456, 0.000789, 0.01, 1000, and 0.5, the
minimum number of significant digits before the decimal point is 1
(corresponding to the number of digits in the number 0.5).
}
\examples{
x <- c(123.456, 0.000789, 0.01, 1000, 0.5)
min_sig_digits(x) # 3
}
