% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{modules_panel}
\alias{modules_panel}
\title{Generate User Interface for Ready Modules}
\usage{
modules_panel(modules = get_from_globalenv("modules"))
}
\arguments{
\item{modules}{<\code{data.frame}> Tibble of all the pages.}
}
\value{
A list of translated shiny UI elements.
}
\description{
This function prepares a list of user interfaces (UIs) for all pages in the
modules data.frame. The module UIs are grouped by their theme and presented
in alphabetical order. Each module UI is created using its corresponding
key and title.
}
