% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns_doubled.R
\name{ns_doubled}
\alias{ns_doubled}
\title{Create a double nested namespace for an element}
\usage{
ns_doubled(page_id, element)
}
\arguments{
\item{page_id}{<\code{character}> A character string representing the page namespace.
This is typically the page ID.}

\item{element}{<\code{character}> A character string representing the Shiny element's ID.}
}
\value{
A character string representing the double nested namespace for the
Shiny element, in the format 'page_id-page_id-element'.
}
\description{
This function generates a double nested namespace string which is useful for
managing any module that's created in a 'map page'. Ex. that way we could access
potentially the dropdown widget of housing directly from the \code{server.R}, which
is \code{housing-housing-ccpicker_var} (page_id-page_id-element)
}
