% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ntile}
\alias{ntile}
\title{Calculate the ntile for a given vector of values}
\usage{
ntile(x, n)
}
\arguments{
\item{x}{<\verb{numeric vector}> Vector of numeric values}

\item{n}{<\code{integer}> An integer specifying the number of quantiles to
calculate}
}
\value{
An integer vector indicating the bin number for each element in x
}
\description{
This function divides a vector into n bins of equal size and returns the
bin number for each element in the vector. It is a reimplementation of
\code{\link[dplyr]{ntile}} in base R.
}
\seealso{
The dplyr package provides a ntile function with similar
functionality.
}
