% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordinal_form}
\alias{ordinal_form}
\title{Generate English or French ordinal form of a number}
\usage{
ordinal_form(lang, x, en_first = "first")
}
\arguments{
\item{lang}{<\code{character}> String indicating the language for the ordinal form,
either "en" or "fr".}

\item{x}{<\code{numeric}> The number to be transformed into ordinal form.}

\item{en_first}{<\code{character}> The English word for first
(default is \code{"first"}). Useful for place explorer where \verb{ranks 'best'} is more
suitable.}
}
\value{
A character string, the ordinal form of x in the selected language.
}
\description{
This function returns the English or French ordinal form of a number. The
ordinal form can be selected by the lang argument, which takes either
"en" for English or "fr" for French.
}
\examples{
ordinal_form("en", 1) # "first"
ordinal_form("fr", 2) # "deuxième"
ordinal_form("en", 23) # "23rd"
}
