% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{page_get}
\alias{page_get}
\title{Retrieve a specific page}
\usage{
page_get(ns_id)
}
\arguments{
\item{ns_id}{<\code{character}> ID of the page identifier including the namespace.}
}
\value{
A subset of the \code{modules} global data frame matching the \code{ns_id}.
}
\description{
This function retrieves a specific page based on the provided \code{ns_id}.
It substracts the namespace doubling noise and retrieves the page from
the modules stored in the global environment.
}
