% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_helpers.R
\name{panel_view_prepare_text}
\alias{panel_view_prepare_text}
\alias{panel_view_prepare_text.q5}
\alias{panel_view_prepare_text.delta}
\alias{panel_view_prepare_text.bivar}
\alias{panel_view_prepare_text.delta_bivar}
\alias{panel_view_prepare_text.default}
\title{Prepare the panel view text for the given variables.}
\usage{
panel_view_prepare_text(vars, df, dat, lang = NULL, ...)

\method{panel_view_prepare_text}{q5}(vars, df, dat, lang = NULL, ...)

\method{panel_view_prepare_text}{delta}(vars, df, dat, lang = NULL, ...)

\method{panel_view_prepare_text}{bivar}(vars, df, dat, lang = NULL, ...)

\method{panel_view_prepare_text}{delta_bivar}(vars, df, dat, lang = NULL, ...)

\method{panel_view_prepare_text}{default}(vars, df, dat, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function.}

\item{df}{<\verb{reactive character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}.}

\item{dat}{<\code{data.frame}> The data frame containing the columns under analysis}

\item{lang}{<\code{character}> The language to use for translating the texts.
Defaults to NULL for no translation}

\item{...}{Additional arguments to be passed to the underlying methods.}
}
\value{
A string containing the panel view text.
}
\description{
This function returns a panel view text including title and description for
given variables, considering the minimum, maximum, mean, and standard deviation
of the data points in the input dataset.
}
