% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_helpers.R
\name{panel_view_prepare_text_helper}
\alias{panel_view_prepare_text_helper}
\title{Prepare Text Helper for Panel View}
\usage{
panel_view_prepare_text_helper(
  df,
  var,
  dat,
  title,
  explanation,
  title_color = "#73AE80",
  lang = NULL
)
}
\arguments{
\item{df}{<\verb{reactive character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}.}

\item{var}{<\code{character}> A character representing the variable name in the
dataset (e.g. 'housing_tenant') with the right class (pct, dollar, ind, ...).}

\item{dat}{<\code{data.frame}> The data frame containing the columns from which
to retrieve the values.}

\item{title}{<\code{character}> The pretty title of the variable, which must
match the column of the pretty table.}

\item{explanation}{<\code{character}> A character representing the explanation of
the variable. A piece of the \code{variables} table.}

\item{title_color}{<\code{character}> The color code for the title. Default is
#73AE80' for the left variable's green.}

\item{lang}{<\code{character}> The language to use for translating variable names.
Defaults to NULL for no translation}
}
\value{
A character string containing the formatted HTML text with statistics.
}
\description{
This function creates a formatted HTML text containing statistics for a given
variable. It includes the minimum, maximum, mean, and standard deviation. The
title is displayed with a custom color, usually depends on wether it is a left
or right variable.
}
