% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_helpers.R
\name{panel_view_rename_cols}
\alias{panel_view_rename_cols}
\alias{panel_view_rename_cols.q5}
\alias{panel_view_rename_cols.delta}
\alias{panel_view_rename_cols.bivar}
\alias{panel_view_rename_cols.delta_bivar}
\alias{panel_view_rename_cols.bivar_ldelta_rq3}
\title{Rename pretty columns in data frame for panel view}
\usage{
panel_view_rename_cols(vars, dat, lang = NULL, ...)

\method{panel_view_rename_cols}{q5}(vars, dat, lang = NULL, ...)

\method{panel_view_rename_cols}{delta}(vars, dat, lang = NULL, ...)

\method{panel_view_rename_cols}{bivar}(vars, dat, lang = NULL, ...)

\method{panel_view_rename_cols}{delta_bivar}(vars, dat, lang = NULL, ...)

\method{panel_view_rename_cols}{bivar_ldelta_rq3}(vars, dat, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function.}

\item{dat}{<\code{data.frame}> The data frame containing the columns to be
renamed.}

\item{lang}{<\code{character}> The language to use for translating variable names.
Defaults to NULL for no translation}

\item{...}{Additional arguments to be passed to the underlying methods.}
}
\value{
A list with data frame with renamed columns and a list of column
names that need styling.
}
\description{
This function renames columns in a data frame for use in a panel view. The
column names are modified to include similar labels than the ones from the
legend and ensures that information about the time period is being displayed.
}
