% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_module.R
\name{panel_view_server}
\alias{panel_view_server}
\alias{panel_view_UI}
\title{Renders a floating panel with a map view and data view buttons.}
\usage{
panel_view_server(id, r, vars, data)

panel_view_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the panel will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function. The class of the vars object is
used to determine which type of legend to draw.}

\item{data}{<\verb{reactive data.frame}> Data frame containing all the scale and
the \code{var_left} and \code{var_right}. The output of \code{\link{data_get}}.}
}
\value{
Panel view module
}
\description{
This function renders a Shiny server module for the \code{panel_view_UI} UI
element which displays a floating panel with buttons to toggle between the map
an table/data info. It shows a map when the map button is clicked and a table
with data when the data button is clicked. When there is a selection, it also
shows a magnifying glass linking to the place explorer.
}
\section{Functions}{
\itemize{
\item \code{panel_view_UI()}: Create the UI for the legend module

}}
