% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_helpers.R
\name{panel_view_style_cols}
\alias{panel_view_style_cols}
\alias{panel_view_style_cols.pct}
\alias{panel_view_style_cols.dollar}
\alias{panel_view_style_cols.count}
\alias{panel_view_style_cols.default}
\title{Apply styling to columns of a DT::datatable based on data type}
\usage{
panel_view_style_cols(var, table, ...)

\method{panel_view_style_cols}{pct}(var, table, ...)

\method{panel_view_style_cols}{dollar}(var, table, ...)

\method{panel_view_style_cols}{count}(var, table, ...)

\method{panel_view_style_cols}{default}(var, table, ...)
}
\arguments{
\item{var}{<\code{character}> One or more column names of \code{table}. It can be one
of the following classes: "pct" for percentages, "dollar" for currency, or
anything else for the default styling (none).}

\item{table}{<\code{datatable}> A \code{\link[DT]{datatable}} object.}

\item{...}{Additional arguments to be passed to the formatting functions.}
}
\value{
A styled table object with formatting applied to the columns based on
their data type.
}
\description{
This function applies styling to columns of a table based on their data type.
}
