% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_picker_helpers.R
\name{picker_hover_divs}
\alias{picker_hover_divs}
\title{Creates divs with explanations on hover for a given list of variables.}
\usage{
picker_hover_divs(var_list, lang = NULL)
}
\arguments{
\item{var_list}{<\verb{named list}> A named list of variable codes. The output
of \code{\link{dropdown_make}} already translated.}

\item{lang}{<\code{character}> The language to translate the explanations into (optional).}
}
\value{
A list containing a single element named 'content'. The 'content' element is a
character vector containing the HTML for the divs.
}
\description{
This function creates a list of divs, where each div contains a variable's
label and an explanation that appears on hover. The variables are specified
using their codes, which are used to look up the corresponding explanations in
the variables table retrieved from the global environment. If a
language is specified, the explanations are translated into that language
using the \code{\link{cc_t}} function. It is intended to be used by the
\code{widget_picker}.
}
