% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_picker_helpers.R
\name{picker_multi_year_disable}
\alias{picker_multi_year_disable}
\title{Disables variables that are not available for the maximum number of years}
\usage{
picker_multi_year_disable(id, var_list, disable)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{alp}.}

\item{var_list}{<\verb{named list}> A named list of variable codes. The output
of \code{\link{dropdown_make}}.}

\item{disable}{<\code{logical}> A flag indicating whether to disable variables.}
}
\value{
A logical vector of the same length as the input variable list. The
elements of the output vector indicate whether each variable should be
disabled or not.
}
\description{
Given a list of variables and a boolean flag indicating whether to disable
variables, this function disables variables that are not available for the
maximum number of years. If the disable flag is not set, then the function
returns a logical vector of FALSE values.
}
