% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_picker.R
\name{picker_UI}
\alias{picker_UI}
\title{Create a picker input control in an optionally styled div}
\usage{
picker_UI(
  id,
  picker_id = "var",
  var_list,
  label = NULL,
  width = "100\%",
  div_style = NULL,
  selected = NULL,
  open_left = TRUE,
  ...
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{canale}.}

\item{picker_id}{<\code{character}> string giving the identifier for the picker input
object. This will be used as the input's \code{inputId} and will inform the bookmarking.
It needs to be unique in the page. The default value is \code{"var"}.}

\item{var_list}{<\verb{reactive list}> Choices to display in the picker input.
Normally made using \code{\link{dropdown_make}}.}

\item{label}{A character string giving the label for the picker input control.
If \code{NULL}, no label will be displayed.}

\item{width}{<\code{character}> A character string giving the width of the picker
input control. The width of the input : 'auto', 'fit', '100px', '75\%'.}

\item{div_style}{<\code{character}> CSS styles to apply to the \code{div} container.
The default value is \code{NULL} for no additional styling.}

\item{selected}{A character vector of a value to preselect. The default value
is \code{NULL} (the first value will be selected).}

\item{open_left}{<\code{logical}> Should the dropdown open to the left? Any dropdown
placed on the left-side of the screen should do see, and it is the default.
Compare dropdown should open right, and this can be changed.}

\item{...}{Additional arguments to pass to \code{\link[shinyWidgets]{pickerInput}}}
}
\value{
A \code{div} container containing a \code{\link[shinyWidgets]{pickerInput}}
control.
}
\description{
This function creates a \code{\link[shinyWidgets]{pickerInput}} control in a
\code{div} container. The \code{pickerInput} control allows users to select one item
from a list of variables choices.
}
\seealso{
\code{\link{picker_server}}
}
