% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer_helpers.R
\name{place_explorer_create_html}
\alias{place_explorer_create_html}
\title{Create HTML file for Place Explorer with dynamic content}
\usage{
place_explorer_create_html(tmpfile, region, df, select_id, lang)
}
\arguments{
\item{tmpfile}{<\code{character}> Temporary file location, an .html document.}

\item{region}{<\code{character}> Region under study, e.g. \code{CMA}.}

\item{df}{<\code{character}> Scale under study, e.g. \code{DA}.}

\item{select_id}{<\code{character}> Selected identifier for the selected combinasion
of \code{region} and \code{df}.}

\item{lang}{<\code{character}> Language that should be used to produce the main card
output. There need to be a function \code{placeex_main_card_prep_output_x} available
in that language. \code{en} or \code{fr}.}
}
\value{
The function creates an HTML document which represents the Place Explorer
for the given data point (specified by \code{select_id}). It is output at the location
of \code{tmpfile}
}
\description{
This function creates a temporary HTML file using data provided as parameters and
global variables. It first checks the version of the \code{bslib} package, which needs to
be the modified version: \code{devtools::install_github('bdbmax/bslib')} to allow
for tabs colouring. If not, throws a warning. It then retrieves
necessary global variables. The function also dynamically adjusts the map zoom level
based on the scale of the data.
}
