% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer_helpers.R
\name{place_explorer_html_links}
\alias{place_explorer_html_links}
\title{Create and concatenate Place Explorer HTML files}
\usage{
place_explorer_html_links(temp_folder, region, scale, select_id, lang = NULL)
}
\arguments{
\item{temp_folder}{<\code{character}> The temporary folder of the app. By default
will grab the \code{temp_folder} object as it's already supposed to have been assigned
in the \code{global.R} file}

\item{region}{<\code{character}> Region under study, e.g. \code{CMA}.}

\item{scale}{<\code{character}> Scale under study}

\item{select_id}{<\code{character}> The ID of the selected polygon.}

\item{lang}{<\code{character}> String indicating the language of the place explorer
file (either "en" for English or "fr" for French). Default is NULL, which
will use English if not specified.}
}
\value{
A list containing the following elements:
\itemize{
\item src: The relative path to the temporary file, to be used when grabbing
the static file for the web server.
\item file: The absolute path to the temporary file.
}
}
\description{
This function creates a place explorer HTML file by concatenating a header
file with a specific place explorer file, based on the provided df, select_ID
and language. It also generates a temporary file to store the concatenated
output.
}
