% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer_module.R
\name{place_explorer_server}
\alias{place_explorer_server}
\alias{place_explorer_UI}
\title{Launches the place explorer module server}
\usage{
place_explorer_server(
  id,
  r,
  scales_as_DA = shiny::reactive(c("building", "street")),
  map_zoom = get_from_globalenv("map_zoom"),
  map_loc = get_from_globalenv("map_loc"),
  mapbox_username = get_from_globalenv("mapbox_username"),
  tileset_prefix = get_from_globalenv("tileset_prefix"),
  map_base_style = get_from_globalenv("map_base_style"),
  temp_folder = get_from_globalenv("temp_folder")
)

place_explorer_UI(id, scales_as_DA = c("building", "street"))
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{scales_as_DA}{<\verb{reactive character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. These
scales won't be selectable by the user.}

\item{map_zoom}{<\code{numeric}> The default zoom level for the map. By default,
this value is retrieved from the \code{map_zoom} variable in the global environment.}

\item{map_loc}{<\verb{numeric vector}> The default location for the map. By default,
this value is retrieved from the \code{map_loc} variable in the global environment.}

\item{mapbox_username}{<\code{character}> Mapbox account username. Defaults to
grabbing the \code{mapbox_username} object from the global environment.}

\item{tileset_prefix}{<\code{character}> Prefix attached to every tileset. Should
correspond to the Curbcut city, e.g. \code{mtl}. Defaults to grabbing the
\code{tileset_prefix} object from the global environment.}

\item{map_base_style}{<\code{character}> The mapbox basemap style url.
See https://docs.mapbox.com/api/maps/#mapbox-styles}

\item{temp_folder}{<\code{character}> The temporary folder of the app. By default
will grab the \code{temp_folder} object as it's already supposed to have been assigned
in the \code{global.R} file}
}
\description{
This function initializes the server logic for the place explorer module,
which is a shiny module that allows users to explore geographical areas and
retrieve data about them. The module consists of a map, a sidebar, and a main
panel that displays detailed information about a selected area.
}
\section{Functions}{
\itemize{
\item \code{place_explorer_UI()}: Create the UI for the place explorer module

}}
