% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer_helpers.R
\name{place_explorer_vars}
\alias{place_explorer_vars}
\title{Retrieve default region and map zoom levels for the Place Explorer module}
\usage{
place_explorer_vars(scales_as_DA = c("building", "street"))
}
\arguments{
\item{scales_as_DA}{<\verb{character vector}> Map scales to exclude from the
returned map zoom levels. The default value is c("building", "street").}
}
\value{
A list with two elements: \code{default_regions} and
\code{map_zoom_levels}. \code{default_regions} is a character vector with the
default region for the Place Explorer module. \code{map_zoom_levels} is a named
character vector with the map zoom levels for the default region, excluding
any specified map scales.
}
\description{
This function retrieves the default region and map zoom levels for the Place
Explorer module. It does so by grabbing the modules from the global environment
and then getting the default region and map zoom levels based on the ID of the
Place Explorer module. The default region is used to retrieve the map zoom
levels from the global environment. We can also specify one or more map
scales to exclude from the returned map zoom levels, which by default are
buildings and streets.
}
