% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_explorer.R
\name{placeex_main_card}
\alias{placeex_main_card}
\title{Final function for the place explorer main card}
\usage{
placeex_main_card(pe_main_card, region, df, select_id, lang, translation)
}
\arguments{
\item{pe_main_card}{<\code{list}> Place explorer data for the maincard. Usually
the output of \code{\link[cc.buildr]{placeex_main_card}} saved in the
data folder.}

\item{region}{<\code{character}> Region under study, e.g. \code{CMA}.}

\item{df}{<\code{character}> Scale under study, e.g. \code{DA}.}

\item{select_id}{<\code{character}> Selected identifier for the selected combinasion
of \code{region} and \code{df}.}

\item{lang}{<\code{character}> Language to use for translation. Must be one of
en' or 'fr'.}

\item{translation}{<\code{data.frame}> Data frame containing translation data.}
}
\value{
Returns a list of all the main card variables and how the selected
id compares in its dataset.
}
\description{
Final function for the place explorer main card
}
