% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_of_interests.R
\name{poi_get}
\alias{poi_get}
\title{Get nearby points of interests (POIs) based on map view state}
\usage{
poi_get(id, map_id = paste0(id, "-map"), poi)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{canale}.}

\item{map_id}{<\code{character}> The ID of the map. By default, it is
set to the value of \code{paste0(id, "-map")} as the \code{\link[curbcut]{map_server}}
module assigns the \code{map} id to the maps.}

\item{poi}{<\verb{character vector}> The current POIs showing on the map.}
}
\value{
A character vector of nearby POIs if new POIs are found; otherwise,
returns the input POI vector.
}
\description{
This function retrieves the nearby POIs based on the current view state of a
map specified by the \code{map_id} parameter. The function filters the POIs
based on their proximity to the center of the map, keeping only those POIs
that are within 2000 meters. Currently only returns the points of interests
of the \code{stories} data.frame.
}
