% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popup_module.R
\name{popup_server}
\alias{popup_server}
\title{Server-side function for a popup in Curbcut}
\usage{
popup_server(id, content, show_popup)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}.}

\item{content}{<\verb{character reactive}> A reactive expression that supplies the content
to be displayed in the popup.}

\item{show_popup}{<\verb{logical reactive}> A reactive expression that determines whether
the popup should be shown.}
}
\description{
This function defines the server-side logic for a popup element in the Curbcut
app It creates a reactive UI element for displaying a popup when
specific conditions are met.
}
