% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_region_value.R
\name{region_value}
\alias{region_value}
\title{Compute region values based on different methods}
\usage{
region_value(var, data, time, scale, region, select_id, ...)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{data}{<\code{data.frame}>}

\item{time}{<\code{numeric}>}

\item{scale}{<\code{character}>}

\item{region}{<\code{character}>}

\item{...}{Additional arguments to pass to specific method functions.}
}
\value{
Varies based on the method specified in \code{var}. Commonly
returns a list with elements like \code{val} and \code{count}.
}
\description{
This function computes region values based on the method specified by \code{var}'s
class. It uses weighted means, sums, or other calculations.
It fetches the relevant data and parent data, and then calls the
appropriate method for calculation.
}
