% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_region_value.R
\name{region_value_data_grab}
\alias{region_value_data_grab}
\title{Retrieve data values and parent data for specific region and scale}
\usage{
region_value_data_grab(var, data, time, scale, region, col)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{time}{<\code{numeric}> The time period of interest.}

\item{scale}{<\code{character}> The scale of interest.}

\item{region}{<\code{character}> The region of interest.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}

\item{data_path}{<\code{character}> The path to the data directory, defaults to
"data/".}
}
\value{
Returns a list containing two elements: \code{data_vals}, which
holds the data values for the specified variable, and \code{parent_vals},
which holds the parent data values.
}
\description{
This function gathers data values and parent data based on the given
variable, data, time, scale, and region. The function utilizes
helper functions to identify the correct column and fetch parent data.
}
