% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_region_value.R
\name{region_value_method_postgresql}
\alias{region_value_method_postgresql}
\alias{region_value_method_postgresql.pct}
\alias{region_value_method_postgresql.count}
\alias{region_value_method_postgresql.ind}
\alias{region_value_method_postgresql.default}
\title{Methods to compute regional values (postgresql)}
\usage{
region_value_method_postgresql(
  var,
  data,
  scale,
  parent_string,
  parent_string_schema,
  var_schema,
  ...
)

\method{region_value_method_postgresql}{pct}(
  var,
  data,
  scale,
  parent_string,
  parent_string_schema,
  var_schema,
  ...
)

\method{region_value_method_postgresql}{count}(
  var,
  data,
  scale,
  parent_string,
  parent_string_schema,
  var_schema,
  ...
)

\method{region_value_method_postgresql}{ind}(
  var,
  data,
  scale,
  parent_string,
  parent_string_schema,
  var_schema,
  col,
  ...
)

\method{region_value_method_postgresql}{default}(
  var,
  data,
  scale,
  parent_string,
  parent_string_schema,
  var_schema,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{scale}{<\code{character}> The scale of interest.}

\item{parent_string}{<\code{character}> The parent string as character ex. c_population}

\item{parent_string_schema}{<\code{character}> With the schema on ex. c_population_2021}

\item{var_schema}{<\code{character}> With the schema on ex. alp_2021}

\item{...}{Additional arguments (e.g., data, time) used in specific methods.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right}.}
}
\value{
Returns a list with calculated values for the specific method.
Commonly returns elements such as \code{val} and \code{count}.
}
\description{
A set of methods to compute regional values based on the type of
variable provided. Supports weighted mean, sums, and specific
indicators.
}
\section{Methods (by class)}{
\itemize{
\item \code{region_value_method_postgresql(pct)}: The method for percentage variables.

\item \code{region_value_method_postgresql(count)}: The method for count variables.

\item \code{region_value_method_postgresql(ind)}: The method for ind variables.

\item \code{region_value_method_postgresql(default)}: The default method

}}
