% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_region_value.R
\name{region_value_method_qs}
\alias{region_value_method_qs}
\alias{region_value_method_qs.pct}
\alias{region_value_method_qs.count}
\alias{region_value_method_qs.ind}
\alias{region_value_method_qs.ppo}
\alias{region_value_method_qs.default}
\title{Methods to compute regional values (ondisk, .qs files)}
\usage{
region_value_method_qs(var, data_vals, parent_vals, ...)

\method{region_value_method_qs}{pct}(var, data_vals, parent_vals, ...)

\method{region_value_method_qs}{count}(var, data_vals, parent_vals, ...)

\method{region_value_method_qs}{ind}(
  var,
  data_vals,
  parent_vals,
  data,
  time,
  col,
  schemas = schemas,
  ...
)

\method{region_value_method_qs}{ppo}(var, data_vals, parent_vals, ...)

\method{region_value_method_qs}{default}(var, data_vals, parent_vals, ...)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{data_vals}{<\verb{numeric vector}> Numeric vector containing the data values
for the region.}

\item{parent_vals}{<\verb{numeric vector}> Numeric vector containing the parent
data values.}

\item{...}{Additional arguments (e.g., data, time) used in specific methods.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{time}{<\code{numeric}> The time period of interest.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right}.}

\item{schemas}{<\verb{named list}> Current schema information. The additional widget
values that have an impact on which data column to pick. Usually \code{r[[id]]$schema()}.}
}
\value{
Returns a list with calculated values for the specific method.
Commonly returns elements such as \code{val} and \code{count}.
}
\description{
A set of methods to compute regional values based on the type of
variable provided. Supports weighted mean, sums, and specific
indicators.
}
\section{Methods (by class)}{
\itemize{
\item \code{region_value_method_qs(pct)}: The method for percentage variables.

\item \code{region_value_method_qs(count)}: The method for count variables.

\item \code{region_value_method_qs(ind)}: The method for \code{ind} variables.

\item \code{region_value_method_qs(ppo)}: The people per object (ppo) method.

\item \code{region_value_method_qs(default)}: The default method (works for dollar, sqkm, per1k, ...).
Simple weighted mean.

}}
