% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_geocode.R
\name{rev_geocode}
\alias{rev_geocode}
\title{Reverse geocode a location using \verb{https://photon.komoot.io/}}
\usage{
rev_geocode(lon, lat)
}
\arguments{
\item{lon}{<\code{numeric}> Represents the longitude of the location}

\item{lat}{<\code{numeric}> Represents the latitude of the location}
}
\value{
A character vector of the reverse geocoded location (address)
}
\description{
Reverse geocode a location using \verb{https://photon.komoot.io/}
}
\examples{
rev_geocode(12.4924, 41.8902) # "Piazza del Colosseo, Roma"
}
