% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_panel.R
\name{right_panel}
\alias{right_panel}
\title{Map page right panel}
\usage{
right_panel(id, ...)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{canale}.}

\item{...}{Additional arguments to be passed to the \code{\link[shiny]{absolutePanel}}
function. Would normally be the \code{compare_UI}, \code{explore_UI}, \code{dyk_UI}. TKTK LINK}
}
\value{
An absolutePanel object with a custom class for the content div.
}
\description{
This function creates a right panel for the Map page of Curbcut.
The panel is an Shiny absolute panel with a default panel class and two nested
div elements. The first div element has a custom class for styling purposes,
and can contain multiple other Shiny widgets or HTML elements.
}
