% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_unit.R
\name{round_big_marks}
\alias{round_big_marks}
\title{Round a numeric vector and apply a scale function based on its value magnitude}
\usage{
round_big_marks(x, min_dig, scale_fun = scales::comma)
}
\arguments{
\item{x}{<\verb{numeric vector}> Numeric values to be rounded and scaled.}

\item{min_dig}{<\code{numeric}> An integer specifying the minimum number of digits to be
displayed after rounding. If the value magnitude is larger than or equal to
6, 5, or 4, \code{accuracy} is set to \code{1e+03}, \code{1e+02}, or
\code{1e+01}, respectively, to round the value accordingly. Otherwise,
\code{accuracy} is set to \code{1} and \code{min_dig} is used to set the
rounding accuracy. \code{min_dig} is usually the output of \code{\link{min_sig_digits}}}

\item{scale_fun}{<\code{function}> A scale function to be applied to the rounded values. This
should be a function from the \code{scales} package that takes a numeric
vector as input and returns a character vector with formatted values. The
default is \code{scales::comma}, which adds commas to separate thousands
and rounds values to two decimal places by default.}
}
\value{
A character vector with the rounded and scaled values.
}
\description{
This function rounds a numeric vector and applies a scale function to format
it based on its value magnitude. The scale function is applied using the
\code{scales} package, which provides several options for formatting numeric
values, such as adding commas or dollar signs.
}
