% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s_extract}
\alias{s_extract}
\title{Extract substrings from a character vector that matches a regular
expression pattern}
\usage{
s_extract(pattern, x)
}
\arguments{
\item{pattern}{<\code{character}> A regular expression pattern to match}

\item{x}{<\code{character}> A vector from which to extract the substring}
}
\value{
The substrings from \code{x} that matches \code{pattern}, or
\code{NA} if no matches are found
}
\description{
It is a reimplementation of \code{\link[stringr]{str_extract}} in base R.
}
\examples{
s_extract("hello", c("hello world", "goodbye")) # hello"
s_extract("[0-9]+", c("123abc", "def456")) # "123", "456"
}
