% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s_extract_all}
\alias{s_extract_all}
\title{Extract all occurrences of a regular expression pattern from a character vector}
\usage{
s_extract_all(pattern, x)
}
\arguments{
\item{pattern}{<\code{character}> A regular expression pattern to match}

\item{x}{<\code{character}> A vector from which to extract the substring}
}
\value{
A character vector containing all occurrences of the regular
expression pattern found in the input vector.
}
\description{
It is a reimplementation of \code{\link[stringr]{str_extract_all}} in base R.
}
\details{
The function uses the \code{gregexpr} function to locate all occurrences of
the specified regular expression pattern in the input character vector \code{x}.
The resulting match positions are then passed to the \code{regmatches} function to
extract the matching substrings.
}
