% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s_sentence}
\alias{s_sentence}
\title{Capitalize the first letter of a sentence}
\usage{
s_sentence(x)
}
\arguments{
\item{x}{<\code{character}> A character string to capitalize}
}
\value{
A character string with the first letter capitalized
}
\description{
This function takes a string as input and returns the same string with the
first etter capitalized, assuming it is the first letter of a sentence. It does this by
extracting the first letter, capitalizing it, extracting the rest of the string,
and then combining the two parts.
}
\examples{
s_sentence("hello world") # "Hello world"
}
