% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeEvaluate}
\alias{safeEvaluate}
\title{Safely evaluate reactive and non-reactive arguments}
\usage{
safeEvaluate(argList)
}
\arguments{
\item{argList}{<\code{list}> A list containing both reactive and non-reactive
arguments. Reactive arguments are expected to be Shiny reactive expressions
that need to be evaluated to obtain their current values. Non-reactive
arguments are any other values or objects that do not require evaluation.}
}
\value{
A list of the same length as \code{argList}, where each element is the
evaluated value of the corresponding argument in \code{argList}. Reactive
expressions are evaluated to their current value, and non-reactive arguments
are returned unchanged.
}
\description{
This function iterates over a list of arguments and evaluates them. If an
argument is a reactive expression, it is evaluated to obtain its current
value. Non-reactive arguments are returned as is.
}
