% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_advanced.R
\name{settings_advanced}
\alias{settings_advanced}
\title{Advanced options UI and server interactions}
\usage{
settings_advanced(r = r)
}
\arguments{
\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}
}
\value{
NULL. This function is called for its side effects.
}
\description{
This function creates an advanced options modal UI with regional settings and
provides server-side handling of user interactions for regional preferences
and default location selections. It only gets triggered when the phantom
button "proxy_advanced_options" is clicked (which is when the settings button
is clicked on any page).
}
