% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_module.R
\name{settings_server}
\alias{settings_server}
\title{Server logic for the settings UI module}
\usage{
settings_server(id, r)
}
\arguments{
\item{id}{<\code{character}> The ID of the module, used to connect the UI and server
functions. Defaults to \code{"settings"}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}
}
\description{
This function contains the server-side logic for the settings UI module. It
only triggers a phantom button, which triggers the opening of the advanced
settings modal.
}
\seealso{
\code{\link{settings_UI}}
}
