% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_module.R
\name{settings_UI}
\alias{settings_UI}
\title{Create a settings menu UI for Curbcut}
\usage{
settings_UI(id = "settings")
}
\arguments{
\item{id}{A character string specifying the input ID for the settings menu.
Defaults to \code{"settings"}.}
}
\value{
A Shiny actionLink wrapped inside an icon_material_button with a
gear icon.
}
\description{
This function returns a settings menu in the form of an action link button
with a gear icon.The button triggers advanced options modal.
}
\seealso{
\code{\link{settings_server}}
}
