% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_module.R
\name{sidebar_server}
\alias{sidebar_server}
\title{Shiny module server for the sidebar component of a Curbcut page}
\usage{
sidebar_server(id, r)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file.}
}
\value{
A Shiny module server for the sidebar component of Curbcut
}
\description{
This function creates a Shiny module server for the sidebar component of
Curbcut. The sidebar component provides title and explanation on the page,
and other widgets are included in the \code{...} of the UI. The \code{modules} object
must live in the global environment, as it is from where the title texts
are retrieved.
}
