% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar_module.R
\name{sidebar_UI}
\alias{sidebar_UI}
\title{Create the UI for the sidebar component of Curbcut}
\usage{
sidebar_UI(id, ..., bottom = NULL)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{canale}.}

\item{...}{Additional UI elements to be included in the sidebar (widgets). To
be connected to the bookmarking, these must be Curbcut widgets, e.g.
\code{\link{picker_UI}}.}

\item{bottom}{A list of UI elements to be included at the bottom of the sidebar,
normally the UI of \code{\link{zoom_server}} and the
\code{\link{legend_server}}.}
}
\value{
A tag list containing the UI elements for the sidebar component of
Curbcut.
}
\description{
This function creates a Shiny module server for the sidebar component of
Curbcut. The sidebar component provides title and explanation on the page,
and other widgets are included in the \code{...} of the UI. in the \code{bottom}
argument of the \code{sidebar_UI} are the \code{\link{zoom_server}} and the
\code{\link{legend_server}}.
}
