% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_helpers.R
\name{single_scales_combination}
\alias{single_scales_combination}
\title{Map single scales to their longest combinations}
\usage{
single_scales_combination(avail_scale_combinations)
}
\arguments{
\item{avail_scale_combinations}{\if{html}{\out{<character vector>}} A vector of available
scale combinations.}
}
\value{
\if{html}{\out{<list>}} A list where each element represents the longest scale
combination for a single top scale.
}
\description{
This function maps each unique top scale in the available scale combinations
to its longest scale combination. It processes each top scale and finds the
combination with the most scales included.
}
