% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_slider.R
\name{slider_server}
\alias{slider_server}
\title{Create a Slider Input Server Module}
\usage{
slider_server(
  id,
  slider_id = "sld",
  value = shiny::reactive(NULL),
  min = shiny::reactive(NULL),
  max = shiny::reactive(NULL)
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{alp}.}

\item{slider_id}{<\code{character}> A character string giving the identifier for
the picker input object. This will be used as the input's \code{inputId} and will
inform the bookmarking. It needs to be unique in the page. The default value
is \code{"var"}.#' @param value A reactive expression representing the current value of the slider input element.}

\item{value}{<\code{reactive}> Represents the new value of the slider (to update).}

\item{min}{<\code{reactive}> Represents the minimum value of the slider (to update).}

\item{max}{<\code{reactive}> Represents the maximum value of the slider (to update).}
}
\value{
A Shiny server module for a slider input element.
}
\description{
This function creates a server module for a slider input element in Curbcut.
}
