% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_slider_text.R
\name{slider_text_server}
\alias{slider_text_server}
\title{Shiny module for a slider text input control}
\usage{
slider_text_server(
  id,
  r,
  slider_text_id = "slt",
  choices = shiny::reactive(NULL),
  selected = shiny::reactive(NULL)
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{slider_text_id}{<\code{character}> string giving the identifier for the slider input
object. This will be used as the input's \code{inputId} and will inform the bookmarking.
It needs to be unique in the page. The default value is \code{"slt"}.}

\item{choices}{<\verb{reactive character vector}> Choices that the user can
select from.}

\item{selected}{<\verb{reactive character}> Represents the selected value to update
to.}
}
\value{
A reactive value representing the user's selection.
}
\description{
Shiny module for a slider text input control
}
\seealso{
\code{\link{slider_text_UI}}
}
