% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_slider_text.R
\name{slider_text_UI}
\alias{slider_text_UI}
\title{Create a slider text UI element for selecting a value from a list of choices.}
\usage{
slider_text_UI(
  id,
  slider_text_id = "sldt",
  label = "Select a year",
  choices,
  selected = NULL,
  force_edges = TRUE,
  ...
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{canale}.}

\item{slider_text_id}{<\code{character}> string giving the identifier for the slider input
object. This will be used as the input's \code{inputId} and will inform the bookmarking.
It needs to be unique in the page. The default value is \code{"sldt"}.}

\item{label}{<\code{character}> A text label to display next to the checkbox input.
If NULL (default), no label will be displayed.}

\item{choices}{<\verb{character vector}> Choices that the user can select from.}

\item{selected}{<\code{character}> The initially selected value. If length > 1,
create a range slider.}

\item{force_edges}{<\code{logical}> Slider will be always inside it's container.}

\item{...}{Additional arguments to pass to the \code{\link{sliderTextInput}}
function.}
}
\value{
A \code{\link{sliderTextInput}} object that can be used as an input
in the app.
}
\description{
Create a slider text UI element for selecting a value from a list of choices.
}
\seealso{
\code{\link{slider_text_server}}
}
