% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget_slider.R
\name{slider_UI}
\alias{slider_UI}
\title{Create a Slider Input UI Element}
\usage{
slider_UI(
  id,
  slider_id = "sld",
  label = cc_t("Select a year"),
  step = 5,
  min,
  max,
  sep = "",
  value = max,
  width = "95\%",
  ...
)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the widget will appear,
e.g. \code{alp}.}

\item{slider_id}{<\code{character}> string giving the identifier for the slider input
object. This will be used as the input's \code{inputId} and will inform the bookmarking.
It needs to be unique in the page. The default value is \code{"sld"}.}

\item{label}{<\code{character}> The label text for the slider input element.
Defaults to "Select a year".}

\item{step}{<\code{numeric}> Specifies the interval between each selectable value
on the slider. Either NULL, the default, which uses a heuristic to determine
the step size or a single number. If the values are dates, step is in days;
if the values are date-times, step is in seconds. Defaults to 5 has it is the
gab between census years and accessibility calculations goes by steps of 5 too.}

\item{min}{<\code{numeric|date}> The minimum value of the slider.}

\item{max}{<\code{numeric|date}> The maximum value of the slider.}

\item{sep}{<\code{character}> The separator character used for displaying the
thousand places in numbers. Defaults to an empty string.}

\item{value}{<\code{numeric|date}> The initial value of the slider, either a
number, a date (class Date), or a date-time (class POSIXt). A length one
vector will create a regular slider; a length two vector will create a
double-ended range slider. Must lie between min and max. Defaults to \code{max}.}

\item{width}{<\code{character}> The width of the slider input element. Defaults to
"95\%".}

\item{...}{Additional arguments to be passed to \code{\link[shiny]{sliderInput}}}
}
\value{
A Shiny slider input element.
}
\description{
This function creates a slider input element for use in Curbcut using
\code{\link[shiny]{sliderInput}}
}
