% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_stories.R
\name{stories_server}
\alias{stories_server}
\title{Stories Server}
\usage{
stories_server(id, r)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}
}
\description{
This function is responsible for the server-side logic for handling stories
in the application. It manages the interaction, including sidebars, maps,
clicks, rendering stories, handling images and various reactive behaviors.
}
