% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories_module.R
\name{stories_UI}
\alias{stories_UI}
\title{Stories UI}
\usage{
stories_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}.}
}
\value{
A tagList object containing the UI elements.
}
\description{
This function creates the UI for handling stories within the application.
It includes the sidebar, map and main panel configurations and dynamically
sets themes and other attributes based on the provided stories data.
}
