% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_assign_any}
\alias{test_assign_any}
\title{Assign Values to specific Variables in the environment}
\usage{
test_assign_any(
  var_left = "housing_tenant",
  var_right = " ",
  region = "CMA",
  scale = "CSD",
  time = 2021,
  zoom_levels = get_from_globalenv("mzl_CSD_CT_DA"),
  select_id = NA,
  pos = 1,
  data_path = get_data_path()
)
}
\arguments{
\item{var_left}{<\verb{reactive character}> Character string of the selected
variable, e.g. \code{alp}}

\item{var_right}{<\verb{reactive character}> Character string of the selected
compared variable, e.g. \code{housing_value}. Defaults to what no compared
variable is represented by (" ").}

\item{region}{<\code{character}> The region. Defaults to CMA.}

\item{scale}{<\code{character}> The scale. Defaults to CSD.}

\item{time}{<\verb{numeric named list}> Time the user would be interested in. (represents
the widget time).}

\item{zoom_levels}{<\verb{named numeric vector}> One of \verb{mzl_*} with zoom levels (or
scales) under study.}

\item{select_id}{<\code{character}> The select_id to assign in the global environment.
Defaults to NA.}

\item{pos}{<\code{numeric}> An integer value indicating the position in the search list where
the environment to be used for assignment is located. Default is 1, which is
typically the global environment.}

\item{data_path}{<\code{character}> Path to a Curbcut repo data folder.
eg, \verb{C:/.../curbcut-montreal/data/}. Defaults to grabbing
.curbcut_montreal_data from the environment.}
}
\value{
This function doesn't return a value. It modifies the specified
environment by assigning values to certain variables.
}
\description{
This function is used to recreate a variable selection, data, df, ... in the
global environment, helping to test line-by-line functions.
}
\details{
This function uses the vars_build function to create a new set of
variables using var_left, var_right, and df. It then retrieves the
corresponding data using the data_get function.
It then assigns each of these variables, along with df, region (which is
the piece before the underscode in \code{df}), and select_id to the specified environment.
}
