% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_assign_bivar_dollar}
\alias{test_assign_bivar_dollar}
\title{Assign a dollar bivar vars/data to the global environment}
\usage{
test_assign_bivar_dollar(pos = 1)
}
\arguments{
\item{pos}{<\code{numeric}> Defaults to 1, the environment in which to assign the
values. This argument is to appease RMD check.}
}
\value{
Assigns \code{df}, \code{vars}, \code{data}
}
\description{
Assign a dollar bivar vars/data to the global environment
}
