% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_assign_bivar_ind_ord}
\alias{test_assign_bivar_ind_ord}
\title{Assign a ind ordinal bivar vars/data to the global environment}
\usage{
test_assign_bivar_ind_ord()
}
\value{
Assigns \code{df}, \code{vars}, \code{data} corresponding to a q5 index ordinal
variable to the global environment
}
\description{
Assign a ind ordinal bivar vars/data to the global environment
}
