% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_assign_q5_count}
\alias{test_assign_q5_count}
\title{Assign a q5 count vars/data to the global environment}
\usage{
test_assign_q5_count(pos = 1)
}
\arguments{
\item{pos}{<\code{numeric}> Defaults to 1, the environment in which to assign the
values. This argument is to appease RMD check.}
}
\value{
Assigns a constructed \code{df}, \code{vars}, \code{data} in the global environment
}
\description{
Assign a q5 count vars/data to the global environment
}
