% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_resources_creation}
\alias{test_resources_creation}
\title{Create resources folder for testing in curbcut package}
\usage{
test_resources_creation(additional_vars = c())
}
\value{
NULL. The function operates by creating and modifying folders and files
in the working directory.
}
\description{
This function copies a specified set of data files and associated metadata from
the "data" directory to a newly created "resources" directory. The resources
folder can then be cut in the curbcut test/ folder.
}
