% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_setup}
\alias{test_setup}
\title{Setup objects}
\usage{
test_setup()
}
\value{
Assigns objects in the global environment which are needed for testthat and
other testing functions.
}
\description{
Assigns objects in the global environment which are needed for testthat and
other testing functions.
}
