% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dropdown_module.R
\name{theme_dropdown_UI}
\alias{theme_dropdown_UI}
\title{Theme Dropdown UI}
\usage{
theme_dropdown_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{alp}.}
}
\value{
A \code{shiny::div} object that can be included in a Shiny UI.
}
\description{
This module generates a div container for a dropdown menu in Curbcut, with
menu items reflecting the 'theme' attributes of the different pages.
}
