% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_chr}
\alias{time_chr}
\title{Convert time value to character representation}
\usage{
time_chr(var, time_val)
}
\arguments{
\item{var}{<\verb{character vector}> String representing the code of the variable
to retrieve information for.}

\item{time_val}{<\code{numeric}> The time value to be checked.}
}
\value{
<\code{character}> Returns the name of the 'time_val' if its corresponding var in
the \code{dates} column in the variables table is named. If  not named, returns 'time_val' itself.
}
\description{
This function takes a variable and a time value, and returns the character
representation of the time value if it is named in the variables table. If the
\code{dates} column of the variable in the \code{variables} table is not a named vector,
it returns the time value itself.
}
