% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_slider_module.R
\name{time_slider_server}
\alias{time_slider_server}
\alias{time_slider_UI}
\title{Time Slider Server and UI Functions for R Shiny}
\usage{
time_slider_server(id, r)

time_slider_UI(id, min, max, step, double_value)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{min}{<\code{numeric}> The minimum value for the slider.}

\item{max}{<\code{numeric}> The maximum value for the slider.}

\item{step}{<\code{numeric}> Specifies the interval between each selectable value
on the slider. Either NULL, the default, which uses a heuristic to determine
the step size or a single number. If the values are dates, step is in days;
if the values are date-times, step is in seconds. Defaults to 5 has it is the
gap between census years and accessibility calculations goes by steps of 5 too.}

\item{double_value}{<\verb{numeric vector}> A vector of two initial values for
the bi-directional slider.}
}
\value{
A reactive value containing either the selected year or a vector of two years.
}
\description{
A Curbcut module that provides a user interface for selecting a single year or
a range of years. The output is a reactive value containing either the selected
year or a vector of two years.
}
\section{Functions}{
\itemize{
\item \code{time_slider_UI()}: Create the UI for the time slider module

}}
