% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_title.R
\name{title_page_get}
\alias{title_page_get}
\title{Get the title of the current page}
\usage{
title_page_get(active_page, site_name, lang = NULL)
}
\arguments{
\item{active_page}{<\code{character}> The ID of the current page}

\item{site_name}{<\code{character}> The name of the site, usually initiated
in \code{global.R}.}

\item{lang}{<\code{character}> The language used to translate the module's title
(optional). Defaults to NULL.}
}
\value{
The title of the current page
}
\description{
This function returns the title of the current page based on the active_page
ID. If the active_page is a valid module ID, the function will append the
module's title to the site_name and return it.
}
