% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{track_previous_reactive}
\alias{track_previous_reactive}
\title{Track Previous Reactive Values}
\usage{
track_previous_reactive(reactive_expr)
}
\arguments{
\item{reactive_expr}{<\verb{reactive expression}> The reactive expression to track.}
}
\value{
<\code{eventReactive}> An eventReactive object containing the previous value.
}
\description{
This function tracks and stores previous and current values of a reactive
expression. It returns an eventReactive object that holds the previous value.
}
