% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{translation_df}
\alias{translation_df}
\title{Translation dataframe (of front-facing Curbcut strings)}
\format{
\subsection{\code{translation_df}}{

A data.frame:
\describe{
\item{en}{The english version of all the front-facing strings present in the Curbcut package}
\item{fr}{The french translation of these strings}
}
}
}
\usage{
translation_df
}
\description{
Translation dataframe (of front-facing Curbcut strings)
}
\keyword{datasets}
