% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_helpers.R
\name{tutorial_trigger}
\alias{tutorial_trigger}
\title{Trigger the Curbcut map tutorial}
\usage{
tutorial_trigger(
  id,
  session,
  server_session,
  skip_elements,
  lang = NULL,
  window_size
)
}
\arguments{
\item{id}{<\code{character}> A string representing the ID of the page for which
the tutorial will be shown.}

\item{session}{<\code{session}> A shiny session object from the server function.
The actual session of the module, should be  \code{session = session}}

\item{server_session}{<\code{session}> A session object passed to the server function. It is used
to manage cookies in the server-side of the Shiny application. Usually
\code{server_session = r$server_session()}.}

\item{skip_elements}{<\verb{character vector}> A character vector specifying the IDs
of elements to be skipped in the tutorial. Default is NULL, meaning all elements
are included.}

\item{lang}{<\code{character}> A string specifying the language of the tutorial.
Defaults to NULL for no translation.}

\item{window_size}{<\verb{numeric vector}> Numeric vector, normally
\code{r$server_session()$input$dimension} giving window dimension. If on mobile (<768px),
the tutorial will not be shown.}
}
\value{
The function does not explicitly return a value. It triggers the tutorial
using \code{\link[rintrojs]{introjs}} and sets a cookie recording the last time the tutorial
was seen.
}
\description{
This function triggers a tutorial for a Curbcut map page. The tutorial
guides the user through the app using the rintrojs package, highlighting various
elements and providing custom instructions.
}
\details{
The function first builds a dataframe with elements to be highlighted
in the tutorial. Each element is identified by its ID, and associated with
an instruction and a title. If \code{skip_elements} is specified, these elements
are removed from the dataframe. The tutorial is then triggered using
\code{\link[rintrojs]{introjs}}, with instructions and titles drawn from the dataframe.
Finally, a cookie is set to record the last time the tutorial was seen.
}
