% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ui}
\alias{ui}
\title{Generate Curbcut Shiny User Interface}
\usage{
ui(
  web_description,
  web_title,
  twitter_handler,
  google_analytics,
  website_url,
  share_jpg
)
}
\arguments{
\item{web_description}{<\code{character}> The description of the website for metadata.
e.g. \code{"Curbcut est une plateforme d'exploration approfondie, dynamique et intuitive de la durabilité urbaine."}}

\item{web_title}{<\code{character}> The title of the website for metadata.
e.g. \verb{Curbcut Montréal | Vers une ville durable}}

\item{twitter_handler}{<\code{character}> The Twitter handle associated with the
website for metadata.e.g. \code{"@curbcutca"}}

\item{google_analytics}{<\code{character}> The Google Analytics script location
for tracking user activity. e.g. \code{"www/google_analytics.html"}}

\item{website_url}{<\code{character}> The URL of the website for metadata.
e.g. \code{"https://montreal.curbcut.ca"}}

\item{share_jpg}{<\code{character}> The URL of the image for sharing on social media.
Not relative paths. e.g. \code{"https://montreal.curbcut.ca/share.jpg"}}
}
\value{
A Shiny UI object that includes all elements of the Curbcut application interface.
}
\description{
This function generates the main Shiny UI for the Curbcut application.
The UI includes elements such as the import of package dependencies,
application styles, Curbcut scripts, Google Analytics, metadata for sharing,
and the creation of the pages UI.
}
