% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_functions.R
\name{update_lang}
\alias{update_lang}
\title{Update language}
\usage{
update_lang(r, lang)
}
\arguments{
\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{lang}{<\code{character}> Value of the new language (\code{"fr"} or \code{"en"})}
}
\value{
Sets the language to the value of \code{lang}
}
\description{
This function changes the active language dynamically based on the user
preferences. It both uses the JS \code{set_language} function to update which
\code{span} gets shown ("user- lang-en" or "user-lang-fr") and updates the \code{r$lang()}
value.
}
