% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_map_rv}
\alias{update_map_rv}
\title{Update a reactive value object}
\usage{
update_map_rv(id, r, rv_name, default_val, new_val)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{rv_name}{<\code{character}> A character string specifying the name of the
reactive value to create and update.}

\item{default_val}{<\code{vector}> A vector of length one. This parameter is used
to initialize the rv.}

\item{new_val}{<\code{reactive}> A reactive expression returning the new value to
assign to the rv. If the new value is the same as the current one, the function
does nothing.}
}
\value{
This function does not return a value. It updates the reactive value in place.
The updated value can be accessed using the rv object and the rv_name parameter.
If the new value is the same as the current one, the function does nothing and
returns NULL.
}
\description{
This function updates a reactive value object and does not trigger a change
in the reactive value if the new value is the same as the previous value. It
observes a new value, and if this new value is different from the current one,
the function updates the rv. It uses the reactive programming features of Shiny.
}
