% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_poi}
\alias{update_poi}
\title{Get nearby points of interests (POIs) based on map view state}
\usage{
update_poi(id, poi, map_viewstate)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this function will appear,
e.g. \code{alp}.}

\item{poi}{<\verb{character vector}> The current POIs showing on the map.}

\item{map_viewstate}{<\code{list}> The map viewstate. Usually the output of the
\code{\link{map_server}}, or of \code{\link[rdeck]{get_view_state}}.}
}
\value{
A character vector of nearby POIs if new POIs are found; otherwise,
returns the input POI vector.
}
\description{
This function retrieves the nearby POIs based on the current view state of a
map specified by the \code{map_id} parameter. The function filters the POIs
based on their proximity to the center of the map, keeping only those POIs
that are within 2000 meters. Currently only returns the points of interests
of the \code{stories} data.frame.
}
