% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_region}
\alias{update_region}
\title{Update the page's \code{region} reactive value in the reactive list \code{r}}
\usage{
update_region(id, r, new_region)
}
\arguments{
\item{id}{<\code{character}> The ID of the element in the reactive list \code{r} where
the \code{region} reactive value is stored.}

\item{r}{<\verb{reactive list}> A reactive list where the element identified by
\code{id} contains a named list with a \code{region} reactive value. The output
of \code{\link{r_init}}.}

\item{new_region}{<\code{reactive}> A reactive expression that returns the new
value for the \code{region}. The first output of the \code{\link{geography_server}}
function.}
}
\value{
An observer that updates the \code{region} reactive value in \code{r[[id]]}
whenever \code{new_region} changes.
}
\description{
This function observes changes in \code{new_region} and updates the \code{region}
reactive value within the reactive list \code{r} identified by the given \code{id}.
The update is triggered only if \code{new_region} is different from the current
\code{region} value stored in \code{r[[id]]}.
}
