% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_rv_general}
\alias{update_rv_general}
\title{Update a reactive value object}
\usage{
update_rv_general(rv, new_val)
}
\arguments{
\item{rv}{<\code{reactiveVal}> The reactive value to update.}

\item{new_val}{<\code{reactive}> A reactive expression returning the new value to
assign to the rv. If the new value is the same as the current one, the function
does nothing.}
}
\description{
This function updates a reactive value object and does not trigger a change
in the reactive value if the new value is the same as the previous value. It
observes a new value, and if this new value is different from the current one,
the function updates the rv. It uses the reactive programming features of Shiny.
}
