% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_scale}
\alias{update_scale}
\title{Get or update the \code{df} rv output}
\usage{
update_scale(tile, zoom_string)
}
\arguments{
\item{tile}{<\code{character}> a character string indicating the tile, the output
of the \code{\link{zoom_server}}}

\item{zoom_string}{<\code{character}> a character string indicating the zoom string,
the output of \code{\link{zoom_get_string}}}
}
\value{
a character string indicating the string of the data the user is
looking at. The combination of the region and the scale, e.g. \code{CMA_CSD}
}
\description{
The \code{update_df} function returns the zoom string if the tile is on auto-scale,
otherwise it returns the tile.
}
