% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_select_id_from_default}
\alias{update_select_id_from_default}
\title{Update Select ID Based on the default ID (Location lock in advanced settings)}
\usage{
update_select_id_from_default(data, default_select_ids, select_id)
}
\arguments{
\item{data}{<\code{data.frame}> A data frame containing the \code{ID} column to be
checked for any match with the \code{default_select_ids}}

\item{default_select_ids}{<\verb{character vector}> Vector of default IDs created
through  \code{\link{adv_opt_lock_selection}} present in the advanced
options, usually \code{r$default_select_ids()}}

\item{select_id}{<\code{character}> the current selected ID, usually \code{r[[id]]$select_id()}}
}
\value{
The updated selected ID based on the provided default IDs
}
\description{
This function updates the selected ID based on the provided default IDs
created through \code{\link{adv_opt_lock_selection}} present in the advanced
options, if any. If the provided default IDs are not present in the data,
the original select ID is returned.
}
