% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_select_id_helper}
\alias{update_select_id_helper}
\title{Update Select ID Helper}
\usage{
update_select_id_helper(new_ID, select_id)
}
\arguments{
\item{new_ID}{<\code{character}> A character vector indicating the new ID that's
been selected.}

\item{select_id}{<\code{character}> A character indicating the current selected ID.}
}
\value{
If the same ID gets selected twice, the function returns NA.
Otherwise, it returns the new ID.
}
\description{
Update Select ID Helper
}
