% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tab.R
\name{update_tab}
\alias{update_tab}
\title{Update the Active Tab in a Shiny App}
\usage{
update_tab(session, inputId = "cc_page", selected)
}
\arguments{
\item{session}{<\code{session}> The current Shiny session object. Usually \code{r$server_session()}.}

\item{inputId}{<\code{character}> The ID of the tabset panel to be updated. Default is "cc_page".}

\item{selected}{<\code{character}> The tab value to be selected. One entry as \code{id} in
the modules table.}
}
\value{
Updates the tabset panel in the Curbcut app to reflect the specified tab.
}
\description{
This function updates the currently active tab in a Shiny app based on the
given parameters. It first resets the query string to blank and then updates
the tabset panel to reflect the desired tab.
}
