% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_vars}
\alias{update_vars}
\title{Update the vars reactive value in a given reactive list r based on changes
in var_left and var_right}
\usage{
update_vars(id, r, var_left, var_right, widget_time)
}
\arguments{
\item{id}{<\code{character}> The id of the element in the reactive list r where
the vars reactive value is stored.}

\item{r}{<\verb{reactive list}> A reactive list where the element with the given
id (the current page) contains a named list with a vars reactive value, which stores the
output of the curbcut::vars_build function.}

\item{var_left}{<\verb{reactive character}> A reactive character string of the
selected variable, e.g. alp_2016 or c("housing_tenant_2006",
#' "housing_tenant_2016").}

\item{var_right}{<\verb{reactive character}> A reactive character string of the
selected compared variable, e.g. housing_value_2016.}

\item{widget_time}{<\verb{reactive vector}> The time selected by the user on the
time widget. Length 1 (single year) or 2 (compare years).}
}
\value{
An observer that updates the vars reactive value \code{r[[id]]$vars}
whenever var_left, var_right and or df changes.
}
\description{
This function is used to update the vars reactive value in a given reactive
list r whenever the input var_left, var_right and/or df changes. It uses the
\code{\link{vars_build}} function to determine the new vars value, and updates
it only if the new value is different from the current value.
}
