% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_zoom_levels}
\alias{update_zoom_levels}
\title{Update the page's \code{zoom_levels} reactive value in the reactive list \code{r}}
\usage{
update_zoom_levels(id, r, new_zl)
}
\arguments{
\item{id}{<\code{character}> The ID of the element in the reactive list \code{r} where
the \code{region} reactive value is stored.}

\item{r}{<\verb{reactive list}> A reactive list where the element identified by
\code{id} contains a named list with a \code{zoom_levels} reactive value. The output
of \code{\link{r_init}}.}

\item{new_zl}{<\code{reactive}> A reactive expression that returns the new
value for the \code{zoom_levels}. The second output of the \code{\link{geography_server}}
function.}
}
\value{
An observer that updates the \code{zoom_levels} reactive value in \code{r[[id]]}
whenever \code{new_zl} changes.
}
\description{
This function observes changes in \code{new_zl} and updates the \code{zoom_levels}
reactive value within the reactive list \code{r} identified by the given \code{id}.
The update is triggered only if \code{new_zl} is different from the current
\code{zoom_levels} value stored in \code{r[[id]]}.
}
