% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{use_curbcut_cookie}
\alias{use_curbcut_cookie}
\title{Load JavaScript cookie library and custom cookie script for Curbcut}
\usage{
use_curbcut_cookie()
}
\value{
A tag list object with the necessary JavaScript files for using cookies
in curbcut package. Must be placed in the \code{ui} function in \code{ui.R}.
}
\description{
This function generates an HTML tag list that loads the necessary JavaScript
files for using cookies in the curbcut package. It adds a script tag for a
custom cookie.js script and a script tag for the js-cookie library from
cdn.jsdelivr.net. It needs to be placed in the \code{ui} function in \code{ui.R} for
the cookies function to work.
}
