% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{use_curbcut_css}
\alias{use_curbcut_css}
\title{Add Curbcut CSS}
\usage{
use_curbcut_css(lang_init = "en")
}
\arguments{
\item{lang_init}{<\code{character}> The language that should be used as the default
language to display.}
}
\description{
This function adds the Curbcut shared CSS files by creating a resource path to
the curbcut package and linking to the CSS stylesheet in the HTML head.
}
