% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_functions.R
\name{use_curbcut_js}
\alias{use_curbcut_js}
\title{Use Curbcut javascript functions}
\usage{
use_curbcut_js()
}
\value{
A \code{script} tag to be placed in the \code{ui} function, \code{ui.R} file.
}
\description{
Curbcut package includes a few javascript functions to supplement the
available shiny functions. This function will import and declare them.
They can afterwards be used using the \code{shinyjs} package.
}
