% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_get_db.R
\name{val_get_db}
\alias{val_get_db}
\title{Retrieves data from a PostgreSQL database based on dynamic conditions}
\usage{
val_get_db(
  vars,
  grid,
  grid_compare,
  rv_zoom_string,
  highest_grd,
  select_id,
  time
)
}
\arguments{
\item{vars}{<\code{list}> A list of variables used for data retrieval, including
the left variable for the time comparison.}

\item{grid}{<\code{logical}> A boolean indicating if the grid mode is active.}

\item{grid_compare}{<\code{logical}> A boolean indicating if the grid comparison
mode is active.}

\item{rv_zoom_string}{<\code{character}> A string representing the current zoom
level.}

\item{highest_grd}{<\code{character}> A string representing the highest granularity
of grid cells available.}

\item{select_id}{<\code{character}> The ID of the selected element for which data
is being retrieved.}

\item{time}{<\code{list}> A list containing the left time variable for data
retrieval.}
}
\value{
<\code{mixed}> Returns the retrieved value(s) from the database if any of
the conditions are met, otherwise returns NULL.
}
\description{
This function dynamically retrieves data from a PostreSQL database based on
user-defined variables, a grid mode, comparison mode, zoom level, the highest
granularity level, a selected ID, and a time range. It is written for the``
grid pages in CUrbcut.
}
