% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_get_db.R
\name{val_get_db_helper}
\alias{val_get_db_helper}
\alias{val_get_db_helper.q5}
\alias{val_get_db_helper.delta}
\title{A helper function to dispatch the DB data retrieval process}
\usage{
val_get_db_helper(vars, select_id, time, scale)

\method{val_get_db_helper}{q5}(vars, select_id, time, scale)

\method{val_get_db_helper}{delta}(vars, select_id, time, scale)
}
\arguments{
\item{vars}{<\code{list}> A list of variables used for data retrieval, including
the left variable for the time comparison.}

\item{select_id}{<\code{character}> The ID of the selected element for which data
is being retrieved.}

\item{time}{<\code{list}> A list containing the left time variable for data
retrieval.}

\item{scale}{<\code{character}> The scale for which data should be extracted}
}
\value{
<\code{mixed}> Returns the data retrieved from the database.
}
\description{
This function acts as a dispatcher for the 'val_get_db' function, allowing
for method overloading based on the class of the 'vars' object.
}
\section{Methods (by class)}{
\itemize{
\item \code{val_get_db_helper(q5)}: The method for q5.

\item \code{val_get_db_helper(delta)}: The method for delta.

}}
