% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_manip.R
\name{var_closest_year}
\alias{var_closest_year}
\title{Return variable based on time and input}
\usage{
var_closest_year(input, time, variables = get_from_globalenv("variables"))
}
\arguments{
\item{input}{<\code{character}> A character string representing the variable code.}

\item{time}{<\code{numeric}> A numeric value indicating the time in years for
which the variable is needed. If \code{NULL}, the function returns the
\code{input} as is.}

\item{variables}{<\code{data.frame}> The \code{variables} df. Defaults to grabbing it
from the global environment using \code{\link{get_from_globalenv}}.}
}
\value{
A character string representing the variable code with the closest
year attached, or the original \code{input} if \code{time} is \code{NULL} or
\code{input} is an empty string.
}
\description{
This function returns a variable code based on the given \code{time} and
\code{input}. If \code{time} is \code{NULL}, it returns the \code{input} as
is. If \code{input} is an empty string (" "), it returns an empty string.
Otherwise, it finds the closest year in which the variable is available, and
attaches it to the variable code.
}
\details{
The function uses the \code{\link{var_get_info}} function to obtain the
dates at which the variable is available, and then finds the closest year to
the given \code{time} value.
}
