% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_get_breaks}
\alias{var_get_breaks}
\title{Get pretty or raw variable breaks}
\usage{
var_get_breaks(
  var,
  df,
  q3_q5 = "q5",
  break_col = "var",
  pretty = TRUE,
  compact = TRUE,
  lang = NULL
)
}
\arguments{
\item{var}{<\code{character}> String representing the code of the variable
to retrieve the breaks for.}

\item{df}{<\code{character}> Indicates the combination of the region and scale
of interest, e.g. \code{"CMA_DA"}}

\item{q3_q5}{<\code{character}> String indicating whether to return Q3 or Q5 breaks.
Defaults to "q5".}

\item{break_col}{<\code{character}> Which column in the breaks_qx column of the
\code{variables} table should the break be taken from. Defaults to \code{var} which
is the numeric breaks for any numeric variable. Other possibility is
\code{var_name} or \code{var_name_short} for qualitative variables.}

\item{pretty}{<\code{logical}> Indicates whether to return pretty-formatted breaks
using the \code{\link[curbcut]{convert_unit}} function. Defaults to \code{TRUE}.}

\item{compact}{<\code{logical}> Indicates whether to compact the number representation
of breaks. Only applies when pretty is \code{TRUE}. Defaults to \code{TRUE}.}

\item{lang}{<\code{character}> String indicating the language to translate the
breaks to. Defaults to \code{NULL}, which is no translation.}
}
\value{
A vector of breaks for the specified variable, region, and scale.
If pretty is TRUE, the breaks are returned as characters and formatted in an
appropriate format for display.
}
\description{
This function returns the breaks for a given variable, region, scale and date
for q3 (directly extracted from \code{var}) as either raw or pretty. Raw breaks are
always given in the base unit of the variable. Pretty breaks are returned in an
appropriate format for display using \code{\link[curbcut]{convert_unit}}, i.e.,
with the appropriate unit prefix and separator depending on their magnitude.
The pretty formatting is optional, and the function defaults to returning pretty
breaks.
}
