% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_manip.R
\name{var_get_parent_info}
\alias{var_get_parent_info}
\title{Get information about the parent variable of a variable from the \code{variables}
table}
\usage{
var_get_parent_info(
  var,
  what = "explanation",
  translate = FALSE,
  lang = NULL,
  check_year = TRUE
)
}
\arguments{
\item{var}{<\verb{character vector}> String representing the code of the variable
to retrieve the parent variable information for.}

\item{what}{<\code{character}> String indicating the column name in the
\code{variables} table to retrieve information from. Defaults to \code{"var_title"}.}

\item{translate}{<\code{logical}> Indicating whether or not to translate the retrieved
information using the \code{\link{cc_t}} function. Defaults to \code{FALSE}.}

\item{lang}{<\code{character}> String indicating the language to translate to, if
\code{translate} is TRUE. If not specified, the function will not attempt to translate.}

\item{check_year}{<\code{logical}> Should the year be removed from \code{var} to grab
variable's info? Defaults to TRUE}
}
\value{
The requested information about the parent variable, with optional
translation using the \code{\link{cc_t}} function.
}
\description{
This function retrieves information about the parent variable of a given
variable from the \code{variables} table in the global environment.
}
\details{
It is using the \code{\link{var_get_info}} function to retrieve the correct
information.
}
