% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_get_time}
\alias{var_get_time}
\title{Extracts the \code{time} component from variable codes}
\usage{
var_get_time(var)
}
\arguments{
\item{var}{<\verb{character vector}> String representing the variable codes.}
}
\value{
A character string representing the year component of the variable
name.
}
\description{
Extracts the \code{time} component from variable codes
}
\examples{
var_get_time("housing_tenant_2016") # 2016
var_get_time(c("housing_tenant_2016", "housing_tenant_2021")) # "2016" "2021"
}
