% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_get_title}
\alias{var_get_title}
\title{Get the title of a variable code}
\usage{
var_get_title(var, short_treshold = NULL, translate = FALSE, lang = NULL)
}
\arguments{
\item{var}{<\verb{character string}> String representing the code of the variable
to retrieve the title for.}

\item{short_treshold}{<\code{numeric}> An optional threshold for the title length,
above which the short title will be returned.}

\item{translate}{<\code{logical}> Indicates whether to translate the title to a
different language.}

\item{lang}{<\code{character}> String indicating the language to translate the title to.}
}
\value{
A character string representing the variable title.
}
\description{
Given a variable name, get the title of the variable from the 'variables' table.
If a 'short_treshold' is provided, return the short title if the title length
exceeds the treshold. The returned title can be translated to a different
language using the 'lang' argument.
}
