% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_manip.R
\name{var_remove_time}
\alias{var_remove_time}
\title{Remove the time component from variable codes}
\usage{
var_remove_time(var)
}
\arguments{
\item{var}{<\verb{character vector}> String representing the variable codes.}
}
\value{
A character string with the variable code without the time component.
If the input was a variable code duplicate of different year, the output is
the unique variable code.
}
\description{
This function removes the time component from variable codes The time
component is identified by the "_YYYY" format at the end of the variable name.
}
\examples{
var_remove_time("housing_tenant_2016") # "housing_tenant"
var_remove_time(c("housing_tenant_2016", "housing_tenant_2021")) # "housing_tenant"
}
