% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_manip.R
\name{var_row_index}
\alias{var_row_index}
\title{Returns the row number of a given variable in the \code{variables} table.}
\usage{
var_row_index(var)
}
\arguments{
\item{var}{<\code{character}> Representing the variable code to look up.}
}
\value{
If \code{var} is found in the \code{variables} table, returns its
corresponding row number. If not found, returns the original value of
\code{var}.
}
\description{
This function takes a variable code and returns the corresponding row number
in the \code{variables} table. If the variable is not found in the \code{variables}
table, it returns the original value of the input. It first removes any
appended time values from the input, and then looks up the variable in
the \code{variables} table.
}
