% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_build.R
\name{vars_build}
\alias{vars_build}
\title{Determine the data type (class) of different \code{var_left}, \code{var_right} and \code{df}
combinations}
\usage{
vars_build(
  var_left,
  var_right = " ",
  df,
  scales_as_DA = c("building", "street"),
  check_choropleth = TRUE,
  variables = get_from_globalenv("variables")
)
}
\arguments{
\item{var_left}{<\verb{reactive character}> Character string of the selected
variable, e.g. \code{alp_2016} or \code{c("housing_tenant_2006", "housing_tenant_2016")}.}

\item{var_right}{<\verb{reactive character}> Character string of the selected
compared variable, e.g. \code{housing_value_2016}. Defaults to what no compared
variable is represented by (" ").}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales} that
should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.+}

\item{variables}{<\code{data.frame}> The \code{variables} df. Defaults to grabbing it
from the global environment using \code{\link{get_from_globalenv}}.}
}
\value{
A named list containing both \code{var_left} and \code{var_right} variables with
a class attached.
}
\description{
Determine the data type (class) of different \code{var_left}, \code{var_right} and \code{df}
combinations
}
