% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_dep}
\alias{vec_dep}
\title{Compute the depth of a nested data structure}
\usage{
vec_dep(x)
}
\arguments{
\item{x}{A vector, a list or a NULL.}
}
\value{
An integer indicating the depth of the data structure.
}
\description{
This function returns the depth of a nested data structure. The depth of a
vector is the number of lists it is nested within. Atomic vectors have a
depth of 1, and NULL has a depth of 0. It is a reimplementation of
\code{\link[purrr]{vec_depth}} in base R.
}
\examples{
vec_dep(NULL) # 0
vec_dep(1:10) # 1
vec_dep(list(1, list(2, 3))) # 2

}
\seealso{
The purrr package provides a vec_depth function with similar
functionality.
}
