% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnuser_get.R
\name{warnuser_get}
\alias{warnuser_get}
\title{Generate Warning Text for the Warn User Module}
\usage{
warnuser_get(vars, data, time, more_text = NULL, lang = NULL)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{data}{<\code{data.frame}> Data frame containing all the scale and
the \code{var_left} and \code{var_right}. The output of \code{\link{data_get}}.}

\item{time}{<\verb{numeric named list}> Vector of time values to use for
appending a time to the variables picked.}

\item{more_text}{<\code{character}> Optional additional text to be added to the
warning message. Default is NULL.}

\item{lang}{<\code{character}> The language to be used for the warning messages.
Defaults to NULL for no translation}
}
\value{
A string containing the warning message in HTML format.
}
\description{
This function generates warning text based on the provided conditions.
It is used internally by the \code{\link{warnuser_server}} function.
}
