% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnuser_module.R
\name{warnuser_server}
\alias{warnuser_server}
\alias{warnuser_UI}
\title{Warn User Module}
\usage{
warnuser_server(id, r, vars, time, data, more_text = shiny::reactive(NULL))

warnuser_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the legend will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{time}{<\verb{reactive numeric vector}> Vector of time values to use for
appending a time to the variables picked.}

\item{data}{<\verb{reactive data.frame}> Data frame containing all the scale and
the \code{var_left} and \code{var_right}. The output of \code{\link{data_get}}.}

\item{more_text}{<\verb{reactive character}> A reactive expression containing optional
additional text to be added to the warning message. Default is shiny::reactive(NULL).}
}
\value{
The warnuser_UI function returns a tag list for inclusion in a Shiny UI,
while the warnuser_server function returns the server function.
}
\description{
This module provides a server function and a UI function to warn users
based on specific conditions. The warnings are displayed as text.
}
\section{Functions}{
\itemize{
\item \code{warnuser_UI()}: Create the UI for the warnuser module

}}
