% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Calculate Weighted Mean}
\usage{
weighted_mean(x, w, ..., na.rm = FALSE)
}
\arguments{
\item{x}{<\verb{numeric vector}> A numeric vector for which the weighted mean is
to be computed.}

\item{w}{<\verb{numeric vector}> A numeric vector of weights, where each weight
corresponds to the elements in \code{x}. The length of \code{w} should match the
ength of \code{x}.}

\item{...}{Additional arguments passed to \code{stats::weighted.mean}.}

\item{na.rm}{<\code{logical}> A logical value indicating whether NA values in \code{x}
and \code{w} should be stripped before the computation proceeds. Defaults to
\code{FALSE}.}
}
\value{
A numeric value representing the weighted mean of the elements in \code{x}.
}
\description{
This function computes the weighted mean of a numeric vector, with the option
to remove NA values on both x and w. The calculation is based on weights provided.
}
\examples{
x <- c(1, 2, 3, 4, NA)
w <- c(1, 2, 3, 4, 5)
weighted_mean(x, w) # NA due to NA in `x`
weighted_mean(x, w, na.rm = TRUE) # weighted mean excluding NA

}
