% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_utils.R
\name{zoom_get_code}
\alias{zoom_get_code}
\title{Get the scale code based on the name of the scales}
\usage{
zoom_get_code(scales_name, lang = NULL)
}
\arguments{
\item{scales_name}{<\verb{character vector}> The name of the scales, e.g.
\code{c("Borough/City", "Census Tracts", ...)}}

\item{lang}{<\code{character}> String indicating the language to translate the
slider titles to. Defaults to \code{NULL}, which is no translation.}
}
\value{
The corresponding scale code(s).
}
\description{
This function takes in the name of the scales and returns the corresponding
scale code. If a translation is needed, it uses the provided language to
translate the scale names before returning the right scale code.
}
