% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_utils.R
\name{zoom_get_name}
\alias{zoom_get_name}
\title{Get the zoom name for a set of scale codes}
\usage{
zoom_get_name(scales, lang = NULL)
}
\arguments{
\item{scales}{<\verb{character vector}> A character vector of scales codes}

\item{lang}{<\code{character}> String indicating the language to translate the
slider titles to. Defaults to \code{NULL}, which is no translation.}
}
\value{
A character vector of slider titles.
}
\description{
This function takes a set of scale codes ("CMA", "CT", "DA") and returns the
corresponding slider title ("Borough/City" , "Census tract", "Dissemination area")
in the same order in which it was fed.
}
