% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_helpers.R
\name{zoom_level_selection}
\alias{zoom_level_selection}
\title{Internal function for dispatching \code{zoom_level_selection}.}
\usage{
zoom_level_selection(
  var_right,
  top_scale,
  avail_scale_combinations,
  scales_as_DA = c("building", "street"),
  ...
)
}
\arguments{
\item{var_right}{<\code{character}> Character string of the selected
compared variable, e.g. \code{housing_value}. It will be used to check on all
possible scale combinations which combinations have scales that all have
valid data.}

\item{top_scale}{<\code{character}>}

\item{avail_scale_combinations}{<\verb{character vector}>}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. They won't
be taken into account when all the scales need to hold data.}

\item{...}{Additional arguments to be passed to the methods.}
}
\value{
A ggplot object that represents the \code{delta} legend.
}
\description{
Internal function for dispatching \code{zoom_level_selection}.
}
