% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_module.R
\name{zoom_server}
\alias{zoom_server}
\alias{zoom_UI}
\title{zoom_server Function}
\usage{
zoom_server(
  id,
  r = r,
  zoom_string,
  zoom_levels,
  suffix_zoom_levels = NA,
  no_autozoom = shiny::reactive(FALSE)
)

zoom_UI(id, zoom_levels)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this module will appear,
e.g. \code{canale}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{zoom_string}{<\verb{reactive character}> A reactive object representing the
current zoom level, e.g. \code{CSD}.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-zoom).}

\item{suffix_zoom_levels}{<\code{character}> Suffix that should be appended at the
end of an auto-zoom. If the maximum zoom possible is CT and we previously created
such a tileset, we could have the suffix_zoom_levels \code{max_CT}, which would
be appended at the end of the auto-zoom tile name. Defaults to NA for none.}

\item{no_autozoom}{<\verb{reactive logical}> Whether the output tile() reactive
should include autozooms or not.}
}
\value{
A reactive object representing the current tile that should be
displayed on the map.
}
\description{
This function creates a module for Curbcut that allows users to control
the zoom level of a map using a slider input. The function disables the slider
if the user selects auto-zoom. The slider is updated if the available
zoom levels change. If on auto-zoom, the slider gets updated when the zoom
levels threshold are attained. When there is a change in the \code{r$region()},
the auto-zoom gets activated to provide better transition between regions.
The function then creates a reactive tile that returns the appropriate tile
based on the user's selection. If auto-zoom is selected, the function returns
the auto-zoom tile. If a specific scale is selected, the function checks to
see if the scale is available for the current region. If it is not
available, the function defaults back to auto-zoom.
}
\section{Functions}{
\itemize{
\item \code{zoom_UI()}: Create the UI for the zoom module

}}
