% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{cc_t}
\alias{cc_t}
\title{Translate an object between English and French}
\usage{
cc_t(..., .envir = parent.frame(), lang = NULL)
}
\arguments{
\item{...}{<\verb{character objects}> Any objects, including lists or atomic vectors}

\item{.envir}{The parent environment for evaluating the expressions in
\code{...}. Necessary to grab the correct value of variables handled by
\code{\link[glue]{glue}}. Defaults to \code{parent.frame()}}

\item{lang}{<\code{character}> Language to use for translation. Must be one of
en' or 'fr'. Defaults to NULL which is no translation.}
}
\value{
If running in a Shiny context (UI), then return spans in both languages.
If in a Shiny context and in server side, returns translation depending on
\code{lang}. Outside a Shiny context, returns the input as is.
}
\description{
Translate input object from English to French depending on the provided
language and whether the function is running in a Shiny environment.
}
\seealso{
\code{\link[curbcut]{cc_t_list}} for translating lists of input
objects
}
