% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colours_get}
\alias{colours_get}
\title{Retrieve the \code{colours} object}
\usage{
colours_get()
}
\value{
The \code{colours} object from the global environment
}
\description{
This function retrieves the \code{colours} object from the global environment. The
\code{colours} object contains pre-defined color schemes that can be used in
various plots across the package.
}
