% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_unit}
\alias{convert_unit}
\title{Convert a numeric vector to appropriate display units}
\usage{
convert_unit(x, var = NULL, compact = FALSE)
}
\arguments{
\item{x}{<\verb{numeric vector}> Numeric values.}

\item{var}{A variable code present in the \code{variables} table. If supplied, the
function looks up the variable type and returns the numbers in the appropriate
format. This argument is optional.}

\item{compact}{A logical value indicating whether to use compact notation for
large numbers. If \code{TRUE} and a number has at least 4 significant digits,
the function will use compact notation (e.g., 1.2M instead of 1200000). This
argument is optional and defaults to \code{FALSE}.}
}
\value{
A character vector of the same length as \code{x} representing the
numbers in an appropriate format for display.
}
\description{
This function takes a vector of numbers and returns a character vector that
represents the numbers in an appropriate format for display. If the \code{var}
argument is supplied, the function looks up the variable type and returns the
numbers in the appropriate format (e.g., as percentages or dollars).
}
\details{
The function uses the \code{scales} package to format the numbers.
By default, it formats numbers with commas and a varying number of decimal
places depending on the magnitude of the numbers. If the \code{var} argument
is supplied and the variable type is \code{"pct"} or \code{"dollar"}, the
function uses the corresponding function from the \code{scales} package to
format the numbers.
}
