% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.bivar_ldelta_rq3}
\alias{data_get.bivar_ldelta_rq3}
\title{Get data from the SQLite database for a \code{bivar_ldelta_rq3} class}
\usage{
\method{data_get}{bivar_ldelta_rq3}(vars, df, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link[curbcut]{vars_build}} function.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link[curbcut]{df_get}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the percentage change between two
years of one variable with the static one year value of another variable.
\code{q3}for the percent change value is  calculated on the spot and used
to regroup with the \code{q3} column of the second value to create the \code{group}
column for map colouring.
}
\description{
Get data from the SQLite database for a \code{bivar_ldelta_rq3} class
}
