% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.default}
\alias{data_get.default}
\title{Default data method}
\usage{
\method{data_get}{default}(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with an unknown class.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A dataframe with \code{NA} values
}
\description{
This is the default data method, which returns all possible columns filled
with NA values. It is intended to be used when no specific get data method
is available or necessary. In this case, \code{NA} takes place.
}
