% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.q5}
\alias{data_get.q5}
\title{Get data for a \code{q5} class}
\usage{
\method{data_get}{q5}(
  vars,
  scale,
  region = NULL,
  scales_as_DA = c("building", "street"),
  data_path = "data/",
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function. The class of the vars object is
used to determine how to grab de data and output it.}

\item{scale}{<\code{character}> The scale of the data to be retrieved, e.g. \code{CSD}.
The output of \code{\link{update_scale}}.}

\item{region}{<\code{character}> The region from which to grab the vector of IDs with
which to filter the retrieved data for a specific region. Defaults to NULL
if no filtering is needed.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By
default, their colour will be the one of their DA.}

\item{data_path}{<\code{character}> A string representing the path to the
directory containing the QS files. Default is "data/".}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the data, with an ID column, one column per
year of data, and one \code{group} column per year of data.
}
\description{
Get data for a \code{q5} class
}
