% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get.q5}
\alias{data_get.q5}
\title{Get data from the SQLite database for a \code{q5} class}
\usage{
\method{data_get}{q5}(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link[curbcut]{vars_build}} function.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the data fresh out of the sqlite db, with an
added \code{group} column for map colouring.
}
\description{
Get data from the SQLite database for a \code{q5} class
}
