% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get}
\alias{data_get}
\alias{data_get.q5}
\alias{data_get.bivar}
\alias{data_get.delta}
\alias{data_get.delta_bivar}
\alias{data_get.bivar_ldelta_rq3}
\alias{data_get.default}
\title{Get data}
\usage{
data_get(
  vars,
  scale,
  region,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)

\method{data_get}{q5}(
  vars,
  scale,
  region = NULL,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)

\method{data_get}{bivar}(
  vars,
  scale,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)

\method{data_get}{delta}(
  vars,
  scale,
  region,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  time,
  ...
)

\method{data_get}{delta_bivar}(
  vars,
  scale,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)

\method{data_get}{bivar_ldelta_rq3}(
  vars,
  scale,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)

\method{data_get}{default}(
  vars,
  scale,
  region = NULL,
  scales_as_DA = c("building", "street"),
  data_path = get_data_path(),
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function. The class of the vars object is
used to determine how to grab de data and output it.}

\item{scale}{<\code{character}> The scale of the data to be retrieved, e.g. \code{CSD}.
The output of \code{\link{update_scale}}.}

\item{region}{<\verb{character vector}> A vector of IDs with which to filter the
retrieved data for a specific region, probably retrieved from
\code{regions_dictionary$scales}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By
default, their colour will be the one of their DA.}

\item{data_path}{<\code{character}> A string representing the path to the
directory containing the QS files. Default is "data/".}

\item{...}{Additional arguments passed to methods.}

\item{time}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function. It contains the time for both var_left and var_right variables.}
}
\value{
A dataframe containing the data according to the class of \code{vars},
with an ID column, one column per year of data, and one \code{group} column per
year of data.
}
\description{
This function retrieves data from QS files on disk or an SQLite database
using the appropriate method based on the class of the input vars object.
vars should be a named list with a class, built using the
\code{\link{vars_build}} function. Depending on the class of vars, different
methods will be used to retrieve and process the data.
}
\section{Methods (by class)}{
\itemize{
\item \code{data_get(q5)}: The method for q5.

\item \code{data_get(bivar)}: The method for bivar.

\item \code{data_get(delta)}: The method for delta.

\item \code{data_get(delta_bivar)}: The method for bivar.

\item \code{data_get(bivar_ldelta_rq3)}: The method for bivar_ldelta_rq3.

\item \code{data_get(default)}: The default method.

}}
