% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get}
\alias{data_get}
\title{Get data from the SQLite database}
\usage{
data_get(vars, df, scales_as_DA = c("building", "street"), ...)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function. The class of the vars object is
used to determine how to grab de data and output it.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A dataframe containing the data according to the class of \code{vars}
along with a \code{group} column for map colouring.
}
\description{
This function retrieves data from an SQLite database using the appropriate
method based on the class of the input vars object. vars should be a named
list with a class, built using the \code{\link{vars_build}} function.
Depending on the class of vars, different methods will be used to retrieve
and process the data.
}
