% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_delta}
\alias{data_get_delta}
\title{Calculate the percentage change between two variables over two years}
\usage{
data_get_delta(var_two_years, df)
}
\arguments{
\item{var_two_years}{<\verb{character vector}> A character vector of length 2,
where each element is the name of a variable to compare,
e.g. \code{c("housing_tenant_2006", "housing_tenant_2016")}}

\item{df}{<\code{character}> A string specifying the name of the database to retrieve
data from. Combination of the region and the scale, e.g. \code{CMA_DA}.}
}
\value{
A data frame with the following columns: ID, var_1, var_2, and var.
\code{ID} is the ID column from the original data, \code{var_1} and \code{var_2} are the
values of the two variables being compared, and \code{var} is the percentage
change between the two variables.
}
\description{
This function takes two variables representing the same quantity measured two
years apart and calculates the percentage change between the two values.
}
