% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_get.R
\name{data_get_delta_fun}
\alias{data_get_delta_fun}
\title{Inner function to get data based on the type of \code{vars}}
\usage{
data_get_delta_fun(
  vars,
  select_id,
  df,
  data,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  data_path,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{data_path}{<\code{character}> A string representing the path to the directory
containing the QS files. Default is "data/".}

\item{...}{Additional arguments passed to methods.}
}
\description{
This function dispatches the data retrieval based on the class
of the \code{vars} object.
}
\seealso{
\code{\link{data_get.delta}}
}
