% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyk_module.R
\name{dyk_server}
\alias{dyk_server}
\alias{dyk_UI}
\title{Did You Know Server and UI Module}
\usage{
dyk_server(id, r, vars, df, select_id, poi = shiny::reactive(NULL))

dyk_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the module will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function.}

\item{df}{<\verb{reactive character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{poi}{<\code{reactive}> (Optional) Point of interests. The output of
\code{\link{update_df}}. Default is NULL.}
}
\value{
A Shiny module server function for the DYK module.
}
\description{
This function provides a server module for the "Did You Know" (DYK) section
of the a Curbcut map page. It retrieves and displays interesting facts, based
on \code{vars} and \code{poi} (point of interests). The server module retrieves DYKs and
handles the click events on the links.
}
\section{Functions}{
\itemize{
\item \code{dyk_UI()}: Create the UI for the legend module

}}
