% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_context}
\alias{explore_context}
\title{Get region text info}
\usage{
explore_context(region, select_id, df, switch_DA)
}
\arguments{
\item{region}{<\code{character}> String specifying the code of the region to
retrieve, e.g. \code{CMA}. Usually equivalent of \code{r$region()}.}

\item{select_id}{<\code{character}> the current selected ID, usually
\code{r[[id]]$select_id()}. If there is a selection (select_id is not NA), the
name of the selected polygon will appear.}

\item{df}{<\code{character}> The combination of the region under study and the
scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{switch_DA}{<\code{logical}> Is the \code{df} part of the scales that should be
switched as DAs instead.}
}
\value{
A list containing multiple texts used for the explore text panel.
}
\description{
This function retrieves the region df from the global environment given the
region name and returns it in a list with the first letter of the region name
capitalized.
}
