% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_q5_ind.R
\name{explore_graph_q5_ind}
\alias{explore_graph_q5_ind}
\alias{explore_graph_q5_ind.scalar}
\alias{explore_graph_q5_ind.ordinal}
\title{Explore Graph Function for \code{INDEX} type}
\usage{
explore_graph_q5_ind(
  vars,
  select_id,
  scale,
  data,
  time,
  schemas,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph_q5_ind}{scalar}(
  vars,
  select_id,
  scale,
  data,
  time,
  schemas,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)

\method{explore_graph_q5_ind}{ordinal}(
  vars,
  select_id,
  scale,
  data,
  time,
  schemas,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  font_family = "acidgrotesk-book",
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of variable codes with classes. The
output of \code{\link{vars_build}}.}

\item{select_id}{<\code{character}> A string indicating the ID of the currently
selected region (if any). Usually \code{r[[id]]$select_id()}}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations. The output of \code{\link{data_get}}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{schemas}{<\verb{named list}> Current schema information. The additional widget
values that have an impact on which data column to pick. Usually \code{r[[id]]$schema()}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their graph will be the one of their DA.}

\item{lang}{<\code{character}> A character string indicating the language to
translate variable titles to.}

\item{font_family}{<\code{character}> A string specifying the font family for the
plot, default is "acidgrotesk-book".}

\item{...}{Additional arguments passed to the specific method.}

\item{df}{<\code{character}> The combination of the region under study and the
scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_scale}}.}
}
\value{
A ggplot2 object representing the plot.
}
\description{
This function creates a ggplot for the explore panel of Curbcut, based on the
input parameters, for the index type variables.
}
