% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_graph_scales.R
\name{explore_graph_scale_ind}
\alias{explore_graph_scale_ind}
\alias{explore_graph_scale_ind.scalar}
\alias{explore_graph_scale_ind.ordinal}
\title{Explore Graph Scale Indice}
\usage{
explore_graph_scale_ind(var, x_y, scale, limits, lang = NULL, ...)

\method{explore_graph_scale_ind}{scalar}(var, x_y, scale, limits, lang = NULL, data_vals, ...)

\method{explore_graph_scale_ind}{ordinal}(
  var,
  x_y,
  scale,
  limits = NULL,
  lang = NULL,
  breaks,
  ...
)
}
\arguments{
\item{var}{<\code{character}> The variable with its type as a class, either "scale",
"ordinal".}

\item{x_y}{<\code{character}> x or y axis?}

\item{scale}{<\code{character}> Current scale, e.g. \code{"CSD"}.}

\item{limits}{<\code{numeric}> The x or y axis limits. Usually one of the breaks
attributes coming from from \code{data}.}

\item{lang}{<\code{character} Which language should the label be translated to.
Defaults to NULL for no translation.}

\item{...}{Additional arguments passed to the underlying ggplot2 functions.}

\item{data_vals}{<\verb{numeric vector}> Data values}

\item{breaks}{<\code{vector}> Current breaks}
}
\value{
A list containing the ggplot2 scale object generated by the specialized
method.
}
\description{
This function generates a scale for a ggplot2 graph based on the variable \code{var}
and the coordinate axis specified in \code{x_y}. The function dispatches to
specialized methods based on the type of \code{var}, either scalar or ordinal.
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_graph_scale_ind(scalar)}: scalar method for index values

\item \code{explore_graph_scale_ind(ordinal)}: ordinal method for index values

}}
