% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text}
\alias{explore_text}
\alias{explore_text.q5}
\alias{explore_text.bivar}
\alias{explore_text.delta}
\alias{explore_text.delta_bivar}
\title{Generate Explore text for the given variables and region}
\usage{
explore_text(
  vars,
  region,
  select_id,
  scale,
  time,
  data,
  scales_as_DA,
  zoom_levels,
  lang,
  ...
)

\method{explore_text}{q5}(
  vars,
  region,
  select_id,
  scale,
  time,
  data,
  zoom_levels,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  ...
)

\method{explore_text}{bivar}(
  vars,
  region,
  select_id,
  scale,
  data,
  time,
  zoom_levels,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  ...
)

\method{explore_text}{delta}(
  vars,
  region,
  select_id,
  scale,
  data,
  time,
  zoom_levels,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  ...
)

\method{explore_text}{delta_bivar}(
  vars,
  region,
  select_id,
  scale,
  data,
  zoom_levels,
  scales_as_DA = c("building", "street"),
  lang = NULL,
  ...
)
}
\arguments{
\item{vars}{<\code{character}> A list containing the variable names for which the
text needs to be generated. Usually the output of \code{\link{vars_build}}.}

\item{region}{<\code{character}> String specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{select_id}{<\code{character}> A string indicating the ID of the currently
selected region (if any). Usually \code{r[[id]]$select_id()}}

\item{scale}{<\verb{reactive character}> Current scale. The output of
\code{\link{update_scale}}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{data}{<\code{data.frame}> A data frame containing the variables and
observations. The output of \code{\link{data_get}}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By
default, their info will be the one of their DA.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \verb{mzl_*}, or the output of
\code{\link{geography_server}}.}

\item{lang}{<\code{character}> A string indicating the language in which to
translates the variable. Defaults to NULL. Usually is \code{r$lang()}.}

\item{...}{Additional arguments passed to the dispatched function.}
}
\value{
The resulting text.
}
\description{
This function dispatches to the appropriate text-generating function based on
the type of \code{vars} and returns the resulting text.
}
