% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text_delta_change}
\alias{explore_text_delta_change}
\alias{explore_text_delta_change.pct}
\alias{explore_text_delta_change.dollar}
\alias{explore_text_delta_change.ind}
\alias{explore_text_delta_change.default}
\title{Explore Text Delta Change}
\usage{
explore_text_delta_change(var, exp_vals, ...)

\method{explore_text_delta_change}{pct}(var, exp_vals, lang = NULL, ...)

\method{explore_text_delta_change}{dollar}(var, exp_vals, ...)

\method{explore_text_delta_change}{ind}(var, exp_vals, lang, ...)

\method{explore_text_delta_change}{default}(var, exp_vals, ...)
}
\arguments{
\item{var}{<\code{character}> The variable code for which the text and values need
to be generated. Usually \code{vars$var_left}.}

\item{exp_vals}{<\code{list}> A list of values, the output of
\code{\link{explore_text_delta_exp}}}

\item{...}{Additional arguments passed to the method functions.}

\item{lang}{<\code{character}> Active language. Defaults to NULL for no translation}
}
\value{
A character string with the formatted delta change text.
}
\description{
This function calculates and formats the text for delta change in variables,
either in percentage or dollar units, and returns a formatted string.
}
