% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text_delta_exp}
\alias{explore_text_delta_exp}
\alias{explore_text_delta_exp.ind}
\alias{explore_text_delta_exp.default}
\title{Explore text \code{delta} explanation and values for}
\usage{
explore_text_delta_exp(var, region, ...)

\method{explore_text_delta_exp}{ind}(var, region, select_id, ...)

\method{explore_text_delta_exp}{default}(var, region, select_id, data, df, ...)
}
\arguments{
\item{var}{<\code{character}> The variable code for which the text and values need
to be generated. Usually \code{vars$var_left}.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{...}{Additional arguments passed to the dispatched method.}
}
\value{
A list with the following elements:
\item{exp}{A character string describing the explanation of the variable}
\item{region_vals}{A numeric vector containing the variable values in the
current region and time period.}
\item{region_vals_strings}{A character vector with the same length as
\code{region_vals}, containing the variable values formatted according
to the class of \code{var}}
}
\description{
This function delivers the delta text and values of a given variable when
compared between two yeas. It dispatches to specific methods depending on the
structure of the variable.
}
