% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_region_val_df}
\alias{explore_text_region_val_df}
\title{Get region values data frame}
\usage{
explore_text_region_val_df(
  var,
  region,
  select_id,
  col = "var_left",
  scale,
  data,
  lang = NULL,
  time,
  schemas = NULL,
  data_path = get_data_path(),
  ...
)
}
\arguments{
\item{var}{<\code{character}> The variable name for which the region values data
frame needs to be retrieved.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{select_id}{<\code{character}> the current selected ID, usually
\code{r[[id]]$select_id()}.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}

\item{scale}{<\verb{reactive character}> Current scale. The output of
\code{\link{update_scale}}.}

\item{data}{<\code{data.frame}> The output of \code{\link{data_get}}.}

\item{lang}{<\code{character}> Active language. "en" or "fr". Defaults to NULL
for no translation.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{schemas}{<\verb{named list}> Current schema information. The additional widget
values that have an impact on which data column to pick. Usually \code{r[[id]]$schema()}.}

\item{data_path}{<\code{character}> A string representing the path to the
directory containing the QS files. Default is "data/".}

\item{...}{Additional arguments for the \code{\link{explore_text_select_val}}}
}
\value{
The resulting data frame after subsetting or list when there is a
selection.
}
\description{
This function retrieves the region values data frame given the variable name
and region and subsets the data frame based on the region name. If the variable
is a year, then it filters the row based on the year. It then returns the
resulting data frame.
}
\details{
If there is a selection, then
}
