% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_select_val}
\alias{explore_text_select_val}
\alias{explore_text_select_val.pct}
\alias{explore_text_select_val.ind}
\alias{explore_text_select_val.default}
\title{Generate values for the given variable and selection}
\usage{
explore_text_select_val(var, ...)

\method{explore_text_select_val}{pct}(
  var,
  select_id,
  data,
  scale,
  col = "var_left",
  time,
  ...
)

\method{explore_text_select_val}{ind}(
  var,
  data,
  df,
  select_id,
  col = "var_left",
  time,
  lang,
  ...
)

\method{explore_text_select_val}{default}(var, data, df, select_id, col = "var_left", time, ...)
}
\arguments{
\item{var}{<\code{character}> The variable code of the variable for which the
values need to be generated. Usually one element of the output of
\code{\link{vars_build}}.}

\item{...}{Additional arguments passed to the dispatched function.}

\item{select_id}{<\code{character}> The ID of the selected zone.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.}

\item{lang}{<\code{character}> Active language. \code{"en"} or \code{"fr"}}
}
\value{
The resulting values
}
\description{
This function dispatches to the appropriate value-generating function based on
the variable type and returns the resulting values. It is only used when
there is a selection, and replaces the value of \code{\link{explore_text_region_val_df}}
}
\section{Methods (by class)}{
\itemize{
\item \code{explore_text_select_val(pct)}: Method for pct

\item \code{explore_text_select_val(ind)}: Method for \code{ind}

\item \code{explore_text_select_val(default)}: Default method

}}
