% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text_helpers.R
\name{explore_text_select_val}
\alias{explore_text_select_val}
\alias{explore_text_select_val.pct}
\alias{explore_text_select_val.ind}
\alias{explore_text_select_val.default}
\title{Generate values for the given variable and selection}
\usage{
explore_text_select_val(var, ...)

\method{explore_text_select_val}{pct}(var, select_id, data, df, col = "var_left", ...)

\method{explore_text_select_val}{ind}(var, data, df, select_id, col = "var_left", ...)

\method{explore_text_select_val}{default}(var, data, df, select_id, col = "var_left", ...)
}
\arguments{
\item{var}{<\code{character}> The variable code of the variable for which the
values need to be generated. Usually one element of the output of
\code{\link{vars_build}}.}

\item{...}{Additional arguments passed to the dispatched function.}

\item{select_id}{<\code{character}> the current selected ID, usually
\code{r[[id]]$select_id()}.}

\item{data}{<\code{data.frame}>The output of \code{\link{data_get}}.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_df}}.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}
}
\value{
The resulting values
}
\description{
This function dispatches to the appropriate value-generating function based on
the variable type and returns the resulting values. It is only used when
there is a selection, and replaces the value of \code{\link{explore_text_region_val_df}}
}
