% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_text.R
\name{explore_text_values_q5}
\alias{explore_text_values_q5}
\title{Generate text for the given variables and region - q5 version}
\usage{
explore_text_values_q5(var, region, ...)
}
\arguments{
\item{var}{<\code{character}> The variable name for which the text needs to be
generated. Usually \code{vars$var_left}}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{...}{Additional arguments passed to the dispatched function.}
}
\value{
The resulting text.
}
\description{
This function generates text for the given variables and region using the
Q5 version. It dispatches to the appropriate text-generating function based on
the variable type and returns the resulting text.
}
