% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{find_breaks_quintiles}
\alias{find_breaks_quintiles}
\title{Find quintile breaks}
\usage{
find_breaks_quintiles(dist, min_val, max_val, q3_q5 = "q5")
}
\arguments{
\item{dist}{<\code{numeric}> Distribution (numerics) with no NAs.}

\item{min_val}{<\code{numeric}>}

\item{max_val}{<\code{numeric}>}
}
\value{
Returns a numeric vector with quintile breaks.
}
\description{
Find quintile breaks
}
