% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_page.R
\name{home_UI}
\alias{home_UI}
\title{Create the home user interface}
\usage{
home_UI(id = "home", placeholder_video_src, video_src, lang_init = "en")
}
\arguments{
\item{id}{<\code{character}> A unique identifier for the home UI. Normally \code{home}.}

\item{placeholder_video_src}{<\code{character}> External link to a publicly available
mp4 video. The video will be used as the placeholder, until the user click on
to watch the intro video. e.g. \verb{https://s3.amazonaws.com/curbcut.public.resources/mtl_vid_placeholder.mp4}}

\item{video_src}{<\verb{named list}> Every video must be named by the language it uses.
The videos are as external link to a publicly available mp4 video. This is the
full length official video the user will see when they click on to watch the
intro video. e.g. \code{list(en = "https://s3.amazonaws.com/curbcut.public.resources/mtl_vid_en.mp4")}}

\item{lang_init}{<\code{character}> Language which should be used to initiate the
app.}
}
\value{
A Shiny UI object for the home page.
}
\description{
This function generates the user interface for the home page,
retrieving necessary data from the global environment and constructing
the landing page with it.
}
