% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.bivar}
\alias{legend_breaks.bivar}
\title{Compute legend breaks for a comparison of two variables}
\usage{
\method{legend_breaks}{bivar}(vars, df, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{delta}. The output of
\code{\link{vars_build}}.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{df_get}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
The \code{delta} class is used when we compare two scores of the same left-hand
variables but in different years, e.g. "housing_tenant_2016" and
"housing_tenant_2021".
}
