% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.delta}
\alias{legend_breaks.delta}
\title{Compute legend breaks for a single variable of two date times}
\usage{
\method{legend_breaks}{delta}(vars, df, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{delta}.}

\item{df}{<\code{character}> The combination of the region under study
and the scale at which the user is on, e.g. \code{CMA_CSD}. The output of
\code{\link{update_scale}}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with hardcoded labels.
}
\description{
Compute legend breaks for a single variable of two date times
}
