% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.q5_ind}
\alias{legend_breaks.q5_ind}
\title{Compute legend breaks for a single \code{ind} \code{var_left}}
\usage{
\method{legend_breaks}{q5_ind}(vars, data, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{q5_ind}. The output of
\code{\link{vars_build}}.}

\item{data}{<\code{data.frame}> The current data. The output of \code{\link{data_get}}.}

\item{lang}{<\code{character}> String indicating the language to translate the
breaks to. Defaults to \code{NULL}, which is no translation.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
Compute legend breaks for a single \code{ind} \code{var_left}
}
