% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_breaks.R
\name{legend_breaks.qual}
\alias{legend_breaks.qual}
\title{Compute legend breaks for a qualitative variable}
\usage{
\method{legend_breaks}{qual}(vars, region, scale, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object of class \code{qual}. The only necessary
named object in the list is var_left.}

\item{region}{<\code{character}> Indicates the region of interest, e.g. \code{"CMA"}}

\item{scale}{<\code{character}> Indicates the scale of interest, e.g. \code{"CSD"}}

\item{lang}{<\code{character}> String indicating the language to translate the
breaks to. Defaults to \code{NULL}, which is no translation.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A vector of legend breaks with pretty labels.
}
\description{
Compute legend breaks for a qualitative variable
}
