% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_labels.R
\name{legend_labels}
\alias{legend_labels}
\title{Generic legend label function for Curbcut legends}
\usage{
legend_labels(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link[curbcut]{vars_build}}.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}
}
\value{
It returns an output of \code{\link[ggplot2]{labs}}.
}
\description{
\code{legend_labels} is a generic function used to produce a list of
\code{\link[ggplot2]{labs}} for the Curbcut legends. The function invokes
particular methods which depend on the class of the \code{vars} argument.
}
