% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render.delta_bivar}
\alias{legend_render.delta_bivar}
\title{Render the legend for a \code{delta_bivar} class}
\usage{
\method{legend_render}{delta_bivar}(vars, font_family = "SourceSansPro", lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a \code{delta_bivar} class. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{SourceSansPro}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{lang}{<\code{character}> The language to use for the text labels. Defaults
to NULL for no translation.}

\item{...}{Additional arguments passed to other functions.}
}
\value{
A ggplot object that represents the \code{delta_bivar} legend.
}
\description{
This function generates a plot with a bivariate color legend using the
ggplot2 package. It uses the \code{\link{legend_get_info}} function
to show the relationship between the variations of two variables. The legend
consists of a grid of colored squares with text labels indicating the meaning
of each square.
}
