% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render}
\alias{legend_render}
\title{Generic legend render function for Curbcut legends}
\usage{
legend_render(vars, font_family = "SourceSansPro", ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link[curbcut]{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{SourceSansPro}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}
}
\value{
It returns a ggplot object
}
\description{
\code{legend_render} is a generic function used to produce the legend for the
Curbcut map pages. The function invokes particular methods which depend on
the class of the \code{vars} argument.
}
