% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render}
\alias{legend_render}
\alias{legend_render.q5}
\alias{legend_render.q5_ind}
\alias{legend_render.qual}
\alias{legend_render.bivar}
\alias{legend_render.delta}
\alias{legend_render.q100}
\alias{legend_render.delta_bivar}
\alias{legend_render.bivar_ldelta_rq3}
\alias{legend_render.default}
\title{Generic legend render function for Curbcut legends}
\usage{
legend_render(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  data,
  ...
)

\method{legend_render}{q5}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  data,
  ...
)

\method{legend_render}{q5_ind}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  data,
  ...
)

\method{legend_render}{qual}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)

\method{legend_render}{bivar}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  data,
  lang = NULL,
  ...
)

\method{legend_render}{delta}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  data,
  scale,
  time,
  lang = NULL,
  ...
)

\method{legend_render}{q100}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)

\method{legend_render}{delta_bivar}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  data,
  scale,
  time,
  lang = NULL,
  ...
)

\method{legend_render}{bivar_ldelta_rq3}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  data,
  scale,
  time,
  lang = NULL,
  ...
)

\method{legend_render}{default}(vars, ...)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a pre-determined class. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{acidgrotesk-book}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{scale}{<\code{character}> Scale under study. The output of
\code{\link{update_scale}}.}

\item{data}{<\code{data.frame}> The current data. The output of \code{\link{data_get}}.}

\item{...}{Arguments to be passed to the methods, e.g. optionally \code{lang}}

\item{time}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function. It contains the time for both var_left and var_right variables.}
}
\value{
It returns a ggplot object
}
\description{
\code{legend_render} is a generic function used to produce the legend for the
Curbcut map pages. The function invokes particular methods which depend on
the class of the \code{vars} argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{legend_render(q5)}: q5 method

\item \code{legend_render(q5_ind)}: q5_ind method

\item \code{legend_render(qual)}: qual method

\item \code{legend_render(bivar)}: bivar method

\item \code{legend_render(delta)}: delta method

\item \code{legend_render(q100)}: q100 method

\item \code{legend_render(delta_bivar)}: delta_bivar method

\item \code{legend_render(bivar_ldelta_rq3)}: bivar_ldelta_rq3 method

\item \code{legend_render(default)}: default method

}}
