% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render_delta}
\alias{legend_render_delta}
\alias{legend_render_delta.scalar}
\alias{legend_render_delta.ordinal}
\title{Internal function for dispatching \code{legend_render_delta}.}
\usage{
legend_render_delta(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)

\method{legend_render_delta}{scalar}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  ...
)

\method{legend_render_delta}{ordinal}(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  scale,
  lang = NULL,
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> A list object with a \code{delta} class. The
output of \code{\link{vars_build}}.}

\item{font_family}{<\code{character}> Which font family should be used to render
the legend (breaks, axis titles, ...). Defaults to \code{acidgrotesk-book}. To use
the default font family og ggplot2, use \code{NULL}.}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
their colour will be the one of their DA.}

\item{scale}{<\code{character}> Scale under study. The output of
\code{\link{update_scale}}.}

\item{...}{Additional arguments passed to other functions.}

\item{lang}{<\code{character}> Language in use. \code{en} or \code{fr}. Defaults to NULL.}
}
\value{
A ggplot object that represents the \code{delta} legend.

A ggplot object representing the \code{delta} legend for scalar data.

A ggplot object representing the \code{delta} legend for ordinal data.
}
\description{
Internal function for dispatching \code{legend_render_delta}.
}
\section{Methods (by class)}{
\itemize{
\item \code{legend_render_delta(scalar)}: scalar method

\item \code{legend_render_delta(ordinal)}: ordinal method

}}
