% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_render.R
\name{legend_render_delta}
\alias{legend_render_delta}
\title{Internal function for dispatching \code{legend_render_delta}.}
\usage{
legend_render_delta(
  vars,
  font_family = "acidgrotesk-book",
  scales_as_DA = c("building", "street"),
  df,
  ...
)
}
\description{
Internal function for dispatching \code{legend_render_delta}.
}
