% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Update tabset panel and select an ID}
\usage{
link(session, r, page, select_id = NA, df = NULL)
}
\arguments{
\item{session}{<\code{session}> A shiny session object from the server function.
The actual session of the module, should be  \code{session = session}}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{page}{<\code{character}> The id of the tab to be opened.}

\item{select_id}{<\code{character}> ID of the item to be selected within the newly
opened page.}

\item{df}{<\code{character}> Combination of region and scale to update the zoom
of the newly opened page's map.}
}
\description{
This function updates the tabset panel with the given \code{page} and selects
the corresponding \code{select_id}. The \code{page} and \code{select_id}
arguments must be reactive objects.
}
