% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Update tabset panel and select an ID}
\usage{
link(
  r,
  page,
  region = r$region(),
  select_id = NA,
  df = NULL,
  date = NULL,
  var_left = NULL,
  var_right = NULL,
  zoom_levels = get_from_globalenv(paste("map_zoom_levels", region, sep = "_")),
  zoom = link_get_zoom(zoom_levels, df)
)
}
\arguments{
\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{page}{<\code{character}> The id of the tab to be opened.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Defaults to \code{r$region()}.}

\item{select_id}{<\code{character}> ID of the item to be selected within the newly
opened page.}

\item{df}{<\code{character}> Combination of region and scale to update the zoom
of the newly opened page's map.}

\item{date}{<\verb{numeric vector}> Change the date on the page. It will be changing
the dates widgets from the autovars module, with html id as \code{alp-alp-alp-ccslider_slb}.}

\item{var_right}{<\verb{reactive character}> Character string of the compare variable
to update to. It will be changing the html id of the \code{alp-alp-compare-ccpicker_var}
widget.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}.}

\item{zoom}{<\code{numeric}> Zoom used to update the map viewstate. Defaults to
using \code{\link{link_get_zoom}}, which will grab the values in the \code{zoom_levels}
argument.}
}
\description{
This function updates the tabset panel with the given \code{page} and selects
the corresponding \code{select_id}. The \code{page} and \code{select_id}
arguments must be reactive objects.
}
