% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\title{Update tabset panel and select an ID}
\usage{
link(r, page, select_id = NA, df = NULL)
}
\arguments{
\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{page}{A reactive object containing the ID of the tab to be selected.}

\item{select_id}{A reactive object containing the ID of the item to be selected
within the selected tab.}
}
\description{
This function updates the tabset panel with the given \code{page} and selects
the corresponding \code{select_id}. The \code{page} and \code{select_id}
arguments must be reactive objects.
}
