% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scales.R
\name{map_scale_colour}
\alias{map_scale_colour}
\title{Map scale colour}
\usage{
map_scale_colour(...)
}
\arguments{
\item{...}{Unused arguments.}
}
\value{
Returns the hex of the white colour.
}
\description{
Map scale colour
}
