% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_scales.R
\name{map_scale_lwd}
\alias{map_scale_lwd}
\title{Map scale line width}
\usage{
map_scale_lwd(select_id, tile = NULL, zoom = NULL, zoom_levels = NULL, lwd = 1)
}
\arguments{
\item{select_id}{<\code{character}> A string specifying the ID of the selected polygon.
The selected polygon will have an increased in the width of its border.}

\item{tile}{<\code{character}> The name of the tileset currently present on the map.}

\item{zoom}{<\code{numeric}> The current zoom level of the map.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-scale).}

\item{lwd}{<\code{numeric}> Value specifying the line width of the polygon borders,
with a default value of 1. This value controls the thickness of the borders
around the polygons on the map. A larger value results in thicker borders,
while a smaller value results in thinner borders.}
}
\value{
A line width scale for polygons in an rdeck map.
}
\description{
This function creates a line width scale for polygons in an rdeck map based
on the zoom level of the map. If \code{tile}, \code{zoom} and \code{zoom_levels} are given,
the borders will disappear at one zoom digit under the zoom at which they
would have appeared on an auto-scale.
}
