% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_schema_to_col}
\alias{match_schema_to_col}
\title{Match schema list to right column in \code{data}}
\usage{
match_schema_to_col(
  data,
  time,
  col = "var_left",
  data_schema = attr(data, sprintf("schema_\%s", col)),
  schemas,
  closest_time = FALSE
)
}
\arguments{
\item{data}{<\code{data.frame}> A data frame containing the data. The output of
output of \code{\link{get_data}}.}

\item{time}{<\verb{numeric named list}> The \code{time} at which data is displayed.
A list for var_left and var_right. The output of \code{\link{vars_build}}(...)$time.
Can also be a simple numeric.}

\item{col}{<\code{character}> Which column should be extracted? \code{var_left} or \code{var_right}.}

\item{data_schema}{<\verb{named list}> A list containing the schema information, specifically the
'time' attribute. Typically obtained, and defaulted, as an attribute to
\code{data} (output of \code{\link{get_data}}).}

\item{schemas}{<\verb{named list}> Current schema information. The additional widget
values that have an impact on which data column to pick. Usually \code{r[[id]]$schema()}.
Can be NULL if there are none (ex. taking a parent variable, no schemas).}

\item{closest_time}{<\code{logical}> Should the closest time be used if the exact
time is not found? Default is \code{FALSE}.}
}
\value{
Returns the name of the variable that corresponds to the given
schema as a character string.
}
\description{
This function identifies the column in a \code{data} (output of \code{\link{get_data}})
that matches a given schema named list. It is designed to be used in situations where
the schema for the data frame can be dynamic. For example, if the schema is
based on the time of the data, this function can be used to identify the
column in the data frame that corresponds to the given time.
}
