% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_warning.R
\name{mobile_warning}
\alias{mobile_warning}
\title{Display a warning message if the user is accessing the app from a mobile device}
\usage{
mobile_warning(id = "mobile_warning", r, parent_session)
}
\arguments{
\item{id}{<\code{character}> A unique id for the module. Default is "mobile_warning".}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file.}

\item{parent_session}{<\code{session}> The session of the parent module that the
warning is embedded in (\code{server.R}). Usually \code{parent_session = session}.}
}
\description{
This function displays a warning message if the user is accessing the Shiny app
from a mobile device. The message is displayed using the
\code{\link[shinyjs]{info}} function.
}
