% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_helpers.R
\name{panel_view_prepare_text}
\alias{panel_view_prepare_text}
\alias{panel_view_prepare_text.q5}
\alias{panel_view_prepare_text.delta}
\alias{panel_view_prepare_text.bivar}
\alias{panel_view_prepare_text.delta_bivar}
\alias{panel_view_prepare_text.default}
\title{Prepare the panel view text for the given variables.}
\usage{
panel_view_prepare_text(vars, scale, dat, time, lang = NULL, ...)

\method{panel_view_prepare_text}{q5}(vars, scale, dat, time, lang = NULL, ...)

\method{panel_view_prepare_text}{delta}(vars, df, dat, time, lang = NULL, ...)

\method{panel_view_prepare_text}{bivar}(vars, df, dat, time, lang = NULL, ...)

\method{panel_view_prepare_text}{delta_bivar}(vars, df, time, dat, lang = NULL, ...)

\method{panel_view_prepare_text}{default}(vars, df, dat, time, lang = NULL, ...)
}
\arguments{
\item{vars}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function.}

\item{scale}{<\code{character}> Current scale, e.g. \code{"DA"}.}

\item{dat}{<\code{data.frame}> The data frame containing the columns under analysis}

\item{time}{<\verb{named list}> Object built using the \code{\link{vars_build}}
function. It contains the time for both var_left and var_right variables.}

\item{lang}{<\code{character}> The language to use for translating the texts.
Defaults to NULL for no translation}

\item{...}{Additional arguments to be passed to the underlying methods.}
}
\value{
A string containing the panel view text.
}
\description{
This function returns a panel view text including title and description for
given variables, considering the minimum, maximum, mean, and standard deviation
of the data points in the input dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{panel_view_prepare_text(q5)}: The method for q5.

\item \code{panel_view_prepare_text(delta)}: The method for delta.

\item \code{panel_view_prepare_text(bivar)}: The method for bivar.

\item \code{panel_view_prepare_text(delta_bivar)}: The method for delta_bivar.

\item \code{panel_view_prepare_text(default)}: The default method.

}}
