% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_view_module.R
\name{panel_view_server}
\alias{panel_view_server}
\alias{panel_view_UI}
\title{Renders a floating panel with a map view and data view buttons.}
\usage{
panel_view_server(
  id,
  r,
  region,
  vars,
  data,
  zoom_levels,
  temp_folder = get_from_globalenv("temp_folder"),
  scales_as_DA = shiny::reactive(c("building", "street"))
)

panel_view_UI(id)
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which the panel will appear,
e.g. \code{alp}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{region}{<\code{character}> Character string specifying the name of the region.
Usually equivalent of \code{r$region()}.}

\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function. The class of the vars object is
used to determine which type of legend to draw.}

\item{data}{<\verb{reactive data.frame}> Data frame containing all the scale and
the \code{var_left} and \code{var_right}. The output of \code{\link{data_get}}.}

\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-scale).}

\item{temp_folder}{<\code{character}> The temporary folder of the app. By default
will grab the \code{temp_folder} object as it's already supposed to have been assigned
in the \code{global.R} file}

\item{scales_as_DA}{<\verb{reactive character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. By default,
the data shown will be \code{DA}.}
}
\value{
Panel view module
}
\description{
This function renders a Shiny server module for the \code{panel_view_UI} UI
element which displays a floating panel with buttons to toggle between the map
an table/data info. It shows a map when the map button is clicked and a table
with data when the data button is clicked. When there is a selection, it also
shows a magnifying glass linking to the place explorer.
}
\section{Functions}{
\itemize{
\item \code{panel_view_UI()}: Create the UI for the legend module

}}
