% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_region_value.R
\name{region_value}
\alias{region_value}
\title{Compute region values based on different methods}
\usage{
region_value(var, data, time, scale, region, select_id, col, ...)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{time}{<\code{numeric}> The time period of interest.}

\item{scale}{<\code{character}> The scale of interest.}

\item{region}{<\code{character}> The region of interest.}

\item{select_id}{<\code{character}>  The ID of the selected feature.}

\item{col}{<\code{character}> Which column of \code{data} should be selected to grab the
value information. Defaults to \code{var_left}, but could also be \code{var_right} or
\code{var_left_1} in delta.}

\item{...}{Additional arguments to pass to specific method functions.}
}
\value{
Varies based on the method specified in \code{var}. Commonly
returns a list with elements like \code{val} and \code{count}.
}
\description{
This function computes region values based on the method specified by \code{var}'s
class. It uses weighted means, sums, or other calculations.
It fetches the relevant data and parent data, and then calls the
appropriate method for calculation.
}
