% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_region_value.R
\name{region_value_method}
\alias{region_value_method}
\alias{region_value_method.pct}
\alias{region_value_method.dollar}
\alias{region_value_method.count}
\alias{region_value_method.ind}
\alias{region_value_method.default}
\title{Methods to compute regional values}
\usage{
region_value_method(var, data_vals, parent_vals, ...)

\method{region_value_method}{pct}(var, data_vals, parent_vals, ...)

\method{region_value_method}{dollar}(var, data_vals, parent_vals, ...)

\method{region_value_method}{count}(var, data_vals, parent_vals, ...)

\method{region_value_method}{ind}(var, data_vals, parent_vals, data, time, ...)

\method{region_value_method}{default}(var, data_vals, parent_vals, data, time, ...)
}
\arguments{
\item{var}{<\code{character}> The code of the variable of interest, with its class
to dispatch to the right method.}

\item{data_vals}{<\verb{numeric vector}> Numeric vector containing the data values
for the region.}

\item{parent_vals}{<\verb{numeric vector}> Numeric vector containing the parent
data values.}

\item{...}{Additional arguments (e.g., data, time) used in specific methods.}

\item{data}{<\code{data.frame}> The data frame containing the data.}

\item{time}{<\code{numeric}> The time period of interest.}
}
\value{
Returns a list with calculated values for the specific method.
Commonly returns elements such as \code{val} and \code{count}.
}
\description{
A set of methods to compute regional values based on the type of
variable provided. Supports weighted mean, sums, and specific
indicators.
}
\section{Methods (by class)}{
\itemize{
\item \code{region_value_method(pct)}: The method for percentage variables.

\item \code{region_value_method(dollar)}: The method for dollar variables.

\item \code{region_value_method(count)}: The method for count variables.

\item \code{region_value_method(ind)}: The method for \code{ind} variables.

\item \code{region_value_method(default)}: The default method.

}}
