% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Curbcut Shiny Server Function}
\usage{
server(lang_init = "en", show_lang_button = FALSE)
}
\arguments{
\item{lang_init}{<\code{character}> Language which should be used to initiate the
app. Defaults to english.}

\item{show_lang_button}{<\code{logical}> Should there be a language button? Defaults
to FALSE (only english).}
}
\value{
Nothing. This function initiates the Shiny server for the
Curbcut application, and its results are side effects (i.e., launching the
application).
}
\description{
This function initiates a shiny server to render Curbcut, an urban
sustainability exploration platform. It includes reactive variables
initialization, title page updating, bookmarking, modules loading,
advanced settings handling, and a heartbeat function to keep the app alive.
}
