% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_module.R
\name{settings_server}
\alias{settings_server}
\title{Server logic for the settings UI module}
\usage{
settings_server(id = "settings", r)
}
\arguments{
\item{id}{<\code{character}> The ID of the module, used to connect the UI and server
functions. Defaults to \code{"settings"}.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file.}
}
\description{
This function contains the server-side logic for the settings UI module.
It connects to the \code{settings_UI} function through the module ID and a
reactive values object, and handles various events triggered by user
interactions with the UI. Specifically, it opens a modal dialog for advanced
options, updates the region cookie if a user changes the default region,
retrieves the IDs for a searched location and saves it in a reactive value,
and clears the default location when a button is clicked.
}
\seealso{
\code{\link{settings_UI}}
}
