% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_module.R
\name{settings_UI}
\alias{settings_UI}
\title{Create a settings menu UI for Curbcut}
\usage{
settings_UI(id = "settings", contact_email = "contact@curbcut.ca")
}
\arguments{
\item{id}{A character string specifying the input ID for the settings menu.
Defaults to \code{"settings"}.}

\item{contact_email}{A character string specifying the email address for
providing feedback. Defaults to \code{"contact@curbcut.ca"}.}
}
\value{
A dropdown button object from the shinyWidgets package representing
the settings menu UI.
}
\description{
This function creates a settings menu for a Shiny app, with several action
links for bookmarking, providing feedback, exporting data, subscribing to a
newsletter, and accessing advanced options. The UI is returned as a dropdown
button object from the shinyWidgets package.
}
\seealso{
\code{\link{settings_server}}
}
