% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{test_assign_delta_ind_ord}
\alias{test_assign_delta_ind_ord}
\title{Assign a ind ordinal delta vars/data to the global environment}
\usage{
test_assign_delta_ind_ord(pos = 1)
}
\arguments{
\item{pos}{<\code{numeric}> Defaults to 1, the environment in which to assign the
values. This argument is to appease RMD check.}
}
\value{
Assigns \code{df}, \code{vars}, \code{data} corresponding to a delta index ordinal
variable to the global environment
}
\description{
Assign a ind ordinal delta vars/data to the global environment
}
