% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_pages_server.R
\name{trigger_pages_server}
\alias{trigger_pages_server}
\title{Trigger the \code{server} of pages}
\usage{
trigger_pages_server(cc_page, r, output)
}
\arguments{
\item{cc_page}{<\code{reactive}> Character of the currently active tab stored
in a \code{\link[shiny]{reactive}} to ensure reactivity on change. Usually
\code{cc_page = shiny::reactive(input$cc_page)}}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{output}{<\code{shiny::output}> The output slot for storing UI elements. Usually
\code{output = output}.}
}
\value{
Triggers the server module function if it hasn't already been
triggered in the same session.
}
\description{
This function creates a reactive environment for tracking the state of pages
opened in Curbcut and triggers the server module of a new page when it is
opened for the first time. The function takes a single argument cc_page,
which is a reactive expression that tracks the currently selected page.
}
