% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ui}
\alias{ui}
\title{Generate Curbcut Shiny User Interface}
\usage{
ui(
  site_name,
  web_description,
  web_title,
  placeholder_video_src,
  video_src,
  twitter_handler,
  google_analytics,
  website_url,
  share_jpg,
  apple_touch_icon,
  lang_init = "en"
)
}
\arguments{
\item{site_name}{<\code{character}>¨Name of the site with the city, e.g. \verb{Curbcut Montréal}.}

\item{web_description}{<\code{character}> The description of the website for metadata.
e.g. \code{"Curbcut est une plateforme d'exploration approfondie, dynamique et intuitive de la durabilité urbaine."}}

\item{web_title}{<\code{character}> The title of the website for metadata.
e.g. \verb{Curbcut Montréal | Vers une ville durable}}

\item{placeholder_video_src}{<\code{character}> External link to a publicly available
mp4 video. The video will be used as the placeholder, until the user click on
to watch the intro video. e.g. \verb{https://s3.amazonaws.com/curbcut.public.resources/mtl_vid_placeholder.mp4}}

\item{video_src}{<\code{character}> External link to a publicly available
mp4 video. This is the full length official video the user will see when
they click on to watch the intro video. e.g. \verb{https://s3.amazonaws.com/curbcut.public.resources/mtl_vid_placeholder.mp4}}

\item{twitter_handler}{<\code{character}> The Twitter handle associated with the
website for metadata.e.g. \code{"@curbcutca"}}

\item{google_analytics}{<\code{character}> The Google Analytics script location
for tracking user activity. e.g. \code{"www/google_analytics.html"}}

\item{website_url}{<\code{character}> The URL of the website for metadata.
e.g. \code{"https://montreal.curbcut.ca"}}

\item{share_jpg}{<\code{character}> ABSOLUTE path to he URL of the image for sharing
on social media. No relative paths. e.g. \code{"https://montreal.curbcut.ca/share.jpg"}}

\item{apple_touch_icon}{<\code{character}> Absolute path to the 192x192px logo.
e.g. \code{"https://montreal.curbcut.ca/logo192.jpg"}}

\item{lang_init}{<\code{character}> Language which should be used to initiate the
app.}
}
\value{
A Shiny UI object that includes all elements of the Curbcut application interface.
}
\description{
This function generates the main Shiny UI for the Curbcut application.
The UI includes elements such as the import of package dependencies,
application styles, Curbcut scripts, Google Analytics, metadata for sharing,
and the creation of the pages UI.
}
