% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_select_id}
\alias{update_select_id}
\title{Update Select ID}
\usage{
update_select_id(
  id,
  r,
  data = shiny::reactive(NULL),
  id_map = paste0(id, "-map")
)
}
\arguments{
\item{id}{<\code{character}> Indicates the ID of the current page.}

\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}

\item{data}{<\verb{reactive data.frame}> A data frame containing the \code{ID} column to be
checked for any match with the \code{r$default_select_ids()} changed in
the advanced options.}

\item{id_map}{<\code{character}> Indicates the ID of the object map, usually
created by \code{\link{map_server}}. Defaults to \code{paste0(id, "-map")} as
the default of the \code{map_server} function.}
}
\value{
This function does not return a value. Instead, it updates the
\code{select_id} reactive in the provided reactive environment.
}
\description{
This function updates the selected ID on a Curbcut map page. It uses the
\code{\link[rdeck]{get_clicked_object}} to get the ID of the clicked map.
It looks if the selection has been made on a 'stories' bubble and if so,
links to the stories page.
If a new ID is selected, the function updates the \code{select_id} reactive to
the newly selected ID. If the same is selected twice, it returns NA. It also
updates the select_id reactive if a match is found with the IDs from
\code{r$default_select_ids()} if the user has decided to lock in a default location
in the advanced options.
}
