% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_rvs.R
\name{update_select_id}
\alias{update_select_id}
\title{Update selected ID based on clicked object}
\usage{
update_select_id(id, select_id, id_map = paste0(id, "-map"))
}
\arguments{
\item{id}{<\code{character}> The ID of the page in which this function will appear,
e.g. \code{canale}.}

\item{select_id}{<\code{character}> A character string representing the currently
selected ID}

\item{id_map}{<\code{character}> The ID of the map. By default, it is
set to the value of \code{paste0(id, "-map")} as the \code{\link{map_server}}
module assigns the \code{map} id to the maps.}
}
\value{
A character string representing the new selected ID, or \code{NA} if
the same ID gets selected twice
}
\description{
This function updates the currently selected ID in a Shiny app based on the
ID of the object that was clicked on on an \code{rdeck} map.
}
