% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{use_bookmark}
\alias{use_bookmark}
\title{Use Bookmark}
\usage{
use_bookmark(r)
}
\arguments{
\item{r}{<\code{reactiveValues}> The reactive values shared between modules and
pages. Created in the \code{server.R} file. The output of \code{\link{r_init}}.}
}
\value{
handles updating the state of the app based on the URL query string.
}
\description{
This function is a module server that updates reactive values and UI elements
based on URL query string. The function is designed enable bookmarking and sharing
of a specific state of the application. It is designed to be placed only once,
in the \code{server.R} file.
}
