% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_utils.R
\name{zoom_get_label}
\alias{zoom_get_label}
\title{Get the zoom labels for a set of zoom levels}
\usage{
zoom_get_label(zoom_levels, lang = NULL)
}
\arguments{
\item{zoom_levels}{<\verb{named numeric vector}> A named numeric vector of zoom
levels. Usually one of the \code{map_zoom_levels_x}, or the output of
\code{\link{zoom_get_levels}}. It needs to be \code{numeric} as the function
will sort them to make sure the lower zoom level is first, and the highest
is last (so it makes sense on an auto-zoom).}

\item{lang}{<\code{character}> String indicating the language to translate the
slider titles to. Defaults to \code{NULL}, which is no translation.}
}
\value{
A character vector of slider titles.
}
\description{
This function takes a set of zoom levels ("CMA", "CT", "DA") and returns the
corresponding slider titles ("Borough/City" , "Census tract", "Dissemination area")
}
