% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_helpers.R
\name{zoom_level_selection}
\alias{zoom_level_selection}
\alias{zoom_level_selection.bivar}
\alias{zoom_level_selection.default}
\title{Internal function for dispatching \code{zoom_level_selection}.}
\usage{
zoom_level_selection(
  vars,
  top_scale,
  avail_scale_combinations,
  scales_as_DA = c("building", "street"),
  ...
)

\method{zoom_level_selection}{bivar}(
  vars,
  top_scale,
  avail_scale_combinations,
  scales_as_DA = c("building", "street"),
  ...
)

\method{zoom_level_selection}{default}(
  vars,
  top_scale,
  avail_scale_combinations,
  scales_as_DA = c("building", "street"),
  ...
)
}
\arguments{
\item{vars}{<\verb{named list}> Named list with a class. Object built using the
\code{\link{vars_build}} function. The class of the vars object is
used to determine which type of legend to draw.}

\item{top_scale}{<\code{character}>}

\item{avail_scale_combinations}{<\verb{character vector}>}

\item{scales_as_DA}{<\verb{character vector}> A character vector of \code{scales}
that should be handled as a "DA" scale, e.g. \code{building} and \code{street}. They won't
be taken into account when all the scales need to hold data.}

\item{...}{Additional arguments to be passed to the methods.}
}
\value{
A ggplot object that represents the \code{delta} legend.
}
\description{
Internal function for dispatching \code{zoom_level_selection}.
}
\section{Methods (by class)}{
\itemize{
\item \code{zoom_level_selection(bivar)}: bivar method

\item \code{zoom_level_selection(default)}: default method

}}
